new library view, changelog readme,
