+++ src-tauri/src/library/mod.rs
@@ -9,5 +9,6 @@ pub(crate) use commands::handle_opened_paths;
 #[cfg(target_os = "macos")]
 pub use types::EVENT_LIBRARY_RENDERER_READY;
 pub use types::{
-    LibraryFilter, LibraryItem, LibraryItemPatch, LibraryItemStatus, TranscriptSegment,
+    LibraryFilter, LibraryItem, LibraryItemPatch, LibraryItemStatus, Speaker, TranscriptSegment,
 };
+pub(crate) use types::default_item_kind;
