--- src/features/library/components/LibraryModal.tsx
@@ -1,1702 +1,0 @@
-import { useLingui } from "@lingui/react/macro";
-import {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-  type ReactNode,
-} from "react";
-import { AnimatePresence, motion } from "framer-motion";
-import { invoke, convertFileSrc } from "@tauri-apps/api/core";
-import { save } from "@tauri-apps/plugin-dialog";
-import { Howl } from "howler";
-import { Virtuoso, type VirtuosoHandle } from "react-virtuoso";
-import {
-  Warning as AlertTriangle,
-  Check,
-  CaretDown as ChevronDown,
-  CaretLeft as ChevronLeft,
-  CaretRight as ChevronRight,
-  Copy,
-  ArrowElbowDownRight as CornerDownRight,
-  Pause,
-  PencilSimple as Pencil,
-  Play,
-  ArrowClockwise as RotateCw,
-  MagnifyingGlass as Search,
-  Trash as Trash2,
-  X,
-} from "@phosphor-icons/react";
-import LibraryRetranscribeModal from "./LibraryRetranscribeModal";
-import {
-  clampProgress,
-  formatDuration,
-  formatPlaybackRate,
-  formatTimestamp,
-  getLibraryErrorDetails,
-  PLAYBACK_RATES,
-  sanitizeFileName,
-  shouldShowImportProgress,
-  formatLibraryName,
-} from "./library-utils";
-import { resolveSpeechModelLabel } from "../../settings/models-queries";
-import { useClickOutside } from "../../../shared/hooks/useClickOutside";
-import { IntelligencePixel } from "../../../shared/ui/IntelligencePixel";
-import ToggleSwitch from "../../../shared/ui/ToggleSwitch";
-import type {
-  ExportFormat,
-  LibraryItem,
-  LibraryItemPatch,
-  SpeechModel,
-  TranscriptSegment,
-} from "../../../types";
-
-const LibraryModal = ({
-  item,
-  models,
-  shiftHeld,
-  followTimestamps,
-  onFollowTimestampsChange,
-  onClose,
-  onDelete,
-  onRetry,
-  onCancel,
-  onUpdate,
-  onExport,
-  availableTags,
-}: {
-  item: LibraryItem;
-  models: SpeechModel[];
-  shiftHeld: boolean;
-  followTimestamps: boolean;
-  onFollowTimestampsChange: (
-    value: boolean | ((prev: boolean) => boolean),
-  ) => void;
-  onClose: () => void;
-  onDelete: () => void;
-  onRetry: () => Promise<void>;
-  onCancel: () => void;
-  onUpdate: (patch: LibraryItemPatch) => Promise<LibraryItem>;
-  onExport: (format: ExportFormat, outputPath: string) => Promise<void>;
-  availableTags: string[];
-}) => {
-  const { t } = useLingui();
-  const [nameDraft, setNameDraft] = useState(item.name);
-  const [isEditingName, setIsEditingName] = useState(false);
-  const [transcriptDraft, setTranscriptDraft] = useState(item.transcript ?? "");
-  const [tagInput, setTagInput] = useState("");
-  const [tagMenuOpen, setTagMenuOpen] = useState(false);
-  const [showTimestamps, setShowTimestamps] = useState(
-    item.show_timestamps && Boolean(item.segments?.length),
-  );
-  const [exportOpen, setExportOpen] = useState(false);
-  const [isExporting, setIsExporting] = useState(false);
-  const [showDeleteConfirm, setShowDeleteConfirm] = useState(false);
-  const [copyConfirmed, setCopyConfirmed] = useState(false);
-  const [audioDuration, setAudioDuration] = useState(
-    item.duration_seconds || 0,
-  );
-  const [audioCurrentTime, setAudioCurrentTime] = useState(0);
-  const [isPlaying, setIsPlaying] = useState(false);
-  const [audioReady, setAudioReady] = useState(false);
-  const [audioError, setAudioError] = useState<string | null>(null);
-  const [playbackRate, setPlaybackRate] = useState(1);
-  const [isScrubbing, setIsScrubbing] = useState(false);
-  const [streamChunks, setStreamChunks] = useState<string[]>([]);
-  const [showRetranscribe, setShowRetranscribe] = useState(false);
-  const [searchQuery, setSearchQuery] = useState("");
-  const [activeSearchIndex, setActiveSearchIndex] = useState(0);
-  const transcriptTimer = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const copyTimer = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const howlRef = useRef<Howl | null>(null);
-  const tagMenuRef = useRef<HTMLDivElement>(null);
-  const playbackRateRef = useRef(1);
-  const streamTranscriptRef = useRef(item.transcript ?? "");
-  const scrubWasPlayingRef = useRef(false);
-  const scrubValueRef = useRef<number | null>(null);
-  const rafRef = useRef<number | null>(null);
-  const isScrubbingRef = useRef(false);
-  const isPlayingRef = useRef(false);
-  const lastTimestampNavRef = useRef(0);
-  const transcriptAreaRef = useRef<HTMLTextAreaElement | null>(null);
-  const segmentsVirtuosoRef = useRef<VirtuosoHandle | null>(null);
-  const streamVirtuosoRef = useRef<VirtuosoHandle | null>(null);
-
-  const modelLabel =
-    resolveSpeechModelLabel(models, item.speech_model) ?? item.speech_model;
-  const transcriptAvailable =
-    item.status.type === "complete" &&
-    (item.transcript ?? "").trim().length > 0;
-  const canShowTimestamps = !!item.segments && item.segments.length > 0;
-  const isTranscribed = item.status.type === "complete";
-  const importStatusText =
-    item.status.type === "importing"
-      ? shouldShowImportProgress(item.status.progress)
-        ? t({
-            id: "library.modal.import_status.converting_progress",
-            message: `Converting audio... ${Math.round(clampProgress(item.status.progress) * 100)}%`,
-          })
-        : t({
-            id: "library.modal.import_status.converting",
-            message: "Converting audio...",
-          })
-      : t({
-          id: "library.modal.import_status.queued",
-          message: "Queued for transcription...",
-        });
-
-  const audioUrl = useMemo(
-    () => convertFileSrc(item.audio_path),
-    [item.audio_path],
-  );
-
-  const stopSeekLoop = useCallback(() => {
-    if (rafRef.current !== null) {
-      cancelAnimationFrame(rafRef.current);
-      rafRef.current = null;
-    }
-  }, []);
-
-  const updateIsPlaying = useCallback((value: boolean) => {
-    isPlayingRef.current = value;
-    setIsPlaying(value);
-  }, []);
-
-  const updateIsScrubbing = useCallback((value: boolean) => {
-    isScrubbingRef.current = value;
-    setIsScrubbing(value);
-  }, []);
-
-  const setPlaybackRateValue = useCallback((value: number) => {
-    playbackRateRef.current = value;
-    setPlaybackRate(value);
-    howlRef.current?.rate(value);
-  }, []);
-
-  const startSeekLoop = useCallback(() => {
-    stopSeekLoop();
-    const tick = () => {
-      const sound = howlRef.current;
-      if (sound) {
-        const playing = sound.playing();
-        if (playing !== isPlayingRef.current) {
-          isPlayingRef.current = playing;
-          setIsPlaying(playing);
-        }
-        if (playing && !isScrubbingRef.current) {
-          const pos = sound.seek();
-          if (typeof pos === "number") {
-            setAudioCurrentTime(pos);
-          }
-        }
-      }
-      rafRef.current = requestAnimationFrame(tick);
-    };
-    rafRef.current = requestAnimationFrame(tick);
-  }, [stopSeekLoop]);
-
-  useEffect(() => {
-    if (!isEditingName) {
-      setNameDraft(item.name);
-    }
-  }, [isEditingName, item.name]);
-
-  useEffect(() => {
-    setShowTimestamps(item.show_timestamps && canShowTimestamps);
-  }, [item.show_timestamps, canShowTimestamps]);
-
-  useEffect(() => {
-    stopSeekLoop();
-    if (howlRef.current) {
-      howlRef.current.unload();
-      howlRef.current = null;
-    }
-    updateIsPlaying(false);
-    updateIsScrubbing(false);
-    setAudioReady(false);
-    setAudioError(null);
-    setAudioCurrentTime(0);
-    setAudioDuration(item.duration_seconds || 0);
-    scrubWasPlayingRef.current = false;
-    scrubValueRef.current = null;
-
-    const sound = new Howl({
-      src: [audioUrl],
-      html5: true,
-      preload: true,
-      onload: () => {
-        const duration = sound.duration();
-        setAudioDuration(Number.isFinite(duration) ? duration : 0);
-        setAudioReady(true);
-      },
-      onloaderror: (_id: number | string, err: unknown) => {
-        console.error("Audio load error:", err);
-        setAudioError(
-          t({
-            id: "library.modal.audio_unavailable",
-            message: "Audio unavailable",
-          }),
-        );
-        setAudioReady(false);
-      },
-      onplayerror: (_id: number | string, err: unknown) => {
-        console.error("Audio play error:", err);
-        setAudioError(
-          t({
-            id: "library.modal.audio_unavailable",
-            message: "Audio unavailable",
-          }),
-        );
-        setAudioReady(false);
-        updateIsPlaying(false);
-        stopSeekLoop();
-      },
-      onplay: () => {
-        updateIsPlaying(true);
-        startSeekLoop();
-      },
-      onpause: () => {
-        updateIsPlaying(false);
-        stopSeekLoop();
-      },
-      onstop: () => {
-        updateIsPlaying(false);
-        stopSeekLoop();
-      },
-      onend: () => {
-        updateIsPlaying(false);
-        stopSeekLoop();
-        const duration = sound.duration();
-        if (Number.isFinite(duration)) {
-          setAudioCurrentTime(duration);
-        }
-      },
-      onseek: () => {
-        if (isScrubbingRef.current) return;
-        const pos = sound.seek();
-        if (typeof pos === "number") {
-          setAudioCurrentTime(pos);
-        }
-      },
-    });
-
-    sound.rate(playbackRateRef.current);
-    howlRef.current = sound;
-
-    return () => {
-      stopSeekLoop();
-      sound.unload();
-    };
-  }, [
-    audioUrl,
-    item.duration_seconds,
-    startSeekLoop,
-    stopSeekLoop,
-    updateIsPlaying,
-    updateIsScrubbing,
-  ]);
-
-  const handlePlaybackRateStep = useCallback(
-    (direction: -1 | 1) => {
-      const currentIndex = PLAYBACK_RATES.indexOf(playbackRate);
-      const safeIndex =
-        currentIndex === -1 ? PLAYBACK_RATES.indexOf(1) : currentIndex;
-      const nextIndex = Math.min(
-        PLAYBACK_RATES.length - 1,
-        Math.max(0, safeIndex + direction),
-      );
-      setPlaybackRateValue(PLAYBACK_RATES[nextIndex]);
-    },
-    [playbackRate, setPlaybackRateValue],
-  );
-
-  const handleRateScrubStart = useCallback(
-    (
-      event:
-        | React.MouseEvent<HTMLSpanElement>
-        | React.TouchEvent<HTMLSpanElement>,
-    ) => {
-      event.preventDefault();
-      const startX =
-        "touches" in event ? event.touches[0].clientX : event.clientX;
-      const startIndex = PLAYBACK_RATES.indexOf(playbackRateRef.current);
-      const initialIndex =
-        startIndex === -1 ? PLAYBACK_RATES.indexOf(1) : startIndex;
-
-      const handleMove = (e: MouseEvent | TouchEvent) => {
-        const currentX =
-          "touches" in e ? e.touches[0].clientX : (e as MouseEvent).clientX;
-        const diffX = currentX - startX;
-        const steps = Math.round(diffX / 15);
-
-        const nextIndex = Math.min(
-          PLAYBACK_RATES.length - 1,
-          Math.max(0, initialIndex + steps),
-        );
-
-        if (PLAYBACK_RATES[nextIndex] !== playbackRateRef.current) {
-          setPlaybackRateValue(PLAYBACK_RATES[nextIndex]);
-        }
-      };
-
-      const handleEnd = () => {
-        window.removeEventListener("mousemove", handleMove);
-        window.removeEventListener("mouseup", handleEnd);
-        window.removeEventListener("touchmove", handleMove);
-        window.removeEventListener("touchend", handleEnd);
-      };
-
-      window.addEventListener("mousemove", handleMove);
-      window.addEventListener("mouseup", handleEnd);
-      window.addEventListener("touchmove", handleMove, { passive: false });
-      window.addEventListener("touchend", handleEnd);
-    },
-    [setPlaybackRateValue],
-  );
-
-  useEffect(() => {
-    setTranscriptDraft(item.transcript ?? "");
-  }, [item.transcript]);
-
-  useEffect(() => {
-    if (item.status.type !== "transcribing") {
-      setStreamChunks([]);
-      streamTranscriptRef.current = item.transcript ?? "";
-    }
-  }, [item.status.type, item.transcript]);
-
-  useEffect(() => {
-    if (item.status.type !== "transcribing") return;
-    const nextTranscript = item.transcript ?? "";
-    const previousTranscript = streamTranscriptRef.current;
-    if (!nextTranscript || nextTranscript === previousTranscript) return;
-
-    if (nextTranscript.startsWith(previousTranscript)) {
-      const appended = nextTranscript
-        .slice(previousTranscript.length)
-        .replace(/^\n+/, "");
-      const cleaned = appended.trimStart();
-      if (cleaned.trim().length > 0) {
-        setStreamChunks((prev) => [...prev, cleaned]);
-      }
-    } else {
-      const cleaned = nextTranscript.trim();
-      setStreamChunks(cleaned.length > 0 ? [cleaned] : []);
-    }
-
-    streamTranscriptRef.current = nextTranscript;
-  }, [item.status.type, item.transcript]);
-
-  useEffect(() => {
-    return () => {
-      if (copyTimer.current) clearTimeout(copyTimer.current);
-    };
-  }, []);
-
-  useEffect(() => {
-    if (!transcriptAvailable) return;
-    if (transcriptTimer.current) clearTimeout(transcriptTimer.current);
-    transcriptTimer.current = setTimeout(() => {
-      if (transcriptDraft !== (item.transcript ?? "")) {
-        onUpdate({ transcript: transcriptDraft });
-      }
-    }, 600);
-    return () => {
-      if (transcriptTimer.current) clearTimeout(transcriptTimer.current);
-    };
-  }, [transcriptDraft, transcriptAvailable, item.transcript, onUpdate]);
-  useClickOutside(tagMenuRef, () => setTagMenuOpen(false), tagMenuOpen);
-
-  const handleNameCommit = async () => {
-    const value = nameDraft.trim();
-    if (!value || value === item.name) {
-      setNameDraft(item.name);
-      setIsEditingName(false);
-      return;
-    }
-    await onUpdate({ name: value });
-    setIsEditingName(false);
-  };
-
-  const handleAddTag = async (overrideTag?: string) => {
-    const value = (overrideTag ?? tagInput).trim();
-    if (!value) return;
-    if (item.tags.some((tag) => tag.toLowerCase() === value.toLowerCase())) {
-      setTagInput("");
-      return;
-    }
-    await onUpdate({ tags: [...item.tags, value] });
-    setTagInput("");
-  };
-
-  const normalizedTagInput = tagInput.trim().toLowerCase();
-  const filteredTagOptions = availableTags.filter((tag) => {
-    const tagLower = tag.toLowerCase();
-    if (item.tags.some((existing) => existing.toLowerCase() === tagLower)) {
-      return false;
-    }
-    if (!normalizedTagInput) return true;
-    return tagLower.includes(normalizedTagInput);
-  });
-
-  const handleRemoveTag = async (tag: string) => {
-    await onUpdate({ tags: item.tags.filter((entry) => entry !== tag) });
-  };
-
-  const handleExport = async (format: ExportFormat) => {
-    setIsExporting(true);
-    try {
-      const ext = format;
-      const safeName =
-        sanitizeFileName(item.name || "transcript") || "transcript";
-      const suggested = `${safeName}.${ext}`;
-      const outputPath = await save({
-        title: t({
-          id: "library.modal.export.title",
-          message: "Export transcription",
-        }),
-        defaultPath: suggested,
-        filters: [{ name: ext.toUpperCase(), extensions: [ext] }],
-      });
-      if (!outputPath) return;
-      const finalPath = outputPath.toLowerCase().endsWith(`.${ext}`)
-        ? outputPath
-        : `${outputPath}.${ext}`;
-      await onExport(format, finalPath);
-    } catch (err) {
-      const message = err instanceof Error ? err.message : String(err);
-      console.error("Export failed:", message);
-      const lower = message.toLowerCase();
-      let toastMessage =
-        message ||
-        t({
-          id: "library.modal.export.failed",
-          message: "Export failed. Try again.",
-        });
-      if (lower.includes("no timestamp segments")) {
-        toastMessage = t({
-          id: "library.modal.export.no_timestamps",
-          message:
-            "This item doesn't have timestamps. Retranscribe with timestamps to export subtitles.",
-        });
-      } else if (lower.includes("failed to write export file")) {
-        toastMessage = t({
-          id: "library.modal.export.write_failed",
-          message: "Couldn't write the export file. Try a different location.",
-        });
-      } else if (lower.includes("library item not found")) {
-        toastMessage = t({
-          id: "library.modal.export.item_not_found",
-          message: "Couldn't find this library item. Try reopening it.",
-        });
-      }
-      invoke("debug_show_toast", {
-        toastType: "error",
-        message: toastMessage,
-      }).catch(() => {});
-    } finally {
-      setIsExporting(false);
-      setExportOpen(false);
-    }
-  };
-
-  const handleCopy = async () => {
-    if (!transcriptDraft.trim()) return;
-    try {
-      await navigator.clipboard.writeText(transcriptDraft);
-      setCopyConfirmed(true);
-      if (copyTimer.current) clearTimeout(copyTimer.current);
-      copyTimer.current = setTimeout(() => {
-        setCopyConfirmed(false);
-      }, 1400);
-    } catch (err) {
-      console.error("Failed to copy transcript:", err);
-    }
-  };
-
-  const handleTogglePlayback = useCallback(() => {
-    const sound = howlRef.current;
-    if (!sound || audioError || !audioReady) return;
-    if (sound.playing()) {
-      sound.pause();
-    } else {
-      sound.play();
-    }
-  }, [audioError, audioReady]);
-
-  const handleScrubChange = (nextValue: string) => {
-    const sound = howlRef.current;
-    if (!sound || audioError || !audioReady) return;
-    const nextTime = Number(nextValue);
-    if (!Number.isFinite(nextTime)) return;
-    scrubValueRef.current = nextTime;
-    if (isScrubbing) {
-      setAudioCurrentTime(nextTime);
-      sound.seek(nextTime);
-      return;
-    }
-    sound.seek(nextTime);
-    setAudioCurrentTime(nextTime);
-  };
-
-  const handleScrubStart = () => {
-    const sound = howlRef.current;
-    if (!sound || audioError || !audioReady) return;
-    scrubWasPlayingRef.current = sound.playing();
-    updateIsScrubbing(true);
-    sound.pause();
-  };
-
-  const handleScrubEnd = () => {
-    const sound = howlRef.current;
-    if (!sound || audioError || !audioReady) return;
-    updateIsScrubbing(false);
-    if (
-      typeof scrubValueRef.current === "number" &&
-      Number.isFinite(scrubValueRef.current)
-    ) {
-      sound.seek(scrubValueRef.current);
-      setAudioCurrentTime(scrubValueRef.current);
-    }
-    scrubValueRef.current = null;
-    if (scrubWasPlayingRef.current) {
-      try {
-        sound.play();
-      } catch (err) {
-        console.error("Failed to resume audio:", err);
-        setAudioError(
-          t({
-            id: "library.modal.audio_unavailable",
-            message: "Audio unavailable",
-          }),
-        );
-      }
-    }
-    scrubWasPlayingRef.current = false;
-  };
-
-  const handleTimestampClick = (startMs: number) => {
-    const sound = howlRef.current;
-    if (!sound || audioError || !audioReady) return;
-    const nextTime = Math.max(0, startMs / 1000);
-    sound.seek(nextTime);
-    setAudioCurrentTime(nextTime);
-    if (!sound.playing()) {
-      try {
-        sound.play();
-      } catch (err) {
-        console.error("Failed to play audio:", err);
-        setAudioError(
-          t({
-            id: "library.modal.audio_unavailable",
-            message: "Audio unavailable",
-          }),
-        );
-      }
-    }
-  };
-  const scrubberMax = audioDuration > 0 ? audioDuration : 1;
-  const scrubberValue = Math.min(audioCurrentTime, scrubberMax);
-  const scrubberPercent =
-    scrubberMax > 0 ? (scrubberValue / scrubberMax) * 100 : 0;
-  const minPlaybackRate = PLAYBACK_RATES[0];
-  const maxPlaybackRate = PLAYBACK_RATES[PLAYBACK_RATES.length - 1];
-  const canDecreasePlaybackRate = playbackRate > minPlaybackRate;
-  const canIncreasePlaybackRate = playbackRate < maxPlaybackRate;
-  const showStreaming = item.status.type === "transcribing" && !showTimestamps;
-  const showSegmentView = showTimestamps && canShowTimestamps;
-  const followTimestampsActive = followTimestamps && showSegmentView;
-  const normalizedSearchQuery = searchQuery.trim();
-  const activeSegmentIndex = useMemo(() => {
-    if (!showTimestamps || !canShowTimestamps) return -1;
-    const targetMs = Math.max(0, Math.round(audioCurrentTime * 1000));
-    const segments = item.segments ?? [];
-    let match = -1;
-    for (let i = 0; i < segments.length; i += 1) {
-      if (segments[i].start_ms <= targetMs) {
-        match = i;
-        continue;
-      }
-      break;
-    }
-    return match;
-  }, [audioCurrentTime, showTimestamps, canShowTimestamps, item.segments]);
-
-  const segmentMatchIndexes = useMemo(() => {
-    if (!normalizedSearchQuery || !showSegmentView) return [];
-    const query = normalizedSearchQuery.toLowerCase();
-    const segments = item.segments ?? [];
-    const matches: number[] = [];
-    for (let i = 0; i < segments.length; i += 1) {
-      if (segments[i].text.toLowerCase().includes(query)) {
-        matches.push(i);
-      }
-    }
-    return matches;
-  }, [normalizedSearchQuery, item.segments, showSegmentView]);
-
-  const streamMatchIndexes = useMemo(() => {
-    if (!normalizedSearchQuery || !showStreaming) return [];
-    const query = normalizedSearchQuery.toLowerCase();
-    const matches: number[] = [];
-    for (let i = 0; i < streamChunks.length; i += 1) {
-      if (streamChunks[i].toLowerCase().includes(query)) {
-        matches.push(i);
-      }
-    }
-    return matches;
-  }, [normalizedSearchQuery, showStreaming, streamChunks]);
-
-  const textMatchIndex = useMemo(() => {
-    if (!normalizedSearchQuery || showSegmentView || showStreaming) return -1;
-    const query = normalizedSearchQuery.toLowerCase();
-    return transcriptDraft.toLowerCase().indexOf(query);
-  }, [normalizedSearchQuery, showSegmentView, showStreaming, transcriptDraft]);
-
-  const activeSegmentMatch = segmentMatchIndexes.length
-    ? segmentMatchIndexes[
-        Math.min(activeSearchIndex, segmentMatchIndexes.length - 1)
-      ]
-    : -1;
-  const activeStreamMatch = streamMatchIndexes.length
-    ? streamMatchIndexes[
-        Math.min(activeSearchIndex, streamMatchIndexes.length - 1)
-      ]
-    : -1;
-
-  const renderHighlightedText = useCallback(
-    (text: string, isActive: boolean) => {
-      if (!normalizedSearchQuery) return text;
-      const query = normalizedSearchQuery.toLowerCase();
-      const lower = text.toLowerCase();
-      const nodes: Array<string | ReactNode> = [];
-      let startIndex = 0;
-      let matchIndex = lower.indexOf(query);
-      let matchCount = 0;
-      if (matchIndex === -1) return text;
-      while (matchIndex !== -1) {
-        if (matchIndex > startIndex) {
-          nodes.push(text.slice(startIndex, matchIndex));
-        }
-        const matchText = text.slice(matchIndex, matchIndex + query.length);
-        nodes.push(
-          <mark
-            key={`${matchIndex}-${matchCount}`}
-            className={`transcript-search-hit${isActive ? " transcript-search-hit-active" : ""}`}
-          >
-            {matchText}
-          </mark>,
-        );
-        startIndex = matchIndex + query.length;
-        matchIndex = lower.indexOf(query, startIndex);
-        matchCount += 1;
-      }
-      if (startIndex < text.length) {
-        nodes.push(text.slice(startIndex));
-      }
-      return nodes;
-    },
-    [normalizedSearchQuery],
-  );
-
-  const handleSearchChange = (value: string) => {
-    setSearchQuery(value);
-    setActiveSearchIndex(0);
-  };
-
-  const handleSearchNavigate = useCallback(
-    (direction: number) => {
-      if (!normalizedSearchQuery) return;
-      if (showSegmentView && segmentMatchIndexes.length > 0) {
-        setActiveSearchIndex(
-          (prev) =>
-            (prev + direction + segmentMatchIndexes.length) %
-            segmentMatchIndexes.length,
-        );
-        return;
-      }
-      if (showStreaming && streamMatchIndexes.length > 0) {
-        setActiveSearchIndex(
-          (prev) =>
-            (prev + direction + streamMatchIndexes.length) %
-            streamMatchIndexes.length,
-        );
-      }
-    },
-    [
-      normalizedSearchQuery,
-      showSegmentView,
-      showStreaming,
-      segmentMatchIndexes,
-      streamMatchIndexes,
-    ],
-  );
-
-  const handleTimestampStep = useCallback(
-    (direction: number) => {
-      if (!showSegmentView) return;
-      const segments = item.segments ?? [];
-      if (segments.length === 0) return;
-      let nextIndex = activeSegmentIndex;
-      if (nextIndex < 0) {
-        nextIndex = direction > 0 ? 0 : segments.length - 1;
-      } else {
-        nextIndex = Math.max(
-          0,
-          Math.min(segments.length - 1, nextIndex + direction),
-        );
-      }
-      if (nextIndex === activeSegmentIndex) return;
-      const segment = segments[nextIndex];
-      handleTimestampClick(segment.start_ms);
-    },
-    [activeSegmentIndex, item.segments, showSegmentView, handleTimestampClick],
-  );
-
-  useEffect(() => {
-    const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.defaultPrevented) return;
-
-      const target = event.target as HTMLElement | null;
-      const tag = target?.tagName.toLowerCase();
-      const isTextInput =
-        tag === "input" || tag === "textarea" || target?.isContentEditable;
-
-      if (event.key === "Escape") {
-        event.preventDefault();
-        if (showDeleteConfirm) {
-          setShowDeleteConfirm(false);
-        } else {
-          onClose();
-        }
-        return;
-      }
-
-      if (event.key === " ") {
-        if (isTextInput) return;
-        event.preventDefault();
-        handleTogglePlayback();
-        return;
-      }
-
-      if (event.key !== "ArrowDown" && event.key !== "ArrowUp") return;
-      if (!showSegmentView || isTextInput) return;
-      const now = performance.now();
-      if (now - lastTimestampNavRef.current < 140) return;
-      lastTimestampNavRef.current = now;
-      event.preventDefault();
-      handleTimestampStep(event.key === "ArrowDown" ? 1 : -1);
-    };
-
-    window.addEventListener("keydown", handleKeyDown);
-    return () => window.removeEventListener("keydown", handleKeyDown);
-  }, [
-    handleTimestampStep,
-    handleTogglePlayback,
-    onClose,
-    showDeleteConfirm,
-    showSegmentView,
-  ]);
-
-  useEffect(() => {
-    if (!normalizedSearchQuery) return;
-    if (showSegmentView) {
-      if (segmentMatchIndexes.length === 0) return;
-      const targetIndex =
-        segmentMatchIndexes[
-          Math.min(activeSearchIndex, segmentMatchIndexes.length - 1)
-        ];
-      segmentsVirtuosoRef.current?.scrollToIndex({
-        index: targetIndex,
-        align: "center",
-        behavior: "smooth",
-      });
-      return;
-    }
-    if (showStreaming) {
-      if (streamMatchIndexes.length === 0) return;
-      const targetIndex =
-        streamMatchIndexes[
-          Math.min(activeSearchIndex, streamMatchIndexes.length - 1)
-        ];
-      streamVirtuosoRef.current?.scrollToIndex({
-        index: targetIndex,
-        align: "center",
-        behavior: "smooth",
-      });
-      return;
-    }
-    if (textMatchIndex >= 0 && transcriptAreaRef.current) {
-      const endIndex = textMatchIndex + normalizedSearchQuery.length;
-      transcriptAreaRef.current.focus();
-      transcriptAreaRef.current.setSelectionRange(textMatchIndex, endIndex);
-    }
-  }, [
-    normalizedSearchQuery,
-    showSegmentView,
-    showStreaming,
-    segmentMatchIndexes,
-    streamMatchIndexes,
-    activeSearchIndex,
-    textMatchIndex,
-  ]);
-
-  useEffect(() => {
-    if (!followTimestampsActive || activeSegmentIndex < 0) return;
-    segmentsVirtuosoRef.current?.scrollToIndex({
-      index: activeSegmentIndex,
-      align: "center",
-      behavior: "smooth",
-    });
-  }, [activeSegmentIndex, followTimestampsActive]);
-
-  return (
-    <div className="flex h-full w-full min-h-0 overflow-hidden rounded-2xl border border-[var(--color-border-secondary)] bg-[var(--color-bg-overlay)] shadow-[var(--shadow-lg)]">
-      <aside className="flex w-48 shrink-0 flex-col border-r border-[var(--color-border-primary)]">
-        <div className="px-4 pt-5 pb-3">
-          {isEditingName ? (
-            <div className="flex items-center gap-1.5 mt-1">
-              <input
-                value={nameDraft}
-                onChange={(event) => setNameDraft(event.target.value)}
-                onBlur={handleNameCommit}
-                onKeyDown={(event) => {
-                  if (event.key === "Enter") {
-                    event.preventDefault();
-                    handleNameCommit();
-                  }
-                }}
-                className="flex-1 min-w-0 bg-transparent border-b border-[var(--color-border-primary)] px-1.5 py-0.5 ui-text-body text-content-primary focus:border-[var(--color-border-hover)] outline-hidden"
-                autoFocus
-              />
-              <button
-                onClick={handleNameCommit}
-                className="text-content-muted hover:text-content-primary"
-              >
-                <Check size={12} />
-              </button>
-            </div>
-          ) : (
-            <div className="flex items-center gap-1.5 mt-1 group">
-              <h2 className="ui-text-body font-semibold text-content-primary truncate">
-                {formatLibraryName(item.name)}
-              </h2>
-              <button
-                onClick={() => setIsEditingName(true)}
-                className="opacity-0 group-hover:opacity-100 text-content-muted hover:text-content-primary transition-opacity"
-              >
-                <Pencil size={10} />
-              </button>
-            </div>
-          )}
-          <p className="ui-text-meta text-content-disabled mt-0.5 truncate">
-            {modelLabel}
-          </p>
-        </div>
-
-        <nav className="flex-1 px-2 py-2 space-y-3 overflow-y-auto custom-scrollbar scrollbar-gutter">
-          <div className="px-2">
-            <p className="ui-text-meta font-semibold text-content-disabled mb-1.5">
-              {t({
-                id: "library.modal.audio",
-                message: "Audio",
-              })}
-            </p>
-            <div className="px-1 pt-1 pb-2">
-              <div className="flex flex-col gap-1.5">
-                <div className="px-1">
-                  <input
-                    type="range"
-                    min={0}
-                    max={scrubberMax}
-                    step={0.01}
-                    value={scrubberValue}
-                    onChange={(event) => handleScrubChange(event.target.value)}
-                    onMouseDown={handleScrubStart}
-                    onTouchStart={handleScrubStart}
-                    onMouseUp={handleScrubEnd}
-                    onTouchEnd={handleScrubEnd}
-                    className="library-scrubber w-full"
-                    disabled={!audioReady || !!audioError}
-                    style={{
-                      background: `linear-gradient(to right, var(--color-toggle-on) 0%, var(--color-toggle-on) ${scrubberPercent}%, var(--color-border-secondary) ${scrubberPercent}%, var(--color-border-secondary) 100%)`,
-                    }}
-                    aria-label={t({
-                      id: "library.modal.audio_scrubber",
-                      message: "Audio scrubber",
-                    })}
-                  />
-                </div>
-
-                <div className="flex items-center justify-between px-1">
-                  <div className="flex items-center gap-2">
-                    <button
-                      onClick={handleTogglePlayback}
-                      disabled={!audioReady || !!audioError}
-                      className={`text-content-primary hover:text-content-secondary transition-colors ${
-                        !audioReady || audioError
-                          ? "opacity-50 cursor-not-allowed"
-                          : ""
-                      }`}
-                      aria-label={
-                        isPlaying
-                          ? t({
-                              id: "library.modal.pause_audio",
-                              message: "Pause audio",
-                            })
-                          : t({
-                              id: "library.modal.play_audio",
-                              message: "Play audio",
-                            })
-                      }
-                    >
-                      {isPlaying ? (
-                        <Pause size={14} className="fill-current" />
-                      ) : (
-                        <Play size={14} className="fill-current" />
-                      )}
-                    </button>
-                    <span className="ui-text-micro tabular-nums text-content-disabled font-medium tracking-wide">
-                      {formatDuration(audioCurrentTime)}{" "}
-                      <span className="opacity-50">
-                        / {formatDuration(audioDuration)}
-                      </span>
-                    </span>
-                  </div>
-
-                  <div className="flex items-center gap-0.5 ui-text-micro leading-none">
-                    <button
-                      type="button"
-                      onClick={() => handlePlaybackRateStep(-1)}
-                      disabled={
-                        !audioReady || !!audioError || !canDecreasePlaybackRate
-                      }
-                      aria-label={t({
-                        id: "library.modal.playback.decrease",
-                        message: "Decrease speed",
-                      })}
-                      className={`transition-colors p-0.5 ${
-                        !audioReady || audioError || !canDecreasePlaybackRate
-                          ? "text-content-disabled"
-                          : "text-content-muted hover:text-content-primary"
-                      }`}
-                    >
-                      <ChevronLeft size={10} />
-                    </button>
-                    <AnimatePresence mode="popLayout" initial={false}>
-                      <motion.span
-                        key={playbackRate}
-                        initial={{ opacity: 0, y: -2, scale: 0.92 }}
-                        animate={{ opacity: 1, y: 0, scale: 1 }}
-                        exit={{ opacity: 0, y: 2, scale: 0.92 }}
-                        transition={{ duration: 0.16, ease: "easeOut" }}
-                        onMouseDown={handleRateScrubStart}
-                        onTouchStart={handleRateScrubStart}
-                        className="w-[26px] min-w-[26px] text-center font-medium text-content-secondary tabular-nums cursor-ew-resize select-none"
-                      >
-                        {formatPlaybackRate(playbackRate)}x
-                      </motion.span>
-                    </AnimatePresence>
-                    <button
-                      type="button"
-                      onClick={() => handlePlaybackRateStep(1)}
-                      disabled={
-                        !audioReady || !!audioError || !canIncreasePlaybackRate
-                      }
-                      aria-label={t({
-                        id: "library.modal.playback.increase",
-                        message: "Increase speed",
-                      })}
-                      className={`transition-colors p-0.5 ${
-                        !audioReady || audioError || !canIncreasePlaybackRate
-                          ? "text-content-disabled"
-                          : "text-content-muted hover:text-content-primary"
-                      }`}
-                    >
-                      <ChevronRight size={10} />
-                    </button>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-
-          <div className="px-2 space-y-2">
-            <p className="ui-text-meta font-semibold text-content-disabled">
-              {t({
-                id: "library.modal.settings",
-                message: "Settings",
-              })}
-            </p>
-            <div className="flex items-center justify-between">
-              <div>
-                <div className="ui-text-label text-content-primary">
-                  {t({
-                    id: "library.modal.timestamps",
-                    message: "Timestamps",
-                  })}
-                </div>
-                <div className="ui-text-micro text-content-disabled">
-                  {isTranscribed
-                    ? canShowTimestamps
-                      ? t({
-                          id: "library.modal.timestamps.supported",
-                          message: "Supported",
-                        })
-                      : t({
-                          id: "library.modal.timestamps.unsupported",
-                          message: "Not supported",
-                        })
-                    : t({
-                        id: "library.modal.timestamps.available_later",
-                        message: "Available after transcription",
-                      })}
-                </div>
-              </div>
-              <ToggleSwitch
-                enabled={showTimestamps}
-                onToggle={() => {
-                  if (!canShowTimestamps) return;
-                  const nextValue = !showTimestamps;
-                  setShowTimestamps(nextValue);
-                  if (!nextValue) {
-                    onFollowTimestampsChange(false);
-                  }
-                  onUpdate({ show_timestamps: nextValue });
-                }}
-                ariaLabel={t({
-                  id: "library.modal.timestamps",
-                  message: "Timestamps",
-                })}
-                disabled={!canShowTimestamps}
-                size="md"
-              />
-            </div>
-            <div className="flex items-center justify-between pl-3">
-              <div>
-                <div className="flex items-center gap-1.5 ui-text-meta text-content-secondary">
-                  <CornerDownRight
-                    size={10}
-                    className="text-content-disabled"
-                    aria-hidden="true"
-                  />
-                  <span>
-                    {t({
-                      id: "library.modal.follow_timestamp",
-                      message: "Follow timestamp",
-                    })}
-                  </span>
-                </div>
-              </div>
-              <ToggleSwitch
-                enabled={followTimestampsActive}
-                onToggle={() => {
-                  if (!showSegmentView) return;
-                  onFollowTimestampsChange((prev) => !prev);
-                }}
-                ariaLabel={t({
-                  id: "library.modal.follow_timestamp",
-                  message: "Follow timestamp",
-                })}
-                disabled={!showSegmentView}
-                size="sm"
-              />
-            </div>
-          </div>
-
-          <div className="px-2 space-y-1.5">
-            <p className="ui-text-meta font-semibold text-content-disabled">
-              {t({
-                id: "library.modal.tags",
-                message: "Tags",
-              })}
-            </p>
-            <div className="flex flex-wrap gap-2 max-h-20 overflow-auto custom-scrollbar scrollbar-gutter pr-3">
-              {item.tags.length === 0 && (
-                <span className="ui-text-meta text-content-disabled italic">
-                  {t({
-                    id: "library.modal.tags.none",
-                    message: "None",
-                  })}
-                </span>
-              )}
-              {item.tags.map((tag, idx) => (
-                <span
-                  key={`${tag}-${idx}`}
-                  onClick={() => {
-                    if (shiftHeld) {
-                      handleRemoveTag(tag);
-                    }
-                  }}
-                  title={
-                    shiftHeld
-                      ? t({
-                          id: "library.modal.tags.remove",
-                          message: `Remove ${tag}`,
-                        })
-                      : undefined
-                  }
-                  className={`inline-flex items-center cursor-pointer ui-text-meta transition-colors duration-100 ease-out whitespace-nowrap text-content-secondary hover:text-content-primary ${
-                    shiftHeld ? "hover:!text-red-500 hover:line-through" : ""
-                  }`}
-                >
-                  <span className="opacity-40 mr-[1px]">#</span>
-                  <span>
-                    {tag.length > 12 ? `${tag.slice(0, 12)}...` : tag}
-                  </span>
-                </span>
-              ))}
-            </div>
-            <div className="flex items-center gap-1.5 min-h-[24px]">
-              <div ref={tagMenuRef} className="relative flex items-center">
-                <button
-                  type="button"
-                  onMouseDown={(event) => event.preventDefault()}
-                  onClick={() => setTagMenuOpen((prev) => !prev)}
-                  className="flex items-center justify-center w-[16px] h-[16px] shrink-0 ui-color-primary hover:text-[var(--color-text-secondary)] transition-colors"
-                  aria-label={t({
-                    id: "library.modal.tags.select_existing",
-                    message: "Select existing tag",
-                  })}
-                  title={t({
-                    id: "library.modal.tags.select_existing",
-                    message: "Select existing tag",
-                  })}
-                >
-                  <ChevronDown
-                    size={12}
-                    className={`translate-y-[1px] transition-transform duration-150 ${tagMenuOpen ? "rotate-180" : ""}`}
-                  />
-                </button>
-                <AnimatePresence>
-                  {tagMenuOpen && (
-                    <motion.div
-                      initial={{ opacity: 0, scale: 0.98, y: -4 }}
-                      animate={{ opacity: 1, scale: 1, y: 0 }}
-                      exit={{ opacity: 0, scale: 0.98, y: -4 }}
-                      transition={{ duration: 0.12 }}
-                      className="absolute left-0 top-full mt-1 z-[120] w-36 rounded-md border border-border-secondary/80 bg-surface-overlay shadow-lg shadow-black/40 overflow-hidden"
-                    >
-                      <div className="max-h-36 overflow-y-auto">
-                        {filteredTagOptions.length > 0 ? (
-                          filteredTagOptions.map((tag, index) => (
-                            <button
-                              key={`tag-option-${index}-${tag || "empty"}`}
-                              type="button"
-                              onMouseDown={(event) => event.preventDefault()}
-                              onClick={() => {
-                                handleAddTag(tag);
-                                setTagMenuOpen(false);
-                              }}
-                              className="w-full text-left px-2.5 py-1.5 ui-text-meta font-medium text-content-secondary hover:bg-surface-elevated/70 hover:text-content-primary transition-colors"
-                            >
-                              {tag}
-                            </button>
-                          ))
-                        ) : (
-                          <div className="px-2.5 py-2 ui-text-micro text-content-muted">
-                            {availableTags.length === 0
-                              ? t({
-                                  id: "library.modal.tags.no_tags_yet",
-                                  message: "No tags yet",
-                                })
-                              : t({
-                                  id: "library.modal.tags.no_other_tags",
-                                  message: "No other tags",
-                                })}
-                          </div>
-                        )}
-                      </div>
-                    </motion.div>
-                  )}
-                </AnimatePresence>
-              </div>
-              <input
-                value={tagInput}
-                onChange={(event) => setTagInput(event.target.value)}
-                onKeyDown={(event) => {
-                  if (event.key === "Enter") {
-                    event.preventDefault();
-                    handleAddTag();
-                  }
-                }}
-                placeholder={t({
-                  id: "library.modal.tags.new_tag",
-                  message: "New tag...",
-                })}
-                className="flex-1 min-w-0 h-6 bg-transparent border-b border-border-primary px-0.5 py-0 ui-text-meta text-content-secondary outline-hidden focus:border-border-hover placeholder:text-content-disabled"
-              />
-            </div>
-          </div>
-        </nav>
-
-        <div className="px-2 py-3 border-t border-[var(--color-border-primary)] space-y-1.5">
-          {(item.status.type === "transcribing" ||
-            item.status.type === "cancelling" ||
-            item.status.type === "pending" ||
-            item.status.type === "importing") && (
-            <button
-              onClick={onCancel}
-              className="w-full flex justify-center rounded-lg px-2 py-1.5 ui-text-meta text-content-secondary hover:text-content-primary hover:bg-surface-elevated transition-colors"
-            >
-              {t({
-                id: "library.modal.cancel",
-                message: "Cancel",
-              })}
-            </button>
-          )}
-          {item.status.type === "error" && (
-            <button
-              onClick={onRetry}
-              className="w-full flex items-center justify-center gap-1.5 rounded-lg px-2 py-1.5 ui-text-meta text-content-secondary hover:text-content-primary hover:bg-surface-elevated transition-colors"
-            >
-              <RotateCw size={10} />
-              {t({
-                id: "library.modal.retry",
-                message: "Retry",
-              })}
-            </button>
-          )}
-          <button
-            onClick={() => setShowDeleteConfirm(true)}
-            className="w-full flex items-center justify-center gap-1.5 rounded-lg px-2 py-1.5 ui-text-meta ui-color-error-soft hover:bg-[var(--color-error)]/10 transition-colors"
-          >
-            <Trash2 size={10} />
-            {t({
-              id: "library.modal.delete",
-              message: "Delete",
-            })}
-          </button>
-        </div>
-      </aside>
-
-      <main className="flex flex-1 flex-col min-h-0 min-w-0 bg-transparent">
-        <div className="flex items-center gap-3 px-4 py-2 border-b border-[var(--color-border-primary)] shrink-0 bg-transparent">
-          <button
-            onClick={() => setShowRetranscribe(true)}
-            disabled={
-              item.status.type === "transcribing" ||
-              item.status.type === "cancelling" ||
-              item.status.type === "pending" ||
-              item.status.type === "importing"
-            }
-            className="flex items-center gap-1.5 rounded-md px-2.5 py-1 ui-text-meta text-content-secondary hover:text-content-primary hover:bg-surface-surface disabled:opacity-40 disabled:cursor-not-allowed"
-          >
-            <RotateCw size={11} />
-            {t({
-              id: "library.modal.retranscribe",
-              message: "Retranscribe",
-            })}
-          </button>
-          <div className="flex-1 flex justify-center">
-            <div className="relative w-full max-w-[240px]">
-              <div className="relative flex items-center gap-2 bg-transparent border-b border-[var(--color-border-primary)] px-2.5 py-1.5 focus-within:border-[var(--color-border-hover)] transition-colors">
-                <Search
-                  size={12}
-                  className="text-content-disabled shrink-0"
-                  aria-hidden="true"
-                />
-                <input
-                  type="text"
-                  value={searchQuery}
-                  onChange={(event) => handleSearchChange(event.target.value)}
-                  onKeyDown={(event) => {
-                    if (event.key === "Enter") {
-                      event.preventDefault();
-                      handleSearchNavigate(event.shiftKey ? -1 : 1);
-                    }
-                    if (event.key === "Escape") {
-                      event.preventDefault();
-                      handleSearchChange("");
-                    }
-                  }}
-                  placeholder={t({
-                    id: "library.modal.search.placeholder",
-                    message: "Search transcript...",
-                  })}
-                  aria-label={t({
-                    id: "library.modal.search.aria",
-                    message: "Search transcript",
-                  })}
-                  className="bg-transparent ui-text-label text-content-secondary placeholder-content-disabled outline-hidden w-full"
-                />
-                <div className="absolute right-2 top-1/2 -translate-y-1/2 flex items-center">
-                  {searchQuery && (
-                    <button
-                      onClick={() => handleSearchChange("")}
-                      aria-label={t({
-                        id: "library.modal.search.clear",
-                        message: "Clear search",
-                      })}
-                      className="text-content-disabled hover:text-content-muted transition-colors"
-                    >
-                      <X size={12} aria-hidden="true" />
-                    </button>
-                  )}
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="flex items-center gap-2">
-            <button
-              onClick={handleCopy}
-              disabled={!transcriptAvailable}
-              className={`flex items-center gap-1.5 rounded-md px-2.5 py-1 ui-text-meta disabled:opacity-50 transition-colors ${
-                copyConfirmed
-                  ? "ui-color-success bg-[color-mix(in_srgb,var(--color-success)_12%,transparent)]"
-                  : "text-content-secondary hover:text-content-primary hover:bg-surface-surface"
-              }`}
-            >
-              {copyConfirmed ? <Check size={10} /> : <Copy size={10} />}
-              <span className="inline-block min-w-[38px] text-left">
-                {copyConfirmed
-                  ? t({
-                      id: "library.modal.copy.copied",
-                      message: "Copied",
-                    })
-                  : t({
-                      id: "library.modal.copy",
-                      message: "Copy",
-                    })}
-              </span>
-            </button>
-            <div className="relative">
-              <button
-                onClick={() => setExportOpen(!exportOpen)}
-                disabled={isExporting || !transcriptAvailable}
-                className="flex items-center gap-1.5 rounded-md px-2.5 py-1 ui-text-meta text-content-secondary hover:text-content-primary hover:bg-surface-surface disabled:opacity-50"
-              >
-                {t({
-                  id: "library.modal.export",
-                  message: "Export",
-                })}
-                <ChevronDown size={10} />
-              </button>
-              <AnimatePresence>
-                {exportOpen && (
-                  <motion.div
-                    initial={{ opacity: 0, y: 4 }}
-                    animate={{ opacity: 1, y: 0 }}
-                    exit={{ opacity: 0, y: 4 }}
-                    transition={{ duration: 0.1 }}
-                    className="absolute right-0 top-full mt-1 w-32 rounded-lg border border-[var(--color-border-secondary)] bg-[var(--color-bg-overlay)] shadow-xl overflow-hidden z-[120]"
-                  >
-                    {(["txt", "md", "srt", "vtt"] as ExportFormat[]).map(
-                      (format) => {
-                        const requiresSegments =
-                          format === "srt" || format === "vtt";
-                        const disabled =
-                          requiresSegments &&
-                          !(item.segments && item.segments.length);
-                        return (
-                          <button
-                            key={format}
-                            onClick={() => handleExport(format)}
-                            disabled={disabled}
-                            className="w-full px-3 py-1.5 text-left ui-text-meta text-content-secondary hover:bg-surface-overlay disabled:opacity-40 disabled:cursor-not-allowed"
-                          >
-                            {format.toUpperCase()}
-                          </button>
-                        );
-                      },
-                    )}
-                  </motion.div>
-                )}
-              </AnimatePresence>
-            </div>
-            <button
-              onClick={onClose}
-              className="flex items-center justify-center rounded-md p-1.5 text-content-muted hover:text-content-primary hover:bg-surface-surface"
-              aria-label={t({
-                id: "library.modal.close",
-                message: "Close",
-              })}
-            >
-              <X size={12} />
-            </button>
-          </div>
-        </div>
-
-        <div className="flex-1 min-h-0 overflow-hidden px-4 pb-0 pt-0 flex flex-col gap-3">
-          {item.status.type === "error" ? (
-            <div className="flex-1 min-h-0 flex items-center justify-center">
-              {(() => {
-                const details = getLibraryErrorDetails(item.status.message);
-                return (
-                  <div className="max-w-[240px] rounded-xl border border-red-500/30 bg-red-500/10 px-4 py-3 text-center">
-                    <div className="flex items-center justify-center gap-2 ui-color-error-tint">
-                      <AlertTriangle size={14} />
-                      <span className="ui-text-label font-medium">
-                        {t({
-                          id: "library.modal.import_failed",
-                          message: "Import failed",
-                        })}
-                      </span>
-                    </div>
-                    <p className="mt-2 ui-text-meta leading-[14px] ui-color-error-tint select-text cursor-text">
-                      {details.message}
-                    </p>
-                    {details.showFfmpegHelp && (
-                      <button
-                        type="button"
-                        onClick={() =>
-                          invoke("open_ffmpeg_install").catch(() => {})
-                        }
-                        className="mt-2 ui-text-meta ui-color-error-faint underline decoration-red-400/60 ui-hover-error-50"
-                      >
-                        {t({
-                          id: "library.modal.ffmpeg_help",
-                          message: "FFmpeg Help",
-                        })}
-                      </button>
-                    )}
-                  </div>
-                );
-              })()}
-            </div>
-          ) : (
-            <div className="relative flex-1 min-h-0">
-              <div
-                className="pointer-events-none absolute left-0 right-3 top-0 h-5 z-10"
-                style={{
-                  background:
-                    "linear-gradient(to bottom, var(--color-bg-overlay), transparent)",
-                }}
-                aria-hidden="true"
-              />
-              <div
-                className="pointer-events-none absolute left-0 right-3 bottom-0 h-6 z-10"
-                style={{
-                  background:
-                    "linear-gradient(to top, var(--color-bg-overlay), transparent)",
-                }}
-                aria-hidden="true"
-              />
-              {showSegmentView ? (
-                <Virtuoso
-                  ref={segmentsVirtuosoRef}
-                  style={{ height: "100%" }}
-                  data={item.segments ?? []}
-                  overscan={200}
-                  className="custom-scrollbar ui-text-body text-content-secondary leading-relaxed"
-                  computeItemKey={(index: number, segment: TranscriptSegment) =>
-                    `${segment.start_ms}-${index}`
-                  }
-                  components={{
-                    Header: () => <div className="h-3" />,
-                    Footer: () => <div className="h-3" />,
-                  }}
-                  itemContent={(idx, segment) => {
-                    const isActive = idx === activeSegmentIndex;
-                    return (
-                      <div className="pb-1.5 pr-4">
-                        <motion.div
-                          initial={{ opacity: 0, y: 6 }}
-                          animate={{ opacity: 1, y: 0 }}
-                          transition={{ duration: 0.2, ease: "easeOut" }}
-                          className="grid w-full grid-cols-[auto_1fr] gap-3 rounded-md px-2 py-1 transition-colors select-none border"
-                          style={{
-                            backgroundColor: isActive
-                              ? "var(--color-interactive-10)"
-                              : "transparent",
-                            borderColor: isActive
-                              ? "var(--color-interactive-30)"
-                              : "transparent",
-                          }}
-                        >
-                          <span
-                            className="text-content-disabled font-mono ui-text-label pt-0.5 select-none cursor-pointer hover:text-content-primary"
-                            role="button"
-                            tabIndex={0}
-                            onClick={() =>
-                              handleTimestampClick(segment.start_ms)
-                            }
-                            onKeyDown={(event) => {
-                              if (event.key === "Enter" || event.key === " ") {
-                                event.preventDefault();
-                                handleTimestampClick(segment.start_ms);
-                              }
-                            }}
-                          >
-                            {formatTimestamp(segment.start_ms)}
-                          </span>
-                          <div className="min-w-0 select-none w-fit">
-                            <span className="select-text">
-                              {renderHighlightedText(
-                                segment.text,
-                                idx === activeSegmentMatch,
-                              )}
-                            </span>
-                          </div>
-                        </motion.div>
-                      </div>
-                    );
-                  }}
-                />
-              ) : showStreaming ? (
-                streamChunks.length === 0 ? (
-                  <div className="flex flex-col h-full w-full items-center justify-center gap-5">
-                    <IntelligencePixel active size="md" />
-                    <div className="ui-text-label font-medium text-content-disabled">
-                      {t({
-                        id: "library.modal.transcribing",
-                        message: "Transcribing...",
-                      })}
-                    </div>
-                  </div>
-                ) : (
-                  <Virtuoso
-                    ref={streamVirtuosoRef}
-                    style={{ height: "100%" }}
-                    data={streamChunks}
-                    overscan={200}
-                    className="custom-scrollbar ui-text-body text-content-secondary leading-relaxed"
-                    computeItemKey={(index: number) =>
-                      `${item.id}-chunk-${index}`
-                    }
-                    components={{
-                      Header: () => <div className="h-3" />,
-                      Footer: () => <div className="h-3" />,
-                    }}
-                    itemContent={(idx, chunk) => (
-                      <div className="pb-2 pr-4">
-                        <motion.p
-                          initial={{ opacity: 0, y: 6 }}
-                          animate={{ opacity: 1, y: 0 }}
-                          transition={{ duration: 0.2, ease: "easeOut" }}
-                          className="select-text"
-                        >
-                          {renderHighlightedText(
-                            chunk,
-                            idx === activeStreamMatch,
-                          )}
-                        </motion.p>
-                      </div>
-                    )}
-                  />
-                )
-              ) : item.status.type === "importing" ||
-                item.status.type === "pending" ? (
-                <div className="flex flex-col h-full w-full items-center justify-center gap-5">
-                  <IntelligencePixel active size="md" />
-                  <div className="ui-text-label font-medium text-content-disabled">
-                    {importStatusText}
-                  </div>
-                </div>
-              ) : (
-                <textarea
-                  ref={transcriptAreaRef}
-                  value={transcriptDraft}
-                  onChange={(event) => setTranscriptDraft(event.target.value)}
-                  disabled={!transcriptAvailable}
-                  placeholder={t({
-                    id: "library.modal.transcript_placeholder",
-                    message: "Transcript will appear here.",
-                  })}
-                  className="h-full w-full resize-none bg-transparent ui-text-body text-content-secondary leading-relaxed outline-hidden disabled:opacity-60 custom-scrollbar select-text pr-4 pt-3 pb-3"
-                />
-              )}
-            </div>
-          )}
-        </div>
-      </main>
-
-      <AnimatePresence>
-        {showDeleteConfirm && (
-          <motion.div
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            exit={{ opacity: 0 }}
-            className="fixed inset-0 z-[100] flex items-center justify-center bg-black/70 backdrop-blur-xs px-6"
-            onClick={(event) => {
-              event.stopPropagation();
-              setShowDeleteConfirm(false);
-            }}
-          >
-            <motion.div
-              initial={{ scale: 0.96, opacity: 0 }}
-              animate={{ scale: 1, opacity: 1 }}
-              exit={{ scale: 0.96, opacity: 0 }}
-              transition={{ duration: 0.18 }}
-              className="w-full max-w-sm rounded-2xl border border-border-primary bg-surface-tertiary p-5 ui-shadow-modal-deep"
-              onClick={(event) => event.stopPropagation()}
-              role="dialog"
-              aria-modal="true"
-            >
-              <div className="flex items-center gap-3 mb-3">
-                <AlertTriangle
-                  size={20}
-                  className="ui-color-warning-strong shrink-0"
-                />
-                <div>
-                  <p className="ui-text-body-lg font-semibold text-content-primary">
-                    {t({
-                      id: "library.modal.delete_confirm.title",
-                      message: "Delete this item?",
-                    })}
-                  </p>
-                  <p className="ui-text-label text-content-disabled">
-                    {t({
-                      id: "library.modal.delete_confirm.description",
-                      message:
-                        "This removes the transcript and audio from your library.",
-                    })}
-                  </p>
-                </div>
-              </div>
-              <div className="flex justify-end gap-2">
-                <button
-                  onClick={() => setShowDeleteConfirm(false)}
-                  className="rounded-lg border border-border-secondary px-4 py-2 ui-text-body-sm font-medium text-content-secondary hover:border-border-hover transition-colors"
-                >
-                  {t({
-                    id: "library.modal.cancel",
-                    message: "Cancel",
-                  })}
-                </button>
-                <button
-                  onClick={() => {
-                    setShowDeleteConfirm(false);
-                    onDelete();
-                  }}
-                  className="rounded-lg bg-red-500/90 px-4 py-2 ui-text-body-sm font-semibold ui-color-on-solid hover:bg-red-500 transition-colors"
-                >
-                  {t({
-                    id: "library.modal.delete",
-                    message: "Delete",
-                  })}
-                </button>
-              </div>
-            </motion.div>
-          </motion.div>
-        )}
-      </AnimatePresence>
-
-      <AnimatePresence>
-        {showRetranscribe && (
-          <LibraryRetranscribeModal
-            item={item}
-            models={models}
-            onCancel={() => setShowRetranscribe(false)}
-            onConfirm={async (options) => {
-              try {
-                await onUpdate({
-                  speech_model: options.model_key,
-                  llm_cleanup_enabled: false,
-                  show_timestamps: options.show_timestamps,
-                });
-                await onRetry();
-                setShowRetranscribe(false);
-              } catch (err) {
-                console.error("Failed to retranscribe:", err);
-              }
-            }}
-          />
-        )}
-      </AnimatePresence>
-    </div>
-  );
-};
-
-export default LibraryModal;
