+++ src/features/library/components/LibraryView.tsx
@@ -1,6 +1,6 @@
 import { useLingui } from "@lingui/react/macro";
-import { useCallback, useEffect, useMemo, useState } from "react";
-import { AnimatePresence } from "framer-motion";
+import { useCallback, useMemo, useState } from "react";
+import { AnimatePresence, motion } from "framer-motion";
 import { invoke } from "@tauri-apps/api/core";
 import { open } from "@tauri-apps/plugin-dialog";
 import {
@@ -22,7 +22,7 @@ import {
 } from "../../settings/models-queries";
 import LibraryImportModal from "./LibraryImportModal";
 import LibraryCard from "./LibraryCard";
-import LibraryModal from "./LibraryModal";
+import LibraryDetail from "./LibraryDetail";
 import {
   useLibraryItems as useLibraryItemsQuery,
   useCreateLibraryItem,
@@ -51,14 +51,12 @@ import type {
 type LibraryViewProps = {
   pendingImportPaths: string[] | null;
   onSetImportPaths: (paths: string[] | null) => void;
-  sidebarWidth: number;
   isActive: boolean;
 };
 
 const LibraryView = ({
   pendingImportPaths,
   onSetImportPaths,
-  sidebarWidth,
   isActive,
 }: LibraryViewProps) => {
   const { t } = useLingui();
@@ -166,20 +164,6 @@ const LibraryView = ({
     onComplete: refreshSpeechModels,
     onError: refreshSpeechModels,
   });
-
-  useEffect(() => {
-    if (!selectedItem) return;
-
-    const handleSidebarClick = (event: MouseEvent) => {
-      const sidebar = document.querySelector("[data-app-sidebar]");
-      if (sidebar && sidebar.contains(event.target as Node)) {
-        setSelectedItemId(null);
-      }
-    };
-
-    document.addEventListener("mousedown", handleSidebarClick);
-    return () => document.removeEventListener("mousedown", handleSidebarClick);
-  }, [selectedItem]);
 
   const handleImportClick = async () => {
     try {
@@ -295,237 +279,228 @@ const LibraryView = ({
 
   return (
     <div className="relative flex h-full min-h-0 min-w-0 flex-1 flex-col">
-      <div className="mx-auto flex w-full max-w-7xl min-w-0 flex-col gap-4 pb-4 px-0 text-left">
-        <div className="flex flex-col gap-4 mb-4 mt-2 md:-mt-6">
-          <ScreenHeader
-            icon={
-              <DotMatrix
-                rows={2}
-                cols={3}
-                activeDots={[0, 1, 2, 4]}
-                dotSize={3}
-                gap={3}
-                color="var(--color-section-marker-alt)"
-              />
+      {selectedItem ? (
+        <motion.div
+          key={selectedItem.id || "selected-library-item"}
+          initial={{ opacity: 0, y: 8 }}
+          animate={{ opacity: 1, y: 0 }}
+          transition={{ duration: 0.18, ease: "easeOut" }}
+          className="flex h-full min-h-0 flex-col"
+        >
+          <LibraryDetail
+            item={selectedItem}
+            models={installedModels}
+            shiftHeld={shiftHeld}
+            followTimestamps={followTimestamps}
+            onFollowTimestampsChange={setFollowTimestamps}
+            onClose={() => setSelectedItemId(null)}
+            onDelete={async () => {
+              await deleteItemAndRefreshTags(selectedItem.id);
+              setSelectedItemId(null);
+            }}
+            onRetry={() => retryMutation.mutateAsync(selectedItem.id)}
+            onCancel={() => cancelMutation.mutateAsync(selectedItem.id)}
+            onUpdate={(patch) => updateItemWithTags(selectedItem.id, patch)}
+            onExport={(format, outputPath) =>
+              exportMutation.mutateAsync({
+                id: selectedItem.id,
+                format,
+                outputPath,
+              })
             }
-            title={t({ id: "library.view.title", message: "Library" })}
-            description={t({
-              id: "library.view.description",
-              message:
-                "Manage and search your transcribed audio and documents.",
-            })}
+            availableTags={availableTags}
           />
-
-          <div className="grid grid-cols-1 gap-3 md:grid-cols-[auto_minmax(14rem,1fr)_auto] md:items-center">
-            <button
-              onClick={handleImportClick}
-              className="flex items-center gap-2 rounded-lg border border-[var(--color-border-primary)] bg-[var(--color-bg-surface)] px-3 py-1.5 ui-text-body-sm ui-color-primary hover:border-[var(--color-border-secondary)] hover:bg-[var(--color-bg-overlay)] transition-colors shrink-0"
-            >
-              <Plus size={14} />
-              {t({ id: "library.view.import_button", message: "Import" })}
-            </button>
-
-            <div className="relative min-w-0 w-full group">
-              <Search
-                size={14}
-                className="absolute left-3 top-1/2 -translate-y-1/2 ui-color-muted transition-colors"
-              />
-              <input
-                type="text"
-                placeholder={t({
-                  id: "library.view.search_placeholder",
-                  message: "Search library... or use #tag",
-                })}
-                value={searchQuery}
-                onChange={(e) => setSearchQuery(e.target.value)}
-                className="w-full bg-[var(--color-bg-surface)] border border-[var(--color-border-primary)] rounded-lg focus:border-[var(--color-border-hover)] pl-9 pr-4 py-1.5 ui-text-input ui-color-primary placeholder-[var(--color-text-muted)] outline-none transition-all duration-100 ease-out"
-              />
-            </div>
-
-            <SegmentedControl
-              value={statusFilterValue}
-              options={statusFilterOptions}
-              onChange={(value) =>
-                setStatusFilter(value === "active" ? "transcribing" : value)
-              }
-              ariaLabel={t({
-                id: "library.filter.aria_label",
-                message: "Filter library by status",
-              })}
-              className="relative flex w-full items-center rounded-lg border border-[var(--color-border-primary)] bg-[var(--color-bg-secondary)] p-1 md:w-auto"
-              activeIndicatorLayoutId="library-status-filter"
-            />
-          </div>
-        </div>
-
-        {error && (
-          <div className="rounded-lg border border-[var(--color-error)]/30 bg-[var(--color-error)]/10 px-4 py-3 ui-text-body-sm ui-color-error-tint mx-4 mb-2">
-            {error}
-          </div>
-        )}
-      </div>
-      <div className="flex-1 min-h-0 overflow-y-scroll overflow-x-hidden custom-scrollbar scrollbar-gutter pb-6 pr-3 pt-1">
-        <div key="library-list" className="flex flex-col gap-6 w-full">
-          <div className="mx-auto flex w-full max-w-6xl min-w-0 flex-col gap-6">
-            <div className="grid min-w-0 gap-4 grid-cols-[repeat(auto-fit,minmax(min(100%,180px),1fr))]">
-              {isLoading && items.length === 0 && (
-                <div className="col-span-full py-12 flex items-center justify-center">
+        </motion.div>
+      ) : (
+        <>
+          <div className="mx-auto flex w-full max-w-7xl min-w-0 flex-col gap-4 pt-8 pb-4 px-0 text-left">
+            <div className="flex flex-col gap-4 mb-4 mt-2 md:-mt-6">
+              <ScreenHeader
+                icon={
                   <DotMatrix
                     rows={2}
-                    cols={8}
-                    activeDots={[0, 1, 2, 3, 4, 5, 6, 7]}
+                    cols={3}
+                    activeDots={[0, 1, 2, 4]}
                     dotSize={3}
                     gap={3}
-                    color="var(--color-text-muted)"
-                    animated
-                    className="opacity-50"
+                    color="var(--color-section-marker-alt)"
+                  />
+                }
+                title={t({ id: "library.view.title", message: "Library" })}
+                description={t({
+                  id: "library.view.description",
+                  message:
+                    "Manage and search your transcribed audio and documents.",
+                })}
+              />
+
+              <div className="grid grid-cols-1 gap-3 md:grid-cols-[auto_minmax(14rem,1fr)_auto] md:items-center">
+                <button
+                  onClick={handleImportClick}
+                  className="flex items-center gap-2 rounded-lg border border-[var(--color-border-primary)] bg-[var(--color-bg-surface)] px-3 py-1.5 ui-text-body-sm ui-color-primary hover:border-[var(--color-border-secondary)] hover:bg-[var(--color-bg-overlay)] transition-colors shrink-0"
+                >
+                  <Plus size={14} />
+                  {t({ id: "library.view.import_button", message: "Import" })}
+                </button>
+
+                <div className="relative min-w-0 w-full group">
+                  <Search
+                    size={14}
+                    className="absolute left-3 top-1/2 -translate-y-1/2 ui-color-muted transition-colors"
+                  />
+                  <input
+                    type="text"
+                    placeholder={t({
+                      id: "library.view.search_placeholder",
+                      message: "Search library... or use #tag",
+                    })}
+                    value={searchQuery}
+                    onChange={(e) => setSearchQuery(e.target.value)}
+                    className="w-full bg-[var(--color-bg-surface)] border border-[var(--color-border-primary)] rounded-lg focus:border-[var(--color-border-hover)] pl-9 pr-4 py-1.5 ui-text-input ui-color-primary placeholder-[var(--color-text-muted)] outline-none transition-all duration-100 ease-out"
                   />
                 </div>
-              )}
 
-              {!isLoading && items.length === 0 && (
-                <button
-                  type="button"
-                  onClick={handleImportClick}
-                  className="col-span-full rounded-xl border border-dashed border-border-secondary bg-surface-secondary p-8 flex flex-col items-center justify-center text-center hover:text-content-secondary hover:border-border-hover transition-colors"
-                >
-                  <FolderOpen size={20} className="text-content-disabled" />
-                  <p className="mt-3 ui-text-body ui-color-muted">
-                    {t({
-                      id: "library.view.empty_state",
-                      message: "Click to import files and build your Library.",
-                    })}
-                  </p>
-                </button>
-              )}
-
-              {items.map((item, index) => (
-                <LibraryCard
-                  key={item.id || `library-item-${index}`}
-                  item={item}
-                  onOpen={() => setSelectedItemId(item.id)}
-                  onRemoveTag={async (tag) => {
-                    const nextTags = item.tags.filter((entry) => entry !== tag);
-                    await updateItemWithTags(item.id, { tags: nextTags });
-                  }}
-                  onClickTag={(tag) => setSearchQuery(`#${tag}`)}
-                  editingNameId={editingNameId}
-                  editingNameDraft={editingNameDraft}
-                  onStartNameEdit={() => startNameEdit(item)}
-                  onChangeNameDraft={setEditingNameDraft}
-                  onCommitNameEdit={() => commitNameEdit(item.id)}
-                  onCancelNameEdit={cancelNameEdit}
-                  onRetry={() => retryMutation.mutateAsync(item.id)}
-                  onCancel={() => cancelMutation.mutateAsync(item.id)}
-                  onDelete={async () => {
-                    await deleteItemAndRefreshTags(item.id);
-                    if (selectedItem?.id === item.id) {
-                      setSelectedItemId(null);
-                    }
-                  }}
-                  editingTagId={editingTagId}
-                  tagDraft={tagDraft}
-                  onStartTagEdit={() => startTagEdit(item)}
-                  onChangeTagDraft={setTagDraft}
-                  onCommitTagAdd={(value) => commitTagAdd(item.id, value)}
-                  onCancelTagEdit={cancelTagEdit}
-                  shiftHeld={shiftHeld}
-                  availableTags={availableTags}
+                <SegmentedControl
+                  value={statusFilterValue}
+                  options={statusFilterOptions}
+                  onChange={(value) =>
+                    setStatusFilter(value === "active" ? "transcribing" : value)
+                  }
+                  ariaLabel={t({
+                    id: "library.filter.aria_label",
+                    message: "Filter library by status",
+                  })}
+                  className="relative flex w-full items-center rounded-lg border border-[var(--color-border-primary)] bg-[var(--color-bg-secondary)] p-1 md:w-auto"
+                  activeIndicatorLayoutId="library-status-filter"
                 />
-              ))}
+              </div>
+            </div>
 
-              {items.length > 0 && (
-                <button
-                  onClick={handleImportClick}
-                  className="rounded-xl border border-dashed border-border-secondary bg-surface-secondary p-4 flex flex-col items-center justify-center text-center ui-color-muted hover:text-content-secondary hover:border-border-hover transition-colors"
-                >
-                  <FolderOpen size={18} />
-                  <span className="mt-2 ui-text-body-sm">
-                    {t({
-                      id: "library.view.dropzone",
-                      message: "Click to import files",
-                    })}
-                  </span>
-                </button>
-              )}
+            {error && (
+              <div className="rounded-lg border border-[var(--color-error)]/30 bg-[var(--color-error)]/10 px-4 py-3 ui-text-body-sm ui-color-error-tint mx-4 mb-2">
+                {error}
+              </div>
+            )}
+          </div>
+          <div className="flex-1 min-h-0 overflow-y-scroll overflow-x-hidden custom-scrollbar scrollbar-gutter pb-6 pr-3 pt-1">
+            <div key="library-list" className="flex flex-col gap-6 w-full">
+              <div className="mx-auto flex w-full max-w-6xl min-w-0 flex-col gap-6">
+                <div className="grid min-w-0 gap-4 grid-cols-[repeat(auto-fit,minmax(min(100%,180px),1fr))]">
+                  {isLoading && items.length === 0 && (
+                    <div className="col-span-full py-12 flex items-center justify-center">
+                      <DotMatrix
+                        rows={2}
+                        cols={8}
+                        activeDots={[0, 1, 2, 3, 4, 5, 6, 7]}
+                        dotSize={3}
+                        gap={3}
+                        color="var(--color-text-muted)"
+                        animated
+                        className="opacity-50"
+                      />
+                    </div>
+                  )}
 
-              {items.length > 0 && hasNextPage && (
-                <div className="col-span-full flex items-center justify-center pt-2">
-                  <button
-                    onClick={() => fetchNextPage()}
-                    disabled={isFetchingNextPage}
-                    className="flex items-center gap-2 rounded-lg border border-border-primary bg-surface-surface px-4 py-2 ui-text-body-sm ui-color-secondary hover:text-content-primary hover:border-border-secondary hover:bg-surface-overlay transition-colors disabled:opacity-60 disabled:cursor-not-allowed"
-                  >
-                    {isFetchingNextPage ? (
-                      <>
-                        <Loader2 size={14} className="animate-spin" />
-                        <span>
-                          {t({
-                            id: "library.view.loading_more",
-                            message: "Loading...",
-                          })}
-                        </span>
-                      </>
-                    ) : (
-                      <span>
+                  {!isLoading && items.length === 0 && (
+                    <button
+                      type="button"
+                      onClick={handleImportClick}
+                      className="col-span-full rounded-xl border border-dashed border-border-secondary bg-surface-secondary p-8 flex flex-col items-center justify-center text-center hover:text-content-secondary hover:border-border-hover transition-colors"
+                    >
+                      <FolderOpen size={20} className="text-content-disabled" />
+                      <p className="mt-3 ui-text-body ui-color-muted">
                         {t({
-                          id: "library.view.load_more",
-                          message: "Load more",
+                          id: "library.view.empty_state",
+                          message:
+                            "Click to import files and build your Library.",
+                        })}
+                      </p>
+                    </button>
+                  )}
+
+                  {items.map((item, index) => (
+                    <LibraryCard
+                      key={item.id || `library-item-${index}`}
+                      item={item}
+                      onOpen={() => setSelectedItemId(item.id)}
+                      onRemoveTag={async (tag) => {
+                        const nextTags = item.tags.filter(
+                          (entry) => entry !== tag,
+                        );
+                        await updateItemWithTags(item.id, { tags: nextTags });
+                      }}
+                      onClickTag={(tag) => setSearchQuery(`#${tag}`)}
+                      editingNameId={editingNameId}
+                      editingNameDraft={editingNameDraft}
+                      onStartNameEdit={() => startNameEdit(item)}
+                      onChangeNameDraft={setEditingNameDraft}
+                      onCommitNameEdit={() => commitNameEdit(item.id)}
+                      onCancelNameEdit={cancelNameEdit}
+                      onRetry={() => retryMutation.mutateAsync(item.id)}
+                      onCancel={() => cancelMutation.mutateAsync(item.id)}
+                      onDelete={async () => {
+                        await deleteItemAndRefreshTags(item.id);
+                      }}
+                      editingTagId={editingTagId}
+                      tagDraft={tagDraft}
+                      onStartTagEdit={() => startTagEdit(item)}
+                      onChangeTagDraft={setTagDraft}
+                      onCommitTagAdd={(value) => commitTagAdd(item.id, value)}
+                      onCancelTagEdit={cancelTagEdit}
+                      shiftHeld={shiftHeld}
+                      availableTags={availableTags}
+                    />
+                  ))}
+
+                  {items.length > 0 && (
+                    <button
+                      onClick={handleImportClick}
+                      className="rounded-xl border border-dashed border-border-secondary bg-surface-secondary p-4 flex flex-col items-center justify-center text-center ui-color-muted hover:text-content-secondary hover:border-border-hover transition-colors"
+                    >
+                      <FolderOpen size={18} />
+                      <span className="mt-2 ui-text-body-sm">
+                        {t({
+                          id: "library.view.dropzone",
+                          message: "Click to import files",
                         })}
                       </span>
-                    )}
-                  </button>
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
-      </div>
+                    </button>
+                  )}
 
-      <AnimatePresence>
-        {selectedItem && (
-          <div
-            key="library-detail"
-            className="fixed top-0 right-0 bottom-0 z-20 flex items-center justify-center p-6 transition-[left] duration-200 ease-out"
-            style={{
-              left: sidebarWidth,
-              backdropFilter: "blur(12px)",
-              WebkitBackdropFilter: "blur(12px)",
-            }}
-            onClick={() => setSelectedItemId(null)}
-          >
-            <div
-              className="w-full h-full min-h-0"
-              onClick={(e) => e.stopPropagation()}
-            >
-              <LibraryModal
-                key={selectedItem.id || "selected-library-item"}
-                item={selectedItem}
-                models={installedModels}
-                shiftHeld={shiftHeld}
-                followTimestamps={followTimestamps}
-                onFollowTimestampsChange={setFollowTimestamps}
-                onClose={() => setSelectedItemId(null)}
-                onDelete={async () => {
-                  await deleteItemAndRefreshTags(selectedItem.id);
-                  setSelectedItemId(null);
-                }}
-                onRetry={() => retryMutation.mutateAsync(selectedItem.id)}
-                onCancel={() => cancelMutation.mutateAsync(selectedItem.id)}
-                onUpdate={(patch) => updateItemWithTags(selectedItem.id, patch)}
-                onExport={(format, outputPath) =>
-                  exportMutation.mutateAsync({
-                    id: selectedItem.id,
-                    format,
-                    outputPath,
-                  })
-                }
-                availableTags={availableTags}
-              />
+                  {items.length > 0 && hasNextPage && (
+                    <div className="col-span-full flex items-center justify-center pt-2">
+                      <button
+                        onClick={() => fetchNextPage()}
+                        disabled={isFetchingNextPage}
+                        className="flex items-center gap-2 rounded-lg border border-border-primary bg-surface-surface px-4 py-2 ui-text-body-sm ui-color-secondary hover:text-content-primary hover:border-border-secondary hover:bg-surface-overlay transition-colors disabled:opacity-60 disabled:cursor-not-allowed"
+                      >
+                        {isFetchingNextPage ? (
+                          <>
+                            <Loader2 size={14} className="animate-spin" />
+                            <span>
+                              {t({
+                                id: "library.view.loading_more",
+                                message: "Loading...",
+                              })}
+                            </span>
+                          </>
+                        ) : (
+                          <span>
+                            {t({
+                              id: "library.view.load_more",
+                              message: "Load more",
+                            })}
+                          </span>
+                        )}
+                      </button>
+                    </div>
+                  )}
+                </div>
+              </div>
             </div>
           </div>
-        )}
-      </AnimatePresence>
+        </>
+      )}
 
       <AnimatePresence>
         {pendingImportPaths !== null && (
