+++ src/locales/en/messages.js
@@ -1,1 +1,1 @@
-/*eslint-disable*/module.exports={messages:JSON.parse("{\"D-waf4\":[[\"seconds\",\"plural\",{\"one\":[\"#\",\" second spoken today\"],\"other\":[\"#\",\" seconds spoken today\"]}]],\"HQ7Iit\":[[\"0\",\"plural\",{\"one\":[\"#\",\" word\"],\"other\":[\"#\",\" words\"]}]],\"Vebvqp\":[[\"0\",\"plural\",{\"one\":[\"Longest dictation today: \",\"#\",\" word\"],\"other\":[\"Longest dictation today: \",\"#\",\" words\"]}]],\"ane_compile.body\":[\"macOS is compiling \",[\"label\"],\" for the Apple Neural Engine. This happens once and can take a few minutes.\"],\"ane_compile.title\":[\"Optimizing for the Neural Engine\"],\"dictionary.add_prompt\":[\"Press Enter to add \\\"\",[\"0\"],\"\\\" as a new entry.\"],\"dictionary.combined.description\":[\"Add custom words the system should recognize, and set automatic word replacements.\"],\"dictionary.combined.title\":[\"Dictionary & Replacements\"],\"dictionary.delete\":[\"Delete\"],\"dictionary.delete_entry\":[\"Delete \",[\"entry\"]],\"dictionary.edit_hint\":[\"Press Enter to save · Esc to cancel\"],\"dictionary.entry_count.multiple\":[[\"0\"],\" entries\"],\"dictionary.entry_count.single\":[\"1 entry\"],\"dictionary.full_add_prompt\":[\"Delete an entry before adding another.\"],\"dictionary.full_hint\":[\"Dictionary is full\"],\"dictionary.no_entries\":[\"No entries yet\"],\"dictionary.no_entries.description\":[\"Add words, phrases, or names above and press Enter to save them here.\"],\"dictionary.no_matches\":[\"No matches found\"],\"dictionary.press_enter_to_add\":[\"Press Enter to add\"],\"dictionary.press_enter_to_add_match\":[\"Press Enter to add this word\"],\"dictionary.replacements.count.multiple\":[[\"0\"],\" replacements\"],\"dictionary.replacements.count.single\":[\"1 replacement\"],\"dictionary.replacements.delete\":[\"Delete replacement for \",[\"0\"]],\"dictionary.replacements.find\":[\"Find word...\"],\"dictionary.replacements.find_aria\":[\"Find word to replace\"],\"dictionary.replacements.none\":[\"No replacements yet\"],\"dictionary.replacements.none_description\":[\"Add word pairs to automatically swap in transcriptions. Matches are case-insensitive.\"],\"dictionary.replacements.press_enter_to_add\":[\"Press Enter in either field to add\"],\"dictionary.replacements.remove_value\":[\"remove\"],\"dictionary.replacements.replace_with\":[\"Replace with...\"],\"dictionary.replacements.replace_with_aria\":[\"Replace with\"],\"dictionary.search_matches\":[[\"0\"],\" of \",[\"1\"],\" matches\"],\"dictionary.search_only\":[\"Search dictionary...\"],\"dictionary.search_or_add\":[\"Search or add a word...\"],\"dictionary.search_or_add_aria\":[\"Add or search dictionary entry\"],\"dictionary.section.dictionary_description\":[\"Add custom words Glimpse should recognize.\"],\"dictionary.section.dictionary_title\":[\"Dictionary\"],\"dictionary.section.replacements_description\":[\"Swap common phrases automatically after transcription.\"],\"dictionary.section.replacements_title\":[\"Replacements\"],\"dictionary.warning\":[\"Dictionary works only for models with dictionary support. Current model \",[\"0\"],\" will ignore these entries until you switch to a compatible model.\"],\"dictionary.warning_aria\":[\"Warning: model compatibility issue\"],\"dropdown.no_options\":[\"No options found\"],\"dropdown.placeholder\":[\"Select...\"],\"dropdown.search_aria\":[\"Search options\"],\"dropdown.search_placeholder\":[\"Search...\"],\"dropdown.toggle_menu\":[\"Toggle options\"],\"faq.ai_writing.answer\":[\"Only if you enable AI writing and set up a provider under Settings → Providers. Cleanup, Edit Mode, and Personalization then send the relevant text directly to that provider. Your API key stays stored locally in Glimpse.\"],\"faq.ai_writing.question\":[\"When does text leave my device?\"],\"faq.close_aria\":[\"Close FAQ\"],\"faq.delete.answer\":[\"Delete recordings from History, remove imported files from Library, or uninstall models from Settings → Models. Settings → App can auto-delete Audio only or full Transcripts (including linked audio). You can also open your Glimpse data folder from About → Storage to manage files directly.\"],\"faq.delete.question\":[\"How do I manage or delete my data?\"],\"faq.free.answer\":[\"Core dictation is free: local transcription, dictionary, replacements, and history. There are no per-minute fees or subscriptions for that. Library, AI Cleanup, Edit Mode, personalization with an LLM, the local API server, and the CLI are part of Glimpse Personal. You get a 14-day trial first; after that, activate a one-time Personal or Commercial license in Settings → Account.\"],\"faq.free.question\":[\"What is free vs Glimpse Personal?\"],\"faq.how_it_works.answer\":[\"Press your dictation shortcut, speak, and release. Glimpse transcribes on your device and inserts the text where your cursor is. Core dictation works offline. No account required.\"],\"faq.how_it_works.question\":[\"How does Glimpse work?\"],\"faq.permissions.answer\":[\"Microphone access to record your voice, and Accessibility access to insert text and read selected text for Edit Mode. Glimpse only uses these while you are actively dictating.\"],\"faq.permissions.question\":[\"What permissions does Glimpse need?\"],\"faq.privacy.answer\":[\"Your audio and transcripts stay on your Mac. Glimpse does not collect recordings, transcripts, API keys, or prompts. Optional anonymous usage analytics (things like session length and feature usage, never your content) help us improve the app. You can turn this off anytime in Settings → App.\"],\"faq.privacy.link\":[\"Privacy policy\"],\"faq.privacy.question\":[\"Where does my data go?\"],\"faq.subtitle\":[\"How Glimpse works, privacy, and AI writing\"],\"faq.title\":[\"Frequently Asked Questions\"],\"g960RG\":[[\"0\",\"plural\",{\"one\":[\"#\",\" dictation\"],\"other\":[\"#\",\" dictations\"]}]],\"h_9Xcr\":[[\"0\",\"plural\",{\"one\":[\"#\",\" dictation polished with AI today\"],\"other\":[\"#\",\" dictations polished with AI today\"]}]],\"home.drag_import.eyebrow\":[\"Library Import\"],\"home.drag_import.subtitle\":[\"MP3, WAV, M4A, MP4, MOV, and more\"],\"home.drag_import.title\":[\"Drop files to transcribe\"],\"home.greeting.afternoon\":[\"Good afternoon\"],\"home.greeting.evening\":[\"Good evening\"],\"home.greeting.morning\":[\"Good morning\"],\"home.local_api.copy_url\":[\"Copy base URL\"],\"home.local_api.open_settings\":[\"Open API server settings\"],\"home.local_api.requests\":[\"Requests served\"],\"home.local_api.running\":[\"Local API\"],\"home.local_api.status_hint\":[\"Your API server is running in the background.\"],\"home.mode.cloud\":[\"Cloud\"],\"home.mode.local\":[\"Local\"],\"home.sidebar.collapse\":[\"Collapse sidebar\"],\"home.sidebar.dictionary\":[\"Dictionary\"],\"home.sidebar.expand\":[\"Expand sidebar\"],\"home.sidebar.home\":[\"Home\"],\"home.sidebar.library\":[\"Library\"],\"home.sidebar.personalization\":[\"Personalization\"],\"home.sidebar.settings\":[\"Settings\"],\"home.support.about.title\":[\"About\"],\"home.support.about.version_mode\":[\"v\",[\"version\"],\" • \",[\"mode\"]],\"home.support.faq.subtitle\":[\"Common questions\"],\"home.support.faq.title\":[\"FAQ\"],\"home.support.feedback.email\":[\"Email\"],\"home.support.feedback.email_copied\":[\"Copied!\"],\"home.support.feedback.github\":[\"GitHub Issue\"],\"home.support.feedback.title\":[\"Feedback\"],\"home.support.label\":[\"Support\"],\"home.support.menu_aria\":[\"Support menu\"],\"home.support.title\":[\"Get Support\"],\"home.today.stats\":[[\"0\",\"plural\",{\"one\":[\"#\",\" dictation\"],\"other\":[\"#\",\" dictations\"]}],\" · \",[\"1\",\"plural\",{\"one\":[\"#\",\" word\"],\"other\":[\"#\",\" words\"]}],\" today\"],\"home.update_available\":[\"Update available\"],\"import.app_picker.aria\":[\"Select app to import from\"],\"import.apply.failed\":[\"Import failed. Try again or skip.\"],\"import.cat.auto_launch\":[\"Launch at login\"],\"import.cat.auto_launch.off\":[\"Off\"],\"import.cat.auto_launch.on\":[\"On\"],\"import.cat.dictionary\":[\"Dictionary\"],\"import.cat.dictionary.detail\":[[\"0\",\"plural\",{\"one\":[\"#\",\" word\"],\"other\":[\"#\",\" words\"]}]],\"import.cat.history\":[\"Transcript history\"],\"import.cat.history.detail\":[[\"0\"],\" transcripts\"],\"import.cat.language\":[\"Language\"],\"import.cat.model\":[\"Transcription model\"],\"import.cat.personalities\":[\"Personalities\"],\"import.cat.personalities.detail\":[[\"0\"],\" saved\"],\"import.cat.replacements\":[\"Text replacements\"],\"import.cat.replacements.detail\":[[\"0\",\"plural\",{\"one\":[\"#\",\" rule\"],\"other\":[\"#\",\" rules\"]}]],\"import.cat.shortcut\":[\"Keyboard shortcut\"],\"import.cta\":[\"Import\"],\"import.empty\":[\"Nothing to bring over from this app.\"],\"import.error\":[\"We couldn't read this app's data. You can skip and set things up manually.\"],\"import.importing\":[\"Importing...\"],\"import.model.unrecognized\":[\"We don't recognize this app's model — you'll pick one on the next step.\"],\"import.skip\":[\"Skip for now\"],\"import.subtitle\":[[\"0\",\"plural\",{\"one\":[\"We found another dictation app. Choose what to import.\"],\"other\":[\"We found other dictation apps. Choose what to import.\"]}]],\"import.subtitle.single\":[\"We noticed you use \",[\"sourceName\"],\". Pick what to carry over.\"],\"import.title\":[\"Want to bring your settings over?\"],\"library.card.cancel\":[\"Cancel\"],\"library.card.delete\":[\"Delete\"],\"library.card.new_tag\":[\"New tag...\"],\"library.card.no_other_tags\":[\"No other tags\"],\"library.card.no_tags_yet\":[\"No tags yet\"],\"library.card.remove_tag\":[\"Remove \",[\"tag\"]],\"library.card.rename\":[\"Rename\"],\"library.card.retranscribe\":[\"Retranscribe\"],\"library.card.retry\":[\"Retry\"],\"library.card.select_existing_tag\":[\"Select existing tag\"],\"library.delete_error.default\":[\"Failed to delete the library item.\"],\"library.delete_error.default_fallback\":[\"Failed to delete the library item.\"],\"library.delete_error.delete_files_failed\":[\"Couldn't delete the library files. Check permissions and try again.\"],\"library.delete_error.invalid_path\":[\"Couldn't delete this item due to an invalid file path.\"],\"library.delete_error.outside_folder\":[\"Couldn't delete this item because its files are outside the library folder.\"],\"library.delete_error.storage_missing\":[\"Couldn't delete this item. Library storage couldn't be found.\"],\"library.error_details.copy_original_failed\":[\"Couldn't copy original file.\"],\"library.error_details.create_storage_failed\":[\"Couldn't create library storage.\"],\"library.error_details.ffmpeg_required\":[\"FFmpeg required for video imports.\"],\"library.error_details.file_not_found\":[\"File not found.\"],\"library.error_details.import_failed\":[\"Import failed.\"],\"library.error_details.insufficient_disk_space\":[\"Not enough disk space.\"],\"library.error_details.invalid_audio\":[\"Not a valid audio file.\"],\"library.error_details.model_not_installed\":[\"Model not installed.\"],\"library.error_details.no_audio_track\":[\"No audio track found.\"],\"library.error_details.unsupported_audio_settings\":[\"Unsupported audio settings.\"],\"library.error_details.unsupported_format\":[\"Unsupported file format.\"],\"library.filter.active\":[\"Active\"],\"library.filter.all\":[\"All\"],\"library.filter.aria_label\":[\"Filter library by status\"],\"library.filter.done\":[\"Done\"],\"library.filter.failed\":[\"Failed\"],\"library.import.cancel\":[\"Cancel\"],\"library.import.confirm\":[\"Import\"],\"library.import.files\":[\"Files\"],\"library.import.importing\":[\"Importing...\"],\"library.import.model\":[\"Model\"],\"library.import.no_models\":[\"No local models are installed. Download a model in Settings -> Models before importing.\"],\"library.import.remote_provider\":[\"Remote provider\"],\"library.import.search_models\":[\"Search installed models...\"],\"library.import.select_model\":[\"Select a model\"],\"library.import.show_timestamps\":[\"Show timestamps\"],\"library.import.show_timestamps.aria\":[\"Show timestamps\"],\"library.import.store_original\":[\"Store original file\"],\"library.import.store_original.aria\":[\"Store original\"],\"library.import.store_original.description\":[\"Keep a copy inside the library folder\"],\"library.import.summary.multiple\":[[\"0\"],\" files\"],\"library.import.summary.single\":[\"1 file\"],\"library.import.timestamps_supported\":[\"Enabled for supported models\"],\"library.import.timestamps_unsupported\":[\"Not supported by this model\"],\"library.import.title\":[\"Import to Library\"],\"library.import_error.audio_conversion_failed\":[\"Couldn't convert this file to audio for transcription.\"],\"library.import_error.copy_original_failed\":[\"Couldn't copy the original file into the library.\"],\"library.import_error.create_storage_failed\":[\"Couldn't create library storage. Check disk permissions.\"],\"library.import_error.decode_failed\":[\"Couldn't decode this audio file. Try installing FFmpeg.\"],\"library.import_error.default\":[\"Import failed for one of the files.\"],\"library.import_error.default_fallback\":[\"Import failed for one of the files.\"],\"library.import_error.file_not_found\":[\"File not found. It may have moved or been deleted.\"],\"library.import_error.model_not_installed\":[\"Selected model isn't installed. Download one in Settings -> Models.\"],\"library.import_error.no_audio_track\":[\"No audio track found in this file.\"],\"library.import_error.unsupported_format\":[\"Unsupported file format.\"],\"library.import_error.unsupported_sample_rate\":[\"This file has an unsupported sample rate.\"],\"library.modal.audio\":[\"Audio\"],\"library.modal.audio_scrubber\":[\"Audio scrubber\"],\"library.modal.audio_unavailable\":[\"Audio unavailable\"],\"library.modal.cancel\":[\"Cancel\"],\"library.modal.close\":[\"Close\"],\"library.modal.copy\":[\"Copy\"],\"library.modal.copy.copied\":[\"Copied\"],\"library.modal.delete\":[\"Delete\"],\"library.modal.delete_confirm.description\":[\"This removes the transcript and audio from your library.\"],\"library.modal.delete_confirm.title\":[\"Delete this item?\"],\"library.modal.export\":[\"Export\"],\"library.modal.export.failed\":[\"Export failed. Try again.\"],\"library.modal.export.item_not_found\":[\"Couldn't find this library item. Try reopening it.\"],\"library.modal.export.no_timestamps\":[\"This item doesn't have timestamps. Retranscribe with timestamps to export subtitles.\"],\"library.modal.export.title\":[\"Export transcription\"],\"library.modal.export.write_failed\":[\"Couldn't write the export file. Try a different location.\"],\"library.modal.ffmpeg_help\":[\"FFmpeg Help\"],\"library.modal.follow_timestamp\":[\"Follow timestamp\"],\"library.modal.import_failed\":[\"Import failed\"],\"library.modal.import_status.converting\":[\"Converting audio...\"],\"library.modal.import_status.converting_progress\":[\"Converting audio... \",[\"0\"],\"%\"],\"library.modal.import_status.queued\":[\"Queued for transcription...\"],\"library.modal.pause_audio\":[\"Pause audio\"],\"library.modal.play_audio\":[\"Play audio\"],\"library.modal.playback.decrease\":[\"Decrease playback speed\"],\"library.modal.playback.increase\":[\"Increase playback speed\"],\"library.modal.retranscribe\":[\"Retranscribe\"],\"library.modal.retry\":[\"Retry\"],\"library.modal.search.aria\":[\"Search transcript\"],\"library.modal.search.clear\":[\"Clear search\"],\"library.modal.search.placeholder\":[\"Search transcript...\"],\"library.modal.settings\":[\"Settings\"],\"library.modal.tags\":[\"Tags\"],\"library.modal.tags.new_tag\":[\"New tag...\"],\"library.modal.tags.no_other_tags\":[\"No other tags\"],\"library.modal.tags.no_tags_yet\":[\"No tags yet\"],\"library.modal.tags.none\":[\"None\"],\"library.modal.tags.remove\":[\"Remove \",[\"tag\"]],\"library.modal.tags.select_existing\":[\"Select existing tag\"],\"library.modal.timestamps\":[\"Timestamps\"],\"library.modal.timestamps.available_later\":[\"Available after transcription\"],\"library.modal.timestamps.supported\":[\"Supported\"],\"library.modal.timestamps.unsupported\":[\"Not supported\"],\"library.modal.transcribing\":[\"Transcribing...\"],\"library.modal.transcript_placeholder\":[\"Transcript will appear here.\"],\"library.retranscribe.cancel\":[\"Cancel\"],\"library.retranscribe.confirm\":[\"Retranscribe\"],\"library.retranscribe.loading\":[\"Retranscribing...\"],\"library.retranscribe.model\":[\"Model\"],\"library.retranscribe.no_models\":[\"No local models are installed. Download a model in Settings -> Models before retranscribing.\"],\"library.retranscribe.remote_provider\":[\"Remote provider\"],\"library.retranscribe.search_models\":[\"Search installed models...\"],\"library.retranscribe.select_model\":[\"Select a model\"],\"library.retranscribe.show_timestamps\":[\"Show timestamps\"],\"library.retranscribe.timestamps_supported\":[\"Enabled for supported models\"],\"library.retranscribe.timestamps_unsupported\":[\"Not supported by this model\"],\"library.retranscribe.title\":[\"Retranscribe\"],\"library.view.description\":[\"Import audio and video files for transcription.\"],\"library.view.dropzone\":[\"Drop files to import\"],\"library.view.empty_state\":[\"Drag files here to build your Library.\"],\"library.view.file_filter\":[\"Audio & Video\"],\"library.view.import_button\":[\"Import\"],\"library.view.import_dialog_error\":[\"Could not open the import dialog.\"],\"library.view.load_more\":[\"Load more\"],\"library.view.loading_more\":[\"Loading...\"],\"library.view.search_placeholder\":[\"Search library...\"],\"library.view.title\":[\"Library\"],\"library.view.unsupported_files\":[[\"0\"],\" file(s) skipped due to unsupported format.\"],\"license.customer_portal\":[\"Customer portal\"],\"member_card.aria\":[\"Glimpse member card\"],\"member_card.coverage_with_devices\":[[\"coverageBase\"],\" · \",[\"activeDevices\"],\" of \",[\"deviceLimit\"],\" devices active\"],\"member_card.draft_aria\":[\"Draft Glimpse member card\"],\"member_card.draft_idle\":[\"Pick a license\"],\"member_card.draft_stamp_empty\":[\"Unissued\"],\"member_card.draft_stamp_issuing\":[\"Issuing\"],\"member_card.label_member_since\":[\"Member since\"],\"member_card.label_words_spoken\":[\"Words spoken\"],\"member_card.tier_commercial\":[\"Commercial\"],\"member_card.tier_contributor\":[\"Contributor\"],\"member_card.tier_founder\":[\"Founder\"],\"member_card.tier_personal\":[\"Personal\"],\"member_card.tier_purchase_aria\":[\"Purchase \",[\"label\"],\" for \",[\"price\"]],\"model_picker.active\":[\"Active\"],\"model_picker.ane.download\":[\"Download Neural Engine encoder\"],\"model_picker.ane.info\":[\"Adds a Core ML encoder that runs on the Apple Neural Engine for faster, more power-efficient transcription. The first load takes longer while macOS optimizes it for your chip.\"],\"model_picker.ane.info_aria\":[\"About the Apple Neural Engine encoder\"],\"model_picker.ane.installed\":[\"Neural Engine encoder installed\"],\"model_picker.ane.toggle\":[\"Include the Apple Neural Engine encoder\"],\"model_picker.cancel\":[\"Cancel\"],\"model_picker.cancelled\":[\"Cancelled\"],\"model_picker.category.experimental\":[\"Experimental\"],\"model_picker.category.legacy\":[\"Legacy\"],\"model_picker.category.standard\":[\"Standard\"],\"model_picker.close\":[\"Close\"],\"model_picker.delete\":[\"Delete\"],\"model_picker.download\":[\"Download\"],\"model_picker.english\":[\"English\"],\"model_picker.filter.all\":[\"All models\"],\"model_picker.filter.aria\":[\"Filter models by category\"],\"model_picker.multilingual\":[\"Multilingual\"],\"model_picker.no_results\":[\"No models match your search.\"],\"model_picker.quant\":[\"Quantization\"],\"model_picker.search\":[\"Search models\"],\"model_picker.search_aria\":[\"Search models\"],\"model_picker.title\":[\"Choose a model\"],\"model_picker.use\":[\"Use\"],\"model_picker.variant\":[\"Model variant\"],\"model_picker.variant_installed\":[\"Model variant (installed)\"],\"models.card.aria\":[[\"0\"],\" model\"],\"models.card.cancel\":[\"Cancel\"],\"models.card.cancel_download\":[\"Cancel download\"],\"models.card.delete\":[\"Delete\"],\"models.card.delete_model\":[\"Delete model\"],\"models.card.download\":[\"Download\"],\"models.card.download_model\":[\"Download model\"],\"models.card.downloading\":[\"Downloading\"],\"models.card.verifying\":[\"Verifying install\"],\"myYTaz\":[\"Trial · \",[\"trialDaysRemaining\"],\" of \",[\"TRIAL_TOTAL_DAYS\"],\" days left\"],\"onboarding.accessibility.app_name\":[\"Glimpse\"],\"onboarding.accessibility.continue\":[\"Continue\"],\"onboarding.accessibility.enable\":[\"Enable in Settings\"],\"onboarding.accessibility.instructions\":[\"Click below to open System Settings, then toggle on \",[\"appName\"]],\"onboarding.accessibility.skip\":[\"Skip\"],\"onboarding.accessibility.subtitle\":[\"Enables auto-paste into any application.\"],\"onboarding.accessibility.title\":[\"Accessibility\"],\"onboarding.back\":[\"Back\"],\"onboarding.brand\":[\"Glimpse\"],\"onboarding.complete.failed\":[\"Could not finish setup. Check your settings and try again.\"],\"onboarding.complete.no_model\":[\"Could not load a local model selection. Try reopening onboarding.\"],\"onboarding.delete.failed\":[\"Delete failed\"],\"onboarding.download.failed\":[\"Download failed\"],\"onboarding.download.starting\":[\"starting...\"],\"onboarding.license.activate.input_aria\":[\"Activation code\"],\"onboarding.license.activate.placeholder\":[\"GLIMPSE_...\"],\"onboarding.license.activate.submit\":[\"Activate\"],\"onboarding.license.active_subtitle\":[\"This Mac is licensed and ready.\"],\"onboarding.license.analytics_notice\":[\"Anonymous usage analytics help improve Glimpse. Turn this off anytime in Settings → App.\"],\"onboarding.license.continue_active\":[\"Continue\"],\"onboarding.license.continue_trial\":[\"Continue with trial\"],\"onboarding.license.saving\":[\"Saving...\"],\"onboarding.license.subtitle\":[\"Buy once, or activate the code from your receipt. You can skip this and keep using the trial.\"],\"onboarding.license.title\":[\"Connect your license\"],\"onboarding.microphone.continue\":[\"Continue\"],\"onboarding.microphone.grant\":[\"Grant Access\"],\"onboarding.microphone.skip\":[\"Skip\"],\"onboarding.microphone.subtitle\":[\"Required to capture your voice for transcription.\"],\"onboarding.microphone.title\":[\"Microphone Access\"],\"onboarding.models.action.delete_aria\":[\"Delete \",[\"modelLabel\"]],\"onboarding.models.action.download_aria\":[\"Download \",[\"modelLabel\"]],\"onboarding.models.action.stop_download_aria\":[\"Stop downloading \",[\"modelLabel\"]],\"onboarding.models.cancelled\":[\"Cancelled\"],\"onboarding.models.confirm_without_model.body\":[\"You haven't downloaded a local model yet. Transcription will not run offline until you add one in Settings.\"],\"onboarding.models.confirm_without_model.continue\":[\"Continue anyway\"],\"onboarding.models.confirm_without_model.stay\":[\"Stay here\"],\"onboarding.models.confirm_without_model.title\":[\"Continue without a model?\"],\"onboarding.models.continue\":[\"Continue\"],\"onboarding.models.download\":[\"Download\"],\"onboarding.models.download_failed\":[\"Download failed\"],\"onboarding.models.downloaded\":[\"Downloaded\"],\"onboarding.models.empty_body\":[\"This build did not return any local models. You can continue setup and manage models later in Settings.\"],\"onboarding.models.empty_title\":[\"No local models found\"],\"onboarding.models.loading_body\":[\"Fetching the available local transcription engines for this build.\"],\"onboarding.models.loading_cta\":[\"Loading...\"],\"onboarding.models.loading_title\":[\"Loading local models\"],\"onboarding.models.select_aria\":[\"Select \",[\"modelLabel\"]],\"onboarding.models.subtitle\":[\"More models and language model setup available in Settings after setup.\"],\"onboarding.models.title\":[\"Choose your local model\"],\"onboarding.models.unavailable_body\":[\"Glimpse couldn't load the local model list. Setup can continue with the default local engine, and you can manage downloads later in Settings.\"],\"onboarding.models.unavailable_title\":[\"Model list unavailable\"],\"onboarding.permissions.continue\":[\"Continue\"],\"onboarding.permissions.skip\":[\"Skip\"],\"onboarding.permissions.subtitle\":[\"Glimpse needs these to record your voice and put text where your cursor is.\"],\"onboarding.permissions.title\":[\"Permissions\"],\"onboarding.ready.analytics_notice\":[\"Glimpse sends anonymous usage analytics to help improve the app. You can disable this anytime in Settings -> App.\"],\"onboarding.ready.continue\":[\"Continue\"],\"onboarding.ready.smart.badge\":[\"Default\"],\"onboarding.ready.smart.capture_aria\":[\"Record new shortcut for Smart, currently \",[\"shortcut\"]],\"onboarding.ready.smart.capture_help_active\":[\"Press your new shortcut, or hit Esc to cancel.\"],\"onboarding.ready.smart.capture_help_idle\":[\"Click the shortcut to change it.\"],\"onboarding.ready.smart.capture_prompt\":[\"Press new shortcut...\"],\"onboarding.ready.smart.description\":[\"tap to toggle, hold to talk\"],\"onboarding.ready.smart.label\":[\"Smart\"],\"onboarding.ready.subtitle\":[\"Smart is on by default, others available in settings.\"],\"onboarding.ready.title\":[\"You're ready!\"],\"onboarding.setup.analytics_notice\":[\"Glimpse sends anonymous usage analytics to help improve the app. You can disable this anytime in Settings -> App.\"],\"onboarding.setup.back\":[\"Back\"],\"onboarding.setup.confirm_without_model.body\":[\"Transcription will not run offline until you download a local model.\"],\"onboarding.setup.confirm_without_model.continue\":[\"Continue anyway\"],\"onboarding.setup.confirm_without_model.stay\":[\"Stay here\"],\"onboarding.setup.confirm_without_model.title\":[\"Continue without a model?\"],\"onboarding.setup.continue\":[\"Continue\"],\"onboarding.setup.language.aria\":[\"Choose primary dictation language\"],\"onboarding.setup.language.english\":[\"English\"],\"onboarding.setup.language.multilingual\":[\"Multiple languages\"],\"onboarding.setup.language.placeholder\":[\"choose a language\"],\"onboarding.setup.language.question\":[\"Which language do you dictate in?\"],\"onboarding.setup.model.cancelled\":[\"Cancelled\"],\"onboarding.setup.model.download\":[\"Download\"],\"onboarding.setup.model.download_failed\":[\"Download failed\"],\"onboarding.setup.model.empty_body\":[\"You can continue setup and manage local models later in Settings.\"],\"onboarding.setup.model.empty_title\":[\"No local models found\"],\"onboarding.setup.model.loading_body\":[\"Glimpse is checking the models available in this build.\"],\"onboarding.setup.model.loading_title\":[\"Finding your local model\"],\"onboarding.setup.model.recommendation_label\":[\"Glimpse recommends\"],\"onboarding.setup.model.remove\":[\"Remove\"],\"onboarding.setup.model.stop\":[\"Stop\"],\"onboarding.setup.model.unavailable_body\":[\"Model list unavailable. You can manage local models later in Settings.\"],\"onboarding.setup.model.unavailable_title\":[\"Model list unavailable\"],\"onboarding.setup.pick_specific\":[\"Browse all models\"],\"onboarding.setup.priority.aria\":[\"Choose local model priority\"],\"onboarding.setup.priority.balanced\":[\"Balanced\"],\"onboarding.setup.priority.balanced.helper\":[\"Accuracy & size\"],\"onboarding.setup.priority.compact\":[\"Smallest\"],\"onboarding.setup.priority.compact.helper\":[\"Less storage\"],\"onboarding.setup.priority.custom\":[\"Custom\"],\"onboarding.setup.priority.placeholder\":[\"choose a priority\"],\"onboarding.setup.priority.quality\":[\"Best quality\"],\"onboarding.setup.priority.quality.helper\":[\"Most accurate\"],\"onboarding.setup.priority.question\":[\"What should your model prioritize?\"],\"onboarding.setup.sentence.language_prefix\":[\"I primarily dictate in\"],\"onboarding.setup.sentence.priority_prefix\":[\"and I want\"],\"onboarding.setup.sentence.shortcut_prefix\":[\"I'll use\"],\"onboarding.setup.sentence.shortcut_suffix\":[\"to dictate.\"],\"onboarding.setup.shortcut.active_help\":[\"Press the keys you want, or hit Esc to cancel.\"],\"onboarding.setup.shortcut.aria\":[\"Record new shortcut, currently \",[\"0\"]],\"onboarding.setup.shortcut.idle_help\":[\"Click the shortcut if you want to change it.\"],\"onboarding.setup.shortcut.label\":[\"Dictate with\"],\"onboarding.setup.shortcut.prompt\":[\"press a shortcut\"],\"onboarding.setup.subtitle\":[\"Answer a couple of plain questions and Glimpse will pick the local model.\"],\"onboarding.setup.title\":[\"Set up Glimpse\"],\"onboarding.status.checking\":[\"Checking...\"],\"onboarding.status.enabled\":[\"Enabled\"],\"onboarding.status.not_enabled\":[\"Not enabled\"],\"onboarding.welcome.cloud.badge\":[\"In progress\"],\"onboarding.welcome.cloud.title\":[\"Cloud\"],\"onboarding.welcome.cta\":[\"Continue\"],\"onboarding.welcome.local.aria\":[\"Select Local mode: privacy-first, on-device transcription\"],\"onboarding.welcome.local.feature.byok\":[\"Bring your own API key for AI features\"],\"onboarding.welcome.local.feature.free\":[\"Free, no account required\"],\"onboarding.welcome.local.feature.models\":[\"On-device transcription models\"],\"onboarding.welcome.local.title\":[\"Local\"],\"onboarding.welcome.subtitle\":[\"Choose how you want to transcribe.\"],\"onboarding.welcome.title\":[\"Welcome\"],\"personalization.apps\":[\"Apps\"],\"personalization.cancel\":[\"Cancel\"],\"personalization.delete\":[\"Delete\"],\"personalization.delete_mode.description\":[\"Delete \\\"\",[\"0\"],\"\\\"? This cannot be undone.\"],\"personalization.delete_mode.title\":[\"Delete mode?\"],\"personalization.description\":[\"Tailor language model behavior to apps, sites, and custom instructions.\"],\"personalization.disable_mode\":[\"Disable mode\"],\"personalization.empty.description\":[\"Create a mode to start customizing your apps and websites.\"],\"personalization.empty.title\":[\"No modes yet\"],\"personalization.enable_mode\":[\"Enable mode\"],\"personalization.modal.add\":[\"Add\"],\"personalization.modal.applications\":[\"Applications\"],\"personalization.modal.applications.add\":[\"Add application\"],\"personalization.modal.applications.none\":[\"No applications selected\"],\"personalization.modal.applications.not_installed\":[\"Not installed\"],\"personalization.modal.applications.search\":[\"Search applications...\"],\"personalization.modal.applications.selected\":[[\"0\"],\" selected\"],\"personalization.modal.applications.toggle_list\":[\"Toggle application list\"],\"personalization.modal.close\":[\"Close modal\"],\"personalization.modal.custom_instructions\":[\"Custom instructions\"],\"personalization.modal.custom_instructions.drag\":[\"Drag to resize\"],\"personalization.modal.custom_instructions.placeholder\":[\"Add custom instructions\"],\"personalization.modal.custom_instructions.resize\":[\"Resize custom instructions\"],\"personalization.modal.delete_mode\":[\"Delete mode\"],\"personalization.modal.edit_name\":[\"Edit mode name\"],\"personalization.modal.header\":[\"Personalization\"],\"personalization.modal.remove\":[\"Remove\"],\"personalization.modal.remove_app\":[\"Remove \",[\"app\"]],\"personalization.modal.remove_site\":[\"Remove \",[\"site\"]],\"personalization.modal.save_name\":[\"Save name\"],\"personalization.modal.snippets.full_list\":[\"Full snippet list\"],\"personalization.modal.snippets.info\":[\"Show personalization snippet examples\"],\"personalization.modal.snippets.summary\":[\"Use snippets to pass live context to the language model, like\"],\"personalization.modal.website.duplicate\":[\"That domain is already added\"],\"personalization.modal.website.invalid\":[\"Enter a valid domain like gmail.com\"],\"personalization.modal.websites\":[\"Websites\"],\"personalization.modal.websites.aria\":[\"Add website domain\"],\"personalization.modal.websites.count\":[[\"0\"],\" sites\"],\"personalization.modal.websites.none\":[\"No websites added\"],\"personalization.modal.websites.placeholder\":[\"Add a site like gmail.com\"],\"personalization.new_mode\":[\"New mode\"],\"personalization.new_mode.default_name\":[\"New Mode\"],\"personalization.no_apps\":[\"No apps yet\"],\"personalization.no_notes\":[\"No notes yet\"],\"personalization.no_sites\":[\"No sites yet\"],\"personalization.notes\":[\"Notes:\"],\"personalization.title\":[\"Personalization\"],\"personalization.websites\":[\"Websites\"],\"pill.background_visualizer\":[\"Background audio visualizer\"],\"pill.status.error\":[\"Error occurred\"],\"pill.status.listening\":[\"Listening...\"],\"pill.status.processing\":[\"Processing...\"],\"pill.visualizer\":[\"Audio visualizer\"],\"qhrhL0\":[[\"minutes\",\"plural\",{\"one\":[\"#\",\" minute spoken today\"],\"other\":[\"#\",\" minutes spoken today\"]}]],\"settings.about.advanced\":[\"Advanced\"],\"settings.about.cli.default_info\":[\"Installs the \",[\"0\"],\" command for Terminal, scripts, and automation tools. Use it when you want to call Glimpse programmatically without opening the app UI.\"],\"settings.about.cli.default_subtitle\":[\"Use Glimpse from Terminal or scripts\"],\"settings.about.cli.installed_info\":[\"The glimpse command is installed at \",[\"cliInstallPath\"],\". Use it from Terminal, scripts, or automation tools to call Glimpse without opening the app UI.\"],\"settings.about.cli.installed_subtitle\":[\"Installed at \",[\"cliInstallPath\"]],\"settings.about.cli.locked_info\":[\"Command line install requires an active license.\"],\"settings.about.cli.locked_subtitle\":[\"Requires an active license\"],\"settings.about.cli.path_missing_info\":[\"Installs \",[\"0\"],\" to \",[\"cliInstallPath\"],\". That folder is not currently on your shell PATH, so you may need to call it by full path or update your shell profile.\"],\"settings.about.cli.unavailable_info\":[\"This build does not include the command line helper.\"],\"settings.about.cli.unavailable_subtitle\":[\"Not available in this build\"],\"settings.about.command_line\":[\"Command line\"],\"settings.about.command_line.info_aria\":[\"More information about command line tools\"],\"settings.about.command_line.open_wiki_aria\":[\"Open the command line documentation\"],\"settings.about.faq_help\":[\"FAQ & Help\"],\"settings.about.faq_help_description\":[\"common questions\"],\"settings.about.install_cli\":[\"Install CLI\"],\"settings.about.mode.cloud\":[\"Cloud\"],\"settings.about.mode.local\":[\"Local\"],\"settings.about.restart_onboarding\":[\"Restart Onboarding\"],\"settings.about.restart_onboarding_description\":[\"re-run setup wizard\"],\"settings.about.restart_onboarding_hold_aria\":[\"Restart Onboarding. Hold to confirm.\"],\"settings.about.setup\":[\"Setup\"],\"settings.about.storage\":[\"Storage\"],\"settings.about.storage.database\":[\"Database\"],\"settings.about.storage.library\":[\"Library\"],\"settings.about.storage.models\":[\"Models\"],\"settings.about.storage.recordings\":[\"Recordings\"],\"settings.about.storage.total\":[\"Total\"],\"settings.about.uninstall_cli\":[\"Uninstall\"],\"settings.about.updates\":[\"Updates\"],\"settings.about.version\":[\"Version\"],\"settings.about.version_label\":[\"Glimpse\"],\"settings.account.action.deactivate\":[\"Deactivate this device\"],\"settings.account.action.deactivate_cancel\":[\"Cancel\"],\"settings.account.action.deactivate_confirm_short\":[\"Deactivate\"],\"settings.account.action.refresh\":[\"Refresh\"],\"settings.account.activate.body\":[\"Paste your activation code from Polar.\"],\"settings.account.activate.input_aria\":[\"Activation code\"],\"settings.account.activate.placeholder\":[\"GLIMPSE_…\"],\"settings.account.activate.submit\":[\"Activate\"],\"settings.account.section.activate\":[\"Already bought?\"],\"settings.account.trial.text\":[\"Trial · \",[\"remaining\"],\" of \",[\"total\"],\" days left\"],\"settings.account.trial.text_ended\":[\"Your trial has ended\"],\"settings.account.trial.text_ended_on\":[\"Your trial ended on\"],\"settings.account.trial.text_one\":[\"Trial · 1 day left\"],\"settings.app.accessibility\":[\"Accessibility\"],\"settings.app.accessibility.description\":[\"required for auto-paste\"],\"settings.app.analytics\":[\"Usage Analytics\"],\"settings.app.analytics.body\":[\"anonymous, no transcripts or audio shared.\"],\"settings.app.analytics.more_info\":[\"More info\"],\"settings.app.analytics.toggle_aria\":[\"Toggle usage analytics\"],\"settings.app.appearance\":[\"Appearance\"],\"settings.app.auto_delete\":[\"Auto-delete\"],\"settings.app.auto_delete.after\":[\"after\"],\"settings.app.auto_delete.body\":[\"Deleting transcripts also removes their saved audio.\"],\"settings.app.auto_delete.confirm.apply\":[\"Apply anyway\"],\"settings.app.auto_delete.confirm.audio_only\":[\"This only removes saved audio files, not your transcripts.\"],\"settings.app.auto_delete.confirm.audio_too\":[\"Deleting transcripts also removes the audio they reference.\"],\"settings.app.auto_delete.confirm.immediately.known_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" will immediately delete \",[\"candidateCount\"],\" existing \",[\"noun\"],\".\"],\"settings.app.auto_delete.confirm.immediately.unknown_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" may immediately delete your existing \",[\"noun\"],\".\"],\"settings.app.auto_delete.confirm.threshold.known_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" will immediately delete \",[\"candidateCount\"],\" \",[\"noun\"],\" already older than \",[\"threshold\"],\".\"],\"settings.app.auto_delete.confirm.threshold.unknown_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" may immediately delete \",[\"noun\"],\" already older than \",[\"threshold\"],\".\"],\"settings.app.auto_delete.confirm.title\":[\"Delete older items now?\"],\"settings.app.auto_delete.confirm.unknown_count\":[\"We couldn't count them right now, but auto-delete will still run as soon as you save this change.\"],\"settings.app.auto_delete_recordings\":[\"Auto-delete Recordings\"],\"settings.app.auto_delete_recordings.after_suffix\":[\" after\"],\"settings.app.auto_delete_recordings.body\":[\"Auto deletes local audio files.\"],\"settings.app.auto_delete_recordings.confirm.apply\":[\"Apply anyway\"],\"settings.app.auto_delete_recordings.confirm.audio_only\":[\"This only removes saved local audio files, not your transcript history.\"],\"settings.app.auto_delete_recordings.confirm.immediately.known_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" will immediately delete \",[\"candidateCount\",\"plural\",{\"one\":[\"#\",\" existing local recording\"],\"other\":[\"#\",\" existing local recordings\"]}],\".\"],\"settings.app.auto_delete_recordings.confirm.immediately.unknown_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" may immediately delete your existing local recordings.\"],\"settings.app.auto_delete_recordings.confirm.threshold.known_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" will immediately delete \",[\"candidateCount\",\"plural\",{\"one\":[\"#\",\" local recording that is already older than \",[\"threshold\"]],\"other\":[\"#\",\" local recordings that are already older than \",[\"threshold\"]]}],\".\"],\"settings.app.auto_delete_recordings.confirm.threshold.unknown_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" may immediately delete local recordings that are already older than \",[\"threshold\"],\".\"],\"settings.app.auto_delete_recordings.confirm.title\":[\"Delete older recordings now?\"],\"settings.app.auto_delete_recordings.confirm.unknown_count\":[\"We couldn't count them right now, but the cleanup will still run as soon as you save this change.\"],\"settings.app.auto_delete_recordings.preview.loading\":[\"Checking how many recordings would be deleted\"],\"settings.app.auto_launch\":[\"Launch at Login\"],\"settings.app.auto_launch.body\":[\"starts Glimpse automatically when you sign in.\"],\"settings.app.auto_launch.toggle_aria\":[\"Toggle launch at login\"],\"settings.app.auto_pause_media\":[\"Auto-pause Media\"],\"settings.app.auto_pause_media.body\":[\"pauses music while recording, resumes when done.\"],\"settings.app.auto_pause_media.body_duck\":[\"Lowers system volume while recording.\"],\"settings.app.auto_pause_media.body_off\":[\"System audio plays while recording.\"],\"settings.app.auto_pause_media.body_pause\":[\"Pauses system audio while recording.\"],\"settings.app.auto_pause_media.lower\":[\"Previous media action\"],\"settings.app.auto_pause_media.off\":[\"Off\"],\"settings.app.auto_pause_media.pause\":[\"Pause\"],\"settings.app.auto_pause_media.raise\":[\"Next media action\"],\"settings.app.auto_pause_media.toggle_aria\":[\"Toggle auto-pause media while recording\"],\"settings.app.auto_update\":[\"Auto-update\"],\"settings.app.auto_update.body\":[\"downloads and installs updates in the background.\"],\"settings.app.auto_update.toggle_aria\":[\"Toggle auto-update\"],\"settings.app.automation\":[\"Automation\"],\"settings.app.cancel\":[\"Cancel\"],\"settings.app.confirm\":[\"Confirm\"],\"settings.app.confirm.checking\":[\"Checking...\"],\"settings.app.input_monitoring\":[\"Input Monitoring\"],\"settings.app.input_monitoring.description\":[\"required for global shortcuts\"],\"settings.app.language.label\":[\"Language\"],\"settings.app.language.search\":[\"Search language...\"],\"settings.app.language.system\":[\"System\"],\"settings.app.microphone\":[\"Microphone\"],\"settings.app.microphone.description\":[\"required for transcription\"],\"settings.app.open_settings\":[\"Open Settings\"],\"settings.app.permission.checking\":[\"Checking permission\"],\"settings.app.permission.enabled\":[\"Enabled\"],\"settings.app.permission.off\":[\"off\"],\"settings.app.permissions_restart_notice\":[\"Permission changes may require a restart.\"],\"settings.app.privacy_permissions\":[\"Privacy & Permissions\"],\"settings.app.prune.day\":[\"A Day\"],\"settings.app.prune.instantly\":[\"Instantly\"],\"settings.app.prune.month\":[\"A Month\"],\"settings.app.prune.never\":[\"Never\"],\"settings.app.prune.noun.audio.one\":[\"audio file\"],\"settings.app.prune.noun.audio.other\":[\"audio files\"],\"settings.app.prune.noun.transcripts.one\":[\"transcript\"],\"settings.app.prune.noun.transcripts.other\":[\"transcripts\"],\"settings.app.prune.threshold.day\":[\"a day\"],\"settings.app.prune.threshold.immediately\":[\"right now\"],\"settings.app.prune.threshold.month\":[\"a month\"],\"settings.app.prune.threshold.week\":[\"a week\"],\"settings.app.prune.threshold.year\":[\"a year\"],\"settings.app.prune.week\":[\"A Week\"],\"settings.app.prune.year\":[\"A Year\"],\"settings.app.prune_target.audio\":[\"Audio\"],\"settings.app.prune_target.transcripts\":[\"Transcripts\"],\"settings.app.start_in_background\":[\"Start in background\"],\"settings.app.start_in_background.toggle_aria\":[\"Toggle start in background\"],\"settings.app.text_size.default\":[\"Default\"],\"settings.app.text_size.label\":[\"Text Size\"],\"settings.app.text_size.large\":[\"Large\"],\"settings.app.text_size.small\":[\"Small\"],\"settings.app.theme.dark\":[\"Dark\"],\"settings.app.theme.label\":[\"Theme\"],\"settings.app.theme.light\":[\"Light\"],\"settings.app.theme.system\":[\"System\"],\"settings.general.account_tab\":[\"Account\"],\"settings.general.auto_dictionary\":[\"Auto Dictionary\"],\"settings.general.auto_dictionary.body\":[\"suggests names and terms after you correct dictated text\"],\"settings.general.auto_dictionary.toggle_aria\":[\"Toggle Auto Dictionary\"],\"settings.general.auto_dictionary.unsupported_body\":[\"requires a model with dictionary support\"],\"settings.general.cloud.aria\":[\"Cloud processing (Coming soon)\"],\"settings.general.cloud.badge\":[\"coming soon\"],\"settings.general.cloud.description\":[\"In development\"],\"settings.general.cloud.label\":[\"Cloud\"],\"settings.general.download_one\":[\"Download one\"],\"settings.general.edit_mode\":[\"Edit Mode\"],\"settings.general.edit_mode.body\":[\"transform selected text with voice\"],\"settings.general.edit_mode.configure_prefix\":[\"Set up AI writing in\"],\"settings.general.edit_mode.help\":[\"Select text in any app, and speak a command like \\\"make this formal\\\" or \\\"fix my grammar\\\".\"],\"settings.general.edit_mode.help_license_requirement\":[\"Requires Glimpse Personal.\"],\"settings.general.edit_mode.help_requirement\":[\"Requires an enabled and configured writing provider.\"],\"settings.general.edit_mode.info_aria\":[\"More information about Edit Mode\"],\"settings.general.edit_mode.license_prefix\":[\"Activate Glimpse Personal in\"],\"settings.general.edit_mode.models_suffix\":[\"to use Edit Mode.\"],\"settings.general.edit_mode.toggle_aria\":[\"Toggle Edit Mode\"],\"settings.general.features\":[\"Features\"],\"settings.general.language_info.active_model\":[\"Language list is filtered to the active transcription model.\"],\"settings.general.language_info_aria\":[\"More information about transcription language support\"],\"settings.general.local.badge\":[\"private\"],\"settings.general.local.description\":[\"Runs entirely on your device\"],\"settings.general.local.label\":[\"Local\"],\"settings.general.microphone\":[\"Microphone\"],\"settings.general.microphone_test.busy_error\":[\"That microphone is already in use.\"],\"settings.general.microphone_test.done\":[\"Done\"],\"settings.general.microphone_test.generic_error\":[\"Couldn't start microphone test.\"],\"settings.general.microphone_test.not_found_error\":[\"No microphone was found.\"],\"settings.general.microphone_test.permission_error\":[\"Microphone access was denied.\"],\"settings.general.microphone_test.start_error\":[\"Couldn't start microphone test.\"],\"settings.general.microphone_test.stop\":[\"Stop\"],\"settings.general.microphone_test.test\":[\"Test\"],\"settings.general.microphone_test.unsupported\":[\"Microphone testing isn't available in this window.\"],\"settings.general.no_model\":[\"No model installed.\"],\"settings.general.processing\":[\"Processing\"],\"settings.general.processing_mode\":[\"Processing Mode\"],\"settings.general.providers_tab\":[\"Providers\"],\"settings.general.search_language\":[\"Search language...\"],\"settings.general.select_microphone\":[\"Select microphone...\"],\"settings.general.shortcut.toggle_aria\":[\"Toggle \",[\"label\"],\" shortcut\"],\"settings.general.shortcuts\":[\"Shortcuts\"],\"settings.general.shortcuts.add_shortcut\":[\"+ Add shortcut\"],\"settings.general.shortcuts.cleanup\":[\"Cleanup\"],\"settings.general.shortcuts.help_cleanup\":[\"Runs Cleanup for that shortcut only.\"],\"settings.general.shortcuts.help_temporary\":[\"Makes a shortcut temporary. It will not save audio, transcript, or history.\"],\"settings.general.shortcuts.hide_shortcuts\":[\"Hide shortcuts\"],\"settings.general.shortcuts.hold\":[\"Hold\"],\"settings.general.shortcuts.hold_description\":[\"hold to talk, release to stop\"],\"settings.general.shortcuts.info_aria\":[\"More information about shortcut options\"],\"settings.general.shortcuts.remove_shortcut\":[\"Remove shortcut\"],\"settings.general.shortcuts.show_shortcuts\":[\"Show shortcuts\"],\"settings.general.shortcuts.smart\":[\"Smart\"],\"settings.general.shortcuts.smart_description\":[\"tap to toggle, hold to talk\"],\"settings.general.shortcuts.temporary\":[\"Temporary\"],\"settings.general.shortcuts.toggle\":[\"Toggle\"],\"settings.general.shortcuts.toggle_description\":[\"tap to start, tap to stop\"],\"settings.general.system_default\":[\"System Default\"],\"settings.general.to_use_local\":[\"to use local.\"],\"settings.general.transcription_language\":[\"Transcription Language\"],\"settings.language_model.api_key\":[\"API Key\"],\"settings.language_model.api_key.aria\":[\"LLM API Key\"],\"settings.language_model.api_key.optional\":[\"Optional\"],\"settings.language_model.api_key.optional_hint\":[\"(if required)\"],\"settings.language_model.api_key.required\":[\"Required\"],\"settings.language_model.description\":[\"Used by Cleanup, Edit Mode, and Personalization.\"],\"settings.language_model.endpoint\":[\"Endpoint\"],\"settings.language_model.endpoint.aria\":[\"LLM Endpoint URL\"],\"settings.language_model.endpoint.autofilled\":[\"(auto-filled)\"],\"settings.language_model.endpoint.placeholder\":[\"https://your-llm-endpoint.com\"],\"settings.language_model.model\":[\"Model\"],\"settings.language_model.model.placeholder\":[\"Model (default: \",[\"0\"],\")\"],\"settings.language_model.model.search\":[\"Search available models...\"],\"settings.language_model.provider\":[\"Provider\"],\"settings.language_model.provider.cloud\":[\"Cloud (API Key)\"],\"settings.language_model.provider.custom\":[\"Custom\"],\"settings.language_model.provider.custom.description\":[\"Enter your own endpoint URL\"],\"settings.language_model.provider.local\":[\"Local\"],\"settings.language_model.provider.search\":[\"Search providers...\"],\"settings.language_model.provider.select\":[\"Select provider...\"],\"settings.language_model.title\":[\"Writing Model Provider\"],\"settings.language_model.toggle\":[\"Use this provider for AI writing features\"],\"settings.local_api.allow_browser_requests\":[\"Allow browser requests\"],\"settings.local_api.allow_browser_requests_aria\":[\"Allow cross-origin browser requests\"],\"settings.local_api.allow_browser_requests_help\":[\"send CORS headers so web apps can call the API\"],\"settings.local_api.api_key\":[\"API key\"],\"settings.local_api.api_key.lan_required\":[\"Required when listening on LAN\"],\"settings.local_api.api_key.placeholder_optional\":[\"Optional - blank to disable\"],\"settings.local_api.api_key.placeholder_required\":[\"Required for LAN access\"],\"settings.local_api.behavior\":[\"Behavior\"],\"settings.local_api.configuration\":[\"Configuration\"],\"settings.local_api.copy.copied_exclaim\":[\"Copied!\"],\"settings.local_api.listen_on_lan\":[\"Listen on LAN\"],\"settings.local_api.listen_on_lan_aria\":[\"Listen on LAN\"],\"settings.local_api.listen_on_lan_help\":[\"expose to other devices on your network\"],\"settings.local_api.logs\":[\"Logs\"],\"settings.local_api.logs.clear\":[\"Clear\"],\"settings.local_api.logs.copied\":[\"Copied\"],\"settings.local_api.logs.copy\":[\"Copy\"],\"settings.local_api.logs.copy_aria\":[\"Copy logs\"],\"settings.local_api.logs.empty\":[\"No logs yet.\"],\"settings.local_api.model.currently_loaded\":[\"currently loaded: \",[\"loadedModelLabel\"]],\"settings.local_api.model.kept_warm\":[\"kept warm in memory for fast responses\"],\"settings.local_api.model.none\":[\"None\"],\"settings.local_api.model.none_held\":[\"no model held in memory between requests\"],\"settings.local_api.port\":[\"Port\"],\"settings.local_api.preloaded_model\":[\"Preloaded model\"],\"settings.local_api.restart_required\":[\"Restart to apply changes\"],\"settings.local_api.start\":[\"Start API\"],\"settings.local_api.start_on_launch\":[\"Start on launch\"],\"settings.local_api.start_on_launch_aria\":[\"Start local API when Glimpse opens\"],\"settings.local_api.start_on_launch_help\":[\"automatically start when Glimpse opens\"],\"settings.local_api.status.running\":[\"Running\"],\"settings.local_api.status.stopped\":[\"Stopped\"],\"settings.local_api.stop\":[\"Stop API\"],\"settings.local_api.title\":[\"API Server\"],\"settings.modal.close_button\":[\"Close settings\"],\"settings.modal.dialog_label\":[\"Settings\"],\"settings.modal.section.core\":[\"Core\"],\"settings.modal.section.developer\":[\"Developer\"],\"settings.modal.section.local\":[\"Local\"],\"settings.modal.tab.about\":[\"About\"],\"settings.modal.tab.account\":[\"Account\"],\"settings.modal.tab.api_server\":[\"API Server\"],\"settings.modal.tab.app\":[\"App\"],\"settings.modal.tab.general\":[\"General\"],\"settings.modal.tab.models\":[\"Models\"],\"settings.modal.tab.providers\":[\"Providers\"],\"settings.modal.title\":[\"Settings\"],\"settings.models.active\":[\"Active\"],\"settings.models.back\":[\"Back\"],\"settings.models.back_overview\":[\"Overview\"],\"settings.models.browse_all\":[\"Browse all models\"],\"settings.models.cancel\":[\"Cancel\"],\"settings.models.cancel_download\":[\"Cancel download\"],\"settings.models.cancelled\":[\"Cancelled\"],\"settings.models.cloud_active\":[\"Glimpse is using \",[\"provider\"],\" to transcribe. Your active local model will be used as a fallback.\"],\"settings.models.cloud_active.provider_fallback\":[\"your speech provider\"],\"settings.models.delete\":[\"Delete\"],\"settings.models.delete_model\":[\"Delete model\"],\"settings.models.description\":[\"Manage local transcription engines and downloaded speech models.\"],\"settings.models.download\":[\"Download\"],\"settings.models.download_model\":[\"Download model\"],\"settings.models.engine.generic.description\":[[\"engineLabel\"],\" transcription engine.\"],\"settings.models.engine.nvidia.description\":[\"NVIDIA local speech models, including Parakeet for transcription and Nemotron for live streaming.\"],\"settings.models.engine.whisper.description\":[\"OpenAI's speech recognition with custom vocabulary support.\"],\"settings.models.installed\":[\"Installed\"],\"settings.models.installed.active\":[\"Active\"],\"settings.models.installed.ane\":[\"ANE\"],\"settings.models.installed.delete\":[\"Delete\"],\"settings.models.installed.delete_model\":[\"Delete model\"],\"settings.models.installed.english\":[\"English\"],\"settings.models.installed.multilingual\":[\"Multilingual\"],\"settings.models.installed.use\":[\"Use\"],\"settings.models.installed_count\":[[\"installedCount\"],\" installed\"],\"settings.models.recommended\":[\"Recommended\"],\"settings.models.remote_speech_active\":[\"Currently using your Speech Provider\",[\"0\"],\". Local models remain available as fallback.\"],\"settings.models.title\":[\"Models\"],\"settings.models.transcription_engines\":[\"Transcription Engines\"],\"settings.models.use\":[\"Use\"],\"settings.providers.description\":[\"Connect local-mode AI providers for Cleanup, Edit Mode, and Personalization.\"],\"settings.providers.language_label\":[\"Language\"],\"settings.providers.speech_label\":[\"Speech\"],\"settings.providers.title\":[\"Providers\"],\"settings.providers.writing_section\":[\"Writing\"],\"settings.speech_model.api_key\":[\"API Key\"],\"settings.speech_model.api_key.aria\":[\"Remote speech API key\"],\"settings.speech_model.api_key.optional\":[\"Optional\"],\"settings.speech_model.api_key.optional_hint\":[\"(if required)\"],\"settings.speech_model.api_key.required\":[\"Required\"],\"settings.speech_model.description\":[\"Transcribe recordings through OpenAI-compatible cloud or self-hosted APIs.\"],\"settings.speech_model.endpoint.aria\":[\"Remote speech endpoint URL\"],\"settings.speech_model.endpoint.placeholder\":[\"https://your-speech-endpoint.com\"],\"settings.speech_model.model\":[\"Model\"],\"settings.speech_model.model.automatic\":[\"Automatic (\",[\"0\"],\")\"],\"settings.speech_model.model.placeholder\":[\"Model (default: \",[\"0\"],\")\"],\"settings.speech_model.model.search\":[\"Search available speech models...\"],\"settings.speech_model.provider\":[\"Provider\"],\"settings.speech_model.provider.cloud\":[\"Cloud (API Key)\"],\"settings.speech_model.provider.custom\":[\"Custom\"],\"settings.speech_model.provider.custom.description\":[\"Enter your own endpoint URL\"],\"settings.speech_model.provider.local\":[\"Local\"],\"settings.speech_model.provider.search\":[\"Search speech providers...\"],\"settings.speech_model.provider.select\":[\"Select provider...\"],\"settings.speech_model.title\":[\"Remote Speech Provider\"],\"settings.speech_model.toggle\":[\"Use this provider for speech-to-text\"],\"toast.close\":[\"Close notification\"],\"toast.copied\":[\"Copied\"],\"toast.copy_message\":[\"Copy message\"],\"toast.retry_failed\":[\"Retry failed. Please try again.\"],\"toast.retry_transcription\":[\"Retry transcription\"],\"toast.retrying\":[\"Retrying...\"],\"transcription.language.auto\":[\"Auto\"],\"transcription.language.unsupported\":[\"Unsupported\"],\"transcription.language.unsupported_description\":[\"Choose a compatible model to enable these.\"],\"transcriptions.filter.sort\":[\"Sort\"],\"transcriptions.filter.when\":[\"When\"],\"transcriptions.group.today\":[\"Today\"],\"transcriptions.group.yesterday\":[\"Yesterday\"],\"transcriptions.item.cleaning\":[\"Cleaning...\"],\"transcriptions.item.copied\":[\"Copied\"],\"transcriptions.item.copy_selection\":[\"Copy selection\"],\"transcriptions.item.copy_transcription\":[\"Copy transcription\"],\"transcriptions.item.delete\":[\"Delete\"],\"transcriptions.item.error.default\":[\"Transcription failed\"],\"transcriptions.item.failed\":[\"Failed\"],\"transcriptions.item.more_options\":[\"More options\"],\"transcriptions.item.restore_original\":[\"Restore original transcript\"],\"transcriptions.item.retry\":[\"Retry\"],\"transcriptions.item.retry_cleanup\":[\"Retry cleanup\"],\"transcriptions.item.retrying\":[\"Retrying...\"],\"transcriptions.item.reverting\":[\"Reverting...\"],\"transcriptions.item.run_cleanup\":[\"Run cleanup\"],\"transcriptions.item.show_less\":[\"Show less\"],\"transcriptions.item.show_more\":[\"Show more\"],\"transcriptions.item.stop_retry\":[\"Stop transcription\"],\"transcriptions.list.empty\":[\"Your recent transcriptions will appear here\"],\"transcriptions.list.filter.aria\":[\"Sort and filter transcriptions\"],\"transcriptions.list.no_results\":[\"No results for \\\"\",[\"resultSearchText\"],\"\\\"\"],\"transcriptions.list.no_results_filters\":[\"No results for selected filters\"],\"transcriptions.list.search.aria\":[\"Search transcriptions\"],\"transcriptions.list.search.clear\":[\"Clear search\"],\"transcriptions.list.search.close\":[\"Close search\"],\"transcriptions.list.search.open\":[\"Search transcriptions\"],\"transcriptions.list.search.placeholder_short\":[\"Search\"],\"transcriptions.list.sort.aria\":[\"Sort transcriptions\"],\"transcriptions.sort.longest\":[\"Longest\"],\"transcriptions.sort.oldest\":[\"Oldest first\"],\"transcriptions.sort.recent\":[\"Newest first\"],\"transcriptions.sort.shortest\":[\"Shortest\"],\"transcriptions.time.7d\":[\"Past 7 days\"],\"transcriptions.time.any\":[\"Any time\"],\"transcriptions.time.today\":[\"Today\"],\"u-NVCP\":[[\"avg\",\"plural\",{\"one\":[\"About \",\"#\",\" word per dictation today\"],\"other\":[\"About \",\"#\",\" words per dictation today\"]}]],\"updates.available_version\":[\"v\",[\"version\"],\" available\"],\"updates.check_aria\":[\"Check for updates\"],\"updates.check_failed\":[\"Update check failed\"],\"updates.check_title\":[\"Check for updates\"],\"updates.checking\":[\"Checking for updates...\"],\"updates.installed\":[\"Update installed\"],\"updates.ready_to_install\":[\"Ready to install\"],\"updates.restart\":[\"Restart\"],\"updates.restart_to_apply\":[\"Restart to apply\"],\"updates.retry\":[\"Retry\"],\"updates.up_to_date\":[\"You're up to date\"],\"updates.update\":[\"Update\"],\"updates.whats_new\":[\"What's new?\"],\"updates.whats_new.couldnt_load\":[\"Couldn't load releases\"],\"updates.whats_new.github_unavailable\":[\"GitHub may be temporarily unavailable. Check your connection and try again.\"],\"updates.whats_new.load_failed\":[\"Failed to load changelog\"],\"updates.whats_new.major_release\":[\"Major Release\"],\"updates.whats_new.no_changelog\":[\"No changelog available.\"],\"updates.whats_new.retry\":[\"Retry\"],\"updates.whats_new.title\":[\"What's New\"],\"updates.whats_new.view_all\":[\"View all releases on GitHub\"]}")};
+/*eslint-disable*/module.exports={messages:JSON.parse("{\"D-waf4\":[[\"seconds\",\"plural\",{\"one\":[\"#\",\" second spoken today\"],\"other\":[\"#\",\" seconds spoken today\"]}]],\"HQ7Iit\":[[\"0\",\"plural\",{\"one\":[\"#\",\" word\"],\"other\":[\"#\",\" words\"]}]],\"Vebvqp\":[[\"0\",\"plural\",{\"one\":[\"Longest dictation today: \",\"#\",\" word\"],\"other\":[\"Longest dictation today: \",\"#\",\" words\"]}]],\"ane_compile.body\":[\"macOS is compiling \",[\"label\"],\" for the Apple Neural Engine. This happens once and can take a few minutes.\"],\"ane_compile.title\":[\"Optimizing for the Neural Engine\"],\"dictionary.add_prompt\":[\"Press Enter to add \\\"\",[\"0\"],\"\\\" as a new entry.\"],\"dictionary.combined.description\":[\"Add custom words the system should recognize, and set automatic word replacements.\"],\"dictionary.combined.title\":[\"Dictionary & Replacements\"],\"dictionary.delete\":[\"Delete\"],\"dictionary.delete_entry\":[\"Delete \",[\"entry\"]],\"dictionary.edit_hint\":[\"Press Enter to save · Esc to cancel\"],\"dictionary.entry_count.multiple\":[[\"0\"],\" entries\"],\"dictionary.entry_count.single\":[\"1 entry\"],\"dictionary.full_add_prompt\":[\"Delete an entry before adding another.\"],\"dictionary.full_hint\":[\"Dictionary is full\"],\"dictionary.no_entries\":[\"No entries yet\"],\"dictionary.no_entries.description\":[\"Add words, phrases, or names above and press Enter to save them here.\"],\"dictionary.no_matches\":[\"No matches found\"],\"dictionary.press_enter_to_add\":[\"Press Enter to add\"],\"dictionary.press_enter_to_add_match\":[\"Press Enter to add this word\"],\"dictionary.replacements.count.multiple\":[[\"0\"],\" replacements\"],\"dictionary.replacements.count.single\":[\"1 replacement\"],\"dictionary.replacements.delete\":[\"Delete replacement for \",[\"0\"]],\"dictionary.replacements.find\":[\"Find word...\"],\"dictionary.replacements.find_aria\":[\"Find word to replace\"],\"dictionary.replacements.none\":[\"No replacements yet\"],\"dictionary.replacements.none_description\":[\"Add word pairs to automatically swap in transcriptions. Matches are case-insensitive.\"],\"dictionary.replacements.press_enter_to_add\":[\"Press Enter in either field to add\"],\"dictionary.replacements.remove_value\":[\"remove\"],\"dictionary.replacements.replace_with\":[\"Replace with...\"],\"dictionary.replacements.replace_with_aria\":[\"Replace with\"],\"dictionary.search_matches\":[[\"0\"],\" of \",[\"1\"],\" matches\"],\"dictionary.search_only\":[\"Search dictionary...\"],\"dictionary.search_or_add\":[\"Search or add a word...\"],\"dictionary.search_or_add_aria\":[\"Add or search dictionary entry\"],\"dictionary.section.dictionary_description\":[\"Add custom words Glimpse should recognize.\"],\"dictionary.section.dictionary_title\":[\"Dictionary\"],\"dictionary.section.replacements_description\":[\"Swap common phrases automatically after transcription.\"],\"dictionary.section.replacements_title\":[\"Replacements\"],\"dictionary.warning\":[\"Dictionary works only for models with dictionary support. Current model \",[\"0\"],\" will ignore these entries until you switch to a compatible model.\"],\"dictionary.warning_aria\":[\"Warning: model compatibility issue\"],\"dropdown.no_options\":[\"No options found\"],\"dropdown.placeholder\":[\"Select...\"],\"dropdown.search_aria\":[\"Search options\"],\"dropdown.search_placeholder\":[\"Search...\"],\"dropdown.toggle_menu\":[\"Toggle options\"],\"faq.ai_writing.answer\":[\"Only if you enable AI writing and set up a provider under Settings → Providers. Cleanup, Edit Mode, and Personalization then send the relevant text directly to that provider. Your API key stays stored locally in Glimpse.\"],\"faq.ai_writing.question\":[\"When does text leave my device?\"],\"faq.close_aria\":[\"Close FAQ\"],\"faq.delete.answer\":[\"Delete recordings from History, remove imported files from Library, or uninstall models from Settings → Models. Settings → App can auto-delete Audio only or full Transcripts (including linked audio). You can also open your Glimpse data folder from About → Storage to manage files directly.\"],\"faq.delete.question\":[\"How do I manage or delete my data?\"],\"faq.free.answer\":[\"Core dictation is free: local transcription, dictionary, replacements, and history. There are no per-minute fees or subscriptions for that. Library, AI Cleanup, Edit Mode, personalization with an LLM, the local API server, and the CLI are part of Glimpse Personal. You get a 14-day trial first; after that, activate a one-time Personal or Commercial license in Settings → Account.\"],\"faq.free.question\":[\"What is free vs Glimpse Personal?\"],\"faq.how_it_works.answer\":[\"Press your dictation shortcut, speak, and release. Glimpse transcribes on your device and inserts the text where your cursor is. Core dictation works offline. No account required.\"],\"faq.how_it_works.question\":[\"How does Glimpse work?\"],\"faq.permissions.answer\":[\"Microphone access to record your voice, and Accessibility access to insert text and read selected text for Edit Mode. Glimpse only uses these while you are actively dictating.\"],\"faq.permissions.question\":[\"What permissions does Glimpse need?\"],\"faq.privacy.answer\":[\"Your audio and transcripts stay on your Mac. Glimpse does not collect recordings, transcripts, API keys, or prompts. Optional anonymous usage analytics (things like session length and feature usage, never your content) help us improve the app. You can turn this off anytime in Settings → App.\"],\"faq.privacy.link\":[\"Privacy policy\"],\"faq.privacy.question\":[\"Where does my data go?\"],\"faq.subtitle\":[\"How Glimpse works, privacy, and AI writing\"],\"faq.title\":[\"Frequently Asked Questions\"],\"g960RG\":[[\"0\",\"plural\",{\"one\":[\"#\",\" dictation\"],\"other\":[\"#\",\" dictations\"]}]],\"h_9Xcr\":[[\"0\",\"plural\",{\"one\":[\"#\",\" dictation polished with AI today\"],\"other\":[\"#\",\" dictations polished with AI today\"]}]],\"home.drag_import.eyebrow\":[\"Library Import\"],\"home.drag_import.subtitle\":[\"MP3, WAV, M4A, MP4, MOV, and more\"],\"home.drag_import.title\":[\"Drop files to transcribe\"],\"home.greeting.afternoon\":[\"Good afternoon\"],\"home.greeting.evening\":[\"Good evening\"],\"home.greeting.morning\":[\"Good morning\"],\"home.local_api.copy_url\":[\"Copy base URL\"],\"home.local_api.open_settings\":[\"Open API server settings\"],\"home.local_api.requests\":[\"Requests served\"],\"home.local_api.running\":[\"Local API\"],\"home.local_api.status_hint\":[\"Your API server is running in the background.\"],\"home.mode.cloud\":[\"Cloud\"],\"home.mode.local\":[\"Local\"],\"home.sidebar.collapse\":[\"Collapse sidebar\"],\"home.sidebar.dictionary\":[\"Dictionary\"],\"home.sidebar.expand\":[\"Expand sidebar\"],\"home.sidebar.home\":[\"Home\"],\"home.sidebar.library\":[\"Library\"],\"home.sidebar.personalization\":[\"Personalization\"],\"home.sidebar.settings\":[\"Settings\"],\"home.support.about.title\":[\"About\"],\"home.support.about.version_mode\":[\"v\",[\"version\"],\" • \",[\"mode\"]],\"home.support.faq.subtitle\":[\"Common questions\"],\"home.support.faq.title\":[\"FAQ\"],\"home.support.feedback.email\":[\"Email\"],\"home.support.feedback.email_copied\":[\"Copied!\"],\"home.support.feedback.github\":[\"GitHub Issue\"],\"home.support.feedback.title\":[\"Feedback\"],\"home.support.label\":[\"Support\"],\"home.support.menu_aria\":[\"Support menu\"],\"home.support.title\":[\"Get Support\"],\"home.today.stats\":[[\"0\",\"plural\",{\"one\":[\"#\",\" dictation\"],\"other\":[\"#\",\" dictations\"]}],\" · \",[\"1\",\"plural\",{\"one\":[\"#\",\" word\"],\"other\":[\"#\",\" words\"]}],\" today\"],\"home.update_available\":[\"Update available\"],\"import.app_picker.aria\":[\"Select app to import from\"],\"import.apply.failed\":[\"Import failed. Try again or skip.\"],\"import.cat.auto_launch\":[\"Launch at login\"],\"import.cat.auto_launch.off\":[\"Off\"],\"import.cat.auto_launch.on\":[\"On\"],\"import.cat.dictionary\":[\"Dictionary\"],\"import.cat.dictionary.detail\":[[\"0\",\"plural\",{\"one\":[\"#\",\" word\"],\"other\":[\"#\",\" words\"]}]],\"import.cat.history\":[\"Transcript history\"],\"import.cat.history.detail\":[[\"0\"],\" transcripts\"],\"import.cat.language\":[\"Language\"],\"import.cat.model\":[\"Transcription model\"],\"import.cat.personalities\":[\"Personalities\"],\"import.cat.personalities.detail\":[[\"0\"],\" saved\"],\"import.cat.replacements\":[\"Text replacements\"],\"import.cat.replacements.detail\":[[\"0\",\"plural\",{\"one\":[\"#\",\" rule\"],\"other\":[\"#\",\" rules\"]}]],\"import.cat.shortcut\":[\"Keyboard shortcut\"],\"import.cta\":[\"Import\"],\"import.empty\":[\"Nothing to bring over from this app.\"],\"import.error\":[\"We couldn't read this app's data. You can skip and set things up manually.\"],\"import.importing\":[\"Importing...\"],\"import.model.unrecognized\":[\"We don't recognize this app's model — you'll pick one on the next step.\"],\"import.skip\":[\"Skip for now\"],\"import.subtitle\":[[\"0\",\"plural\",{\"one\":[\"We found another dictation app. Choose what to import.\"],\"other\":[\"We found other dictation apps. Choose what to import.\"]}]],\"import.subtitle.single\":[\"We noticed you use \",[\"sourceName\"],\". Pick what to carry over.\"],\"import.title\":[\"Want to bring your settings over?\"],\"library.card.cancel\":[\"Cancel\"],\"library.card.delete\":[\"Delete\"],\"library.card.new_tag\":[\"New tag...\"],\"library.card.no_other_tags\":[\"No other tags\"],\"library.card.no_tags_yet\":[\"No tags yet\"],\"library.card.remove_tag\":[\"Remove \",[\"tag\"]],\"library.card.rename\":[\"Rename\"],\"library.card.retranscribe\":[\"Retranscribe\"],\"library.card.retry\":[\"Retry\"],\"library.card.select_existing_tag\":[\"Select existing tag\"],\"library.delete_error.default\":[\"Failed to delete the library item.\"],\"library.delete_error.default_fallback\":[\"Failed to delete the library item.\"],\"library.delete_error.delete_files_failed\":[\"Couldn't delete the library files. Check permissions and try again.\"],\"library.delete_error.invalid_path\":[\"Couldn't delete this item due to an invalid file path.\"],\"library.delete_error.outside_folder\":[\"Couldn't delete this item because its files are outside the library folder.\"],\"library.delete_error.storage_missing\":[\"Couldn't delete this item. Library storage couldn't be found.\"],\"library.detail.add_speaker\":[\"Add speaker\"],\"library.detail.assign_new_speaker\":[\"Assign new speaker\"],\"library.detail.back\":[\"Back to library\"],\"library.detail.discuss\":[\"ChatGPT\"],\"library.detail.discuss.copied\":[\"Transcript copied — paste it into ChatGPT.\"],\"library.detail.discuss.header\":[\"Here's a transcript titled \\\"\",[\"0\"],\"\\\". Help me review, summarize, and discuss it.\"],\"library.detail.discuss.instruction\":[\"I have a transcript copied to my clipboard and will paste it in my next message. Help me review, summarize, and discuss it.\"],\"library.detail.discuss.tooltip\":[\"Open this transcript in ChatGPT\"],\"library.detail.filter.all\":[\"All speakers\"],\"library.detail.filter.aria\":[\"Filter by speaker\"],\"library.detail.filter.no_speakers\":[\"No speakers yet\"],\"library.detail.more_actions\":[\"More actions\"],\"library.detail.speaker.clear\":[\"Clear speaker\"],\"library.detail.speaker.remove\":[\"Remove \",[\"0\"]],\"library.detail.speaker.rename\":[\"Click to rename\"],\"library.detail.speaker.unassigned\":[\"Assign\"],\"library.detail.speaker_default_name\":[\"Speaker \",[\"nextIndex\"]],\"library.detail.speakers\":[\"Speakers\"],\"library.detail.tags.add\":[\"Add tag\"],\"library.detail.tags.label\":[\"Tag\"],\"library.error_details.copy_original_failed\":[\"Couldn't copy original file.\"],\"library.error_details.create_storage_failed\":[\"Couldn't create library storage.\"],\"library.error_details.ffmpeg_required\":[\"FFmpeg required for video imports.\"],\"library.error_details.file_not_found\":[\"File not found.\"],\"library.error_details.import_failed\":[\"Import failed.\"],\"library.error_details.insufficient_disk_space\":[\"Not enough disk space.\"],\"library.error_details.invalid_audio\":[\"Not a valid audio file.\"],\"library.error_details.model_not_installed\":[\"Model not installed.\"],\"library.error_details.no_audio_track\":[\"No audio track found.\"],\"library.error_details.unsupported_audio_settings\":[\"Unsupported audio settings.\"],\"library.error_details.unsupported_format\":[\"Unsupported file format.\"],\"library.filter.active\":[\"Active\"],\"library.filter.all\":[\"All\"],\"library.filter.aria_label\":[\"Filter library by status\"],\"library.filter.done\":[\"Done\"],\"library.filter.failed\":[\"Failed\"],\"library.import.cancel\":[\"Cancel\"],\"library.import.confirm\":[\"Import\"],\"library.import.files\":[\"Files\"],\"library.import.importing\":[\"Importing...\"],\"library.import.model\":[\"Model\"],\"library.import.no_models\":[\"No local models are installed. Download a model in Settings -> Models before importing.\"],\"library.import.remote_provider\":[\"Remote provider\"],\"library.import.search_models\":[\"Search installed models...\"],\"library.import.select_model\":[\"Select a model\"],\"library.import.show_timestamps\":[\"Show timestamps\"],\"library.import.show_timestamps.aria\":[\"Show timestamps\"],\"library.import.store_original\":[\"Store original file\"],\"library.import.store_original.aria\":[\"Store original\"],\"library.import.store_original.description\":[\"Keep a copy inside the library folder\"],\"library.import.summary.multiple\":[[\"0\"],\" files\"],\"library.import.summary.single\":[\"1 file\"],\"library.import.timestamps_supported\":[\"Enabled for supported models\"],\"library.import.timestamps_unsupported\":[\"Not supported by this model\"],\"library.import.title\":[\"Import to Library\"],\"library.import_error.audio_conversion_failed\":[\"Couldn't convert this file to audio for transcription.\"],\"library.import_error.copy_original_failed\":[\"Couldn't copy the original file into the library.\"],\"library.import_error.create_storage_failed\":[\"Couldn't create library storage. Check disk permissions.\"],\"library.import_error.decode_failed\":[\"Couldn't decode this audio file. Try installing FFmpeg.\"],\"library.import_error.default\":[\"Import failed for one of the files.\"],\"library.import_error.default_fallback\":[\"Import failed for one of the files.\"],\"library.import_error.file_not_found\":[\"File not found. It may have moved or been deleted.\"],\"library.import_error.model_not_installed\":[\"Selected model isn't installed. Download one in Settings -> Models.\"],\"library.import_error.no_audio_track\":[\"No audio track found in this file.\"],\"library.import_error.unsupported_format\":[\"Unsupported file format.\"],\"library.import_error.unsupported_sample_rate\":[\"This file has an unsupported sample rate.\"],\"library.modal.audio\":[\"Audio\"],\"library.modal.audio_scrubber\":[\"Audio scrubber\"],\"library.modal.audio_unavailable\":[\"Audio unavailable\"],\"library.modal.cancel\":[\"Cancel\"],\"library.modal.close\":[\"Close\"],\"library.modal.copy\":[\"Copy\"],\"library.modal.copy.copied\":[\"Copied\"],\"library.modal.delete\":[\"Delete\"],\"library.modal.delete_confirm.description\":[\"This removes the transcript and audio from your library.\"],\"library.modal.delete_confirm.title\":[\"Delete this item?\"],\"library.modal.export\":[\"Export\"],\"library.modal.export.failed\":[\"Export failed. Try again.\"],\"library.modal.export.item_not_found\":[\"Couldn't find this library item. Try reopening it.\"],\"library.modal.export.no_timestamps\":[\"This item doesn't have timestamps. Retranscribe with timestamps to export subtitles.\"],\"library.modal.export.title\":[\"Export transcription\"],\"library.modal.export.write_failed\":[\"Couldn't write the export file. Try a different location.\"],\"library.modal.ffmpeg_help\":[\"FFmpeg Help\"],\"library.modal.follow_timestamp\":[\"Follow timestamp\"],\"library.modal.import_failed\":[\"Import failed\"],\"library.modal.import_status.converting\":[\"Converting audio...\"],\"library.modal.import_status.converting_progress\":[\"Converting audio... \",[\"0\"],\"%\"],\"library.modal.import_status.queued\":[\"Queued for transcription...\"],\"library.modal.pause_audio\":[\"Pause audio\"],\"library.modal.play_audio\":[\"Play audio\"],\"library.modal.playback.decrease\":[\"Decrease playback speed\"],\"library.modal.playback.increase\":[\"Increase playback speed\"],\"library.modal.retranscribe\":[\"Retranscribe\"],\"library.modal.retry\":[\"Retry\"],\"library.modal.search.aria\":[\"Search transcript\"],\"library.modal.search.clear\":[\"Clear search\"],\"library.modal.search.placeholder\":[\"Search transcript...\"],\"library.modal.settings\":[\"Settings\"],\"library.modal.tags\":[\"Tags\"],\"library.modal.tags.new_tag\":[\"New tag...\"],\"library.modal.tags.no_other_tags\":[\"No other tags\"],\"library.modal.tags.no_tags_yet\":[\"No tags yet\"],\"library.modal.tags.none\":[\"None\"],\"library.modal.tags.remove\":[\"Remove \",[\"tag\"]],\"library.modal.tags.select_existing\":[\"Select existing tag\"],\"library.modal.timestamps\":[\"Timestamps\"],\"library.modal.timestamps.available_later\":[\"Available after transcription\"],\"library.modal.timestamps.supported\":[\"Supported\"],\"library.modal.timestamps.unsupported\":[\"Not supported\"],\"library.modal.transcribing\":[\"Transcribing...\"],\"library.modal.transcript_placeholder\":[\"Transcript will appear here.\"],\"library.retranscribe.cancel\":[\"Cancel\"],\"library.retranscribe.confirm\":[\"Retranscribe\"],\"library.retranscribe.loading\":[\"Retranscribing...\"],\"library.retranscribe.model\":[\"Model\"],\"library.retranscribe.no_models\":[\"No local models are installed. Download a model in Settings -> Models before retranscribing.\"],\"library.retranscribe.remote_provider\":[\"Remote provider\"],\"library.retranscribe.search_models\":[\"Search installed models...\"],\"library.retranscribe.select_model\":[\"Select a model\"],\"library.retranscribe.show_timestamps\":[\"Show timestamps\"],\"library.retranscribe.timestamps_supported\":[\"Enabled for supported models\"],\"library.retranscribe.timestamps_unsupported\":[\"Not supported by this model\"],\"library.retranscribe.title\":[\"Retranscribe\"],\"library.view.description\":[\"Import audio and video files for transcription.\"],\"library.view.dropzone\":[\"Drop files to import\"],\"library.view.empty_state\":[\"Drag files here to build your Library.\"],\"library.view.file_filter\":[\"Audio & Video\"],\"library.view.import_button\":[\"Import\"],\"library.view.import_dialog_error\":[\"Could not open the import dialog.\"],\"library.view.load_more\":[\"Load more\"],\"library.view.loading_more\":[\"Loading...\"],\"library.view.search_placeholder\":[\"Search library...\"],\"library.view.title\":[\"Library\"],\"library.view.unsupported_files\":[[\"0\"],\" file(s) skipped due to unsupported format.\"],\"license.customer_portal\":[\"Customer portal\"],\"member_card.aria\":[\"Glimpse member card\"],\"member_card.coverage_with_devices\":[[\"coverageBase\"],\" · \",[\"activeDevices\"],\" of \",[\"deviceLimit\"],\" devices active\"],\"member_card.draft_aria\":[\"Draft Glimpse member card\"],\"member_card.draft_idle\":[\"Pick a license\"],\"member_card.draft_stamp_empty\":[\"Unissued\"],\"member_card.draft_stamp_issuing\":[\"Issuing\"],\"member_card.label_member_since\":[\"Member since\"],\"member_card.label_words_spoken\":[\"Words spoken\"],\"member_card.tier_commercial\":[\"Commercial\"],\"member_card.tier_contributor\":[\"Contributor\"],\"member_card.tier_founder\":[\"Founder\"],\"member_card.tier_personal\":[\"Personal\"],\"member_card.tier_purchase_aria\":[\"Purchase \",[\"label\"],\" for \",[\"price\"]],\"model_picker.active\":[\"Active\"],\"model_picker.ane.download\":[\"Download Neural Engine encoder\"],\"model_picker.ane.info\":[\"Adds a Core ML encoder that runs on the Apple Neural Engine for faster, more power-efficient transcription. The first load takes longer while macOS optimizes it for your chip.\"],\"model_picker.ane.info_aria\":[\"About the Apple Neural Engine encoder\"],\"model_picker.ane.installed\":[\"Neural Engine encoder installed\"],\"model_picker.ane.toggle\":[\"Include the Apple Neural Engine encoder\"],\"model_picker.cancel\":[\"Cancel\"],\"model_picker.cancelled\":[\"Cancelled\"],\"model_picker.category.experimental\":[\"Experimental\"],\"model_picker.category.legacy\":[\"Legacy\"],\"model_picker.category.standard\":[\"Standard\"],\"model_picker.close\":[\"Close\"],\"model_picker.delete\":[\"Delete\"],\"model_picker.download\":[\"Download\"],\"model_picker.english\":[\"English\"],\"model_picker.filter.all\":[\"All models\"],\"model_picker.filter.aria\":[\"Filter models by category\"],\"model_picker.multilingual\":[\"Multilingual\"],\"model_picker.no_results\":[\"No models match your search.\"],\"model_picker.quant\":[\"Quantization\"],\"model_picker.search\":[\"Search models\"],\"model_picker.search_aria\":[\"Search models\"],\"model_picker.title\":[\"Choose a model\"],\"model_picker.use\":[\"Use\"],\"model_picker.variant\":[\"Model variant\"],\"model_picker.variant_installed\":[\"Model variant (installed)\"],\"models.card.aria\":[[\"0\"],\" model\"],\"models.card.cancel\":[\"Cancel\"],\"models.card.cancel_download\":[\"Cancel download\"],\"models.card.delete\":[\"Delete\"],\"models.card.delete_model\":[\"Delete model\"],\"models.card.download\":[\"Download\"],\"models.card.download_model\":[\"Download model\"],\"models.card.downloading\":[\"Downloading\"],\"models.card.verifying\":[\"Verifying install\"],\"myYTaz\":[\"Trial · \",[\"trialDaysRemaining\"],\" of \",[\"TRIAL_TOTAL_DAYS\"],\" days left\"],\"onboarding.accessibility.app_name\":[\"Glimpse\"],\"onboarding.accessibility.continue\":[\"Continue\"],\"onboarding.accessibility.enable\":[\"Enable in Settings\"],\"onboarding.accessibility.instructions\":[\"Click below to open System Settings, then toggle on \",[\"appName\"]],\"onboarding.accessibility.skip\":[\"Skip\"],\"onboarding.accessibility.subtitle\":[\"Enables auto-paste into any application.\"],\"onboarding.accessibility.title\":[\"Accessibility\"],\"onboarding.back\":[\"Back\"],\"onboarding.brand\":[\"Glimpse\"],\"onboarding.complete.failed\":[\"Could not finish setup. Check your settings and try again.\"],\"onboarding.complete.no_model\":[\"Could not load a local model selection. Try reopening onboarding.\"],\"onboarding.delete.failed\":[\"Delete failed\"],\"onboarding.download.failed\":[\"Download failed\"],\"onboarding.download.starting\":[\"starting...\"],\"onboarding.license.activate.input_aria\":[\"Activation code\"],\"onboarding.license.activate.placeholder\":[\"GLIMPSE_...\"],\"onboarding.license.activate.submit\":[\"Activate\"],\"onboarding.license.active_subtitle\":[\"This Mac is licensed and ready.\"],\"onboarding.license.analytics_notice\":[\"Anonymous usage analytics help improve Glimpse. Turn this off anytime in Settings → App.\"],\"onboarding.license.continue_active\":[\"Continue\"],\"onboarding.license.continue_trial\":[\"Continue with trial\"],\"onboarding.license.saving\":[\"Saving...\"],\"onboarding.license.subtitle\":[\"Buy once, or activate the code from your receipt. You can skip this and keep using the trial.\"],\"onboarding.license.title\":[\"Connect your license\"],\"onboarding.microphone.continue\":[\"Continue\"],\"onboarding.microphone.grant\":[\"Grant Access\"],\"onboarding.microphone.skip\":[\"Skip\"],\"onboarding.microphone.subtitle\":[\"Required to capture your voice for transcription.\"],\"onboarding.microphone.title\":[\"Microphone Access\"],\"onboarding.models.action.delete_aria\":[\"Delete \",[\"modelLabel\"]],\"onboarding.models.action.download_aria\":[\"Download \",[\"modelLabel\"]],\"onboarding.models.action.stop_download_aria\":[\"Stop downloading \",[\"modelLabel\"]],\"onboarding.models.cancelled\":[\"Cancelled\"],\"onboarding.models.confirm_without_model.body\":[\"You haven't downloaded a local model yet. Transcription will not run offline until you add one in Settings.\"],\"onboarding.models.confirm_without_model.continue\":[\"Continue anyway\"],\"onboarding.models.confirm_without_model.stay\":[\"Stay here\"],\"onboarding.models.confirm_without_model.title\":[\"Continue without a model?\"],\"onboarding.models.continue\":[\"Continue\"],\"onboarding.models.download\":[\"Download\"],\"onboarding.models.download_failed\":[\"Download failed\"],\"onboarding.models.downloaded\":[\"Downloaded\"],\"onboarding.models.empty_body\":[\"This build did not return any local models. You can continue setup and manage models later in Settings.\"],\"onboarding.models.empty_title\":[\"No local models found\"],\"onboarding.models.loading_body\":[\"Fetching the available local transcription engines for this build.\"],\"onboarding.models.loading_cta\":[\"Loading...\"],\"onboarding.models.loading_title\":[\"Loading local models\"],\"onboarding.models.select_aria\":[\"Select \",[\"modelLabel\"]],\"onboarding.models.subtitle\":[\"More models and language model setup available in Settings after setup.\"],\"onboarding.models.title\":[\"Choose your local model\"],\"onboarding.models.unavailable_body\":[\"Glimpse couldn't load the local model list. Setup can continue with the default local engine, and you can manage downloads later in Settings.\"],\"onboarding.models.unavailable_title\":[\"Model list unavailable\"],\"onboarding.permissions.continue\":[\"Continue\"],\"onboarding.permissions.skip\":[\"Skip\"],\"onboarding.permissions.subtitle\":[\"Glimpse needs these to record your voice and put text where your cursor is.\"],\"onboarding.permissions.title\":[\"Permissions\"],\"onboarding.ready.analytics_notice\":[\"Glimpse sends anonymous usage analytics to help improve the app. You can disable this anytime in Settings -> App.\"],\"onboarding.ready.continue\":[\"Continue\"],\"onboarding.ready.smart.badge\":[\"Default\"],\"onboarding.ready.smart.capture_aria\":[\"Record new shortcut for Smart, currently \",[\"shortcut\"]],\"onboarding.ready.smart.capture_help_active\":[\"Press your new shortcut, or hit Esc to cancel.\"],\"onboarding.ready.smart.capture_help_idle\":[\"Click the shortcut to change it.\"],\"onboarding.ready.smart.capture_prompt\":[\"Press new shortcut...\"],\"onboarding.ready.smart.description\":[\"tap to toggle, hold to talk\"],\"onboarding.ready.smart.label\":[\"Smart\"],\"onboarding.ready.subtitle\":[\"Smart is on by default, others available in settings.\"],\"onboarding.ready.title\":[\"You're ready!\"],\"onboarding.setup.analytics_notice\":[\"Glimpse sends anonymous usage analytics to help improve the app. You can disable this anytime in Settings -> App.\"],\"onboarding.setup.back\":[\"Back\"],\"onboarding.setup.confirm_without_model.body\":[\"Transcription will not run offline until you download a local model.\"],\"onboarding.setup.confirm_without_model.continue\":[\"Continue anyway\"],\"onboarding.setup.confirm_without_model.stay\":[\"Stay here\"],\"onboarding.setup.confirm_without_model.title\":[\"Continue without a model?\"],\"onboarding.setup.continue\":[\"Continue\"],\"onboarding.setup.language.aria\":[\"Choose primary dictation language\"],\"onboarding.setup.language.english\":[\"English\"],\"onboarding.setup.language.multilingual\":[\"Multiple languages\"],\"onboarding.setup.language.placeholder\":[\"choose a language\"],\"onboarding.setup.language.question\":[\"Which language do you dictate in?\"],\"onboarding.setup.model.cancelled\":[\"Cancelled\"],\"onboarding.setup.model.download\":[\"Download\"],\"onboarding.setup.model.download_failed\":[\"Download failed\"],\"onboarding.setup.model.empty_body\":[\"You can continue setup and manage local models later in Settings.\"],\"onboarding.setup.model.empty_title\":[\"No local models found\"],\"onboarding.setup.model.loading_body\":[\"Glimpse is checking the models available in this build.\"],\"onboarding.setup.model.loading_title\":[\"Finding your local model\"],\"onboarding.setup.model.recommendation_label\":[\"Glimpse recommends\"],\"onboarding.setup.model.remove\":[\"Remove\"],\"onboarding.setup.model.stop\":[\"Stop\"],\"onboarding.setup.model.unavailable_body\":[\"Model list unavailable. You can manage local models later in Settings.\"],\"onboarding.setup.model.unavailable_title\":[\"Model list unavailable\"],\"onboarding.setup.pick_specific\":[\"Browse all models\"],\"onboarding.setup.priority.aria\":[\"Choose local model priority\"],\"onboarding.setup.priority.balanced\":[\"Balanced\"],\"onboarding.setup.priority.balanced.helper\":[\"Accuracy & size\"],\"onboarding.setup.priority.compact\":[\"Smallest\"],\"onboarding.setup.priority.compact.helper\":[\"Less storage\"],\"onboarding.setup.priority.custom\":[\"Custom\"],\"onboarding.setup.priority.placeholder\":[\"choose a priority\"],\"onboarding.setup.priority.quality\":[\"Best quality\"],\"onboarding.setup.priority.quality.helper\":[\"Most accurate\"],\"onboarding.setup.priority.question\":[\"What should your model prioritize?\"],\"onboarding.setup.sentence.language_prefix\":[\"I primarily dictate in\"],\"onboarding.setup.sentence.priority_prefix\":[\"and I want\"],\"onboarding.setup.sentence.shortcut_prefix\":[\"I'll use\"],\"onboarding.setup.sentence.shortcut_suffix\":[\"to dictate.\"],\"onboarding.setup.shortcut.active_help\":[\"Press the keys you want, or hit Esc to cancel.\"],\"onboarding.setup.shortcut.aria\":[\"Record new shortcut, currently \",[\"0\"]],\"onboarding.setup.shortcut.idle_help\":[\"Click the shortcut if you want to change it.\"],\"onboarding.setup.shortcut.label\":[\"Dictate with\"],\"onboarding.setup.shortcut.prompt\":[\"press a shortcut\"],\"onboarding.setup.subtitle\":[\"Answer a couple of plain questions and Glimpse will pick the local model.\"],\"onboarding.setup.title\":[\"Set up Glimpse\"],\"onboarding.status.checking\":[\"Checking...\"],\"onboarding.status.enabled\":[\"Enabled\"],\"onboarding.status.not_enabled\":[\"Not enabled\"],\"onboarding.welcome.cloud.badge\":[\"In progress\"],\"onboarding.welcome.cloud.title\":[\"Cloud\"],\"onboarding.welcome.cta\":[\"Continue\"],\"onboarding.welcome.local.aria\":[\"Select Local mode: privacy-first, on-device transcription\"],\"onboarding.welcome.local.feature.byok\":[\"Bring your own API key for AI features\"],\"onboarding.welcome.local.feature.free\":[\"Free, no account required\"],\"onboarding.welcome.local.feature.models\":[\"On-device transcription models\"],\"onboarding.welcome.local.title\":[\"Local\"],\"onboarding.welcome.subtitle\":[\"Choose how you want to transcribe.\"],\"onboarding.welcome.title\":[\"Welcome\"],\"personalization.apps\":[\"Apps\"],\"personalization.cancel\":[\"Cancel\"],\"personalization.delete\":[\"Delete\"],\"personalization.delete_mode.description\":[\"Delete \\\"\",[\"0\"],\"\\\"? This cannot be undone.\"],\"personalization.delete_mode.title\":[\"Delete mode?\"],\"personalization.description\":[\"Tailor language model behavior to apps, sites, and custom instructions.\"],\"personalization.disable_mode\":[\"Disable mode\"],\"personalization.empty.description\":[\"Create a mode to start customizing your apps and websites.\"],\"personalization.empty.title\":[\"No modes yet\"],\"personalization.enable_mode\":[\"Enable mode\"],\"personalization.modal.add\":[\"Add\"],\"personalization.modal.applications\":[\"Applications\"],\"personalization.modal.applications.add\":[\"Add application\"],\"personalization.modal.applications.none\":[\"No applications selected\"],\"personalization.modal.applications.not_installed\":[\"Not installed\"],\"personalization.modal.applications.search\":[\"Search applications...\"],\"personalization.modal.applications.selected\":[[\"0\"],\" selected\"],\"personalization.modal.applications.toggle_list\":[\"Toggle application list\"],\"personalization.modal.close\":[\"Close modal\"],\"personalization.modal.custom_instructions\":[\"Custom instructions\"],\"personalization.modal.custom_instructions.drag\":[\"Drag to resize\"],\"personalization.modal.custom_instructions.placeholder\":[\"Add custom instructions\"],\"personalization.modal.custom_instructions.resize\":[\"Resize custom instructions\"],\"personalization.modal.delete_mode\":[\"Delete mode\"],\"personalization.modal.edit_name\":[\"Edit mode name\"],\"personalization.modal.header\":[\"Personalization\"],\"personalization.modal.remove\":[\"Remove\"],\"personalization.modal.remove_app\":[\"Remove \",[\"app\"]],\"personalization.modal.remove_site\":[\"Remove \",[\"site\"]],\"personalization.modal.save_name\":[\"Save name\"],\"personalization.modal.snippets.full_list\":[\"Full snippet list\"],\"personalization.modal.snippets.info\":[\"Show personalization snippet examples\"],\"personalization.modal.snippets.summary\":[\"Use snippets to pass live context to the language model, like\"],\"personalization.modal.website.duplicate\":[\"That domain is already added\"],\"personalization.modal.website.invalid\":[\"Enter a valid domain like gmail.com\"],\"personalization.modal.websites\":[\"Websites\"],\"personalization.modal.websites.aria\":[\"Add website domain\"],\"personalization.modal.websites.count\":[[\"0\"],\" sites\"],\"personalization.modal.websites.none\":[\"No websites added\"],\"personalization.modal.websites.placeholder\":[\"Add a site like gmail.com\"],\"personalization.new_mode\":[\"New mode\"],\"personalization.new_mode.default_name\":[\"New Mode\"],\"personalization.no_apps\":[\"No apps yet\"],\"personalization.no_notes\":[\"No notes yet\"],\"personalization.no_sites\":[\"No sites yet\"],\"personalization.notes\":[\"Notes:\"],\"personalization.title\":[\"Personalization\"],\"personalization.websites\":[\"Websites\"],\"pill.background_visualizer\":[\"Background audio visualizer\"],\"pill.status.error\":[\"Error occurred\"],\"pill.status.listening\":[\"Listening...\"],\"pill.status.processing\":[\"Processing...\"],\"pill.visualizer\":[\"Audio visualizer\"],\"qhrhL0\":[[\"minutes\",\"plural\",{\"one\":[\"#\",\" minute spoken today\"],\"other\":[\"#\",\" minutes spoken today\"]}]],\"settings.about.advanced\":[\"Advanced\"],\"settings.about.cli.default_info\":[\"Installs the \",[\"0\"],\" command for Terminal, scripts, and automation tools. Use it when you want to call Glimpse programmatically without opening the app UI.\"],\"settings.about.cli.default_subtitle\":[\"Use Glimpse from Terminal or scripts\"],\"settings.about.cli.installed_info\":[\"The glimpse command is installed at \",[\"cliInstallPath\"],\". Use it from Terminal, scripts, or automation tools to call Glimpse without opening the app UI.\"],\"settings.about.cli.installed_subtitle\":[\"Installed at \",[\"cliInstallPath\"]],\"settings.about.cli.locked_info\":[\"Command line install requires an active license.\"],\"settings.about.cli.locked_subtitle\":[\"Requires an active license\"],\"settings.about.cli.path_missing_info\":[\"Installs \",[\"0\"],\" to \",[\"cliInstallPath\"],\". That folder is not currently on your shell PATH, so you may need to call it by full path or update your shell profile.\"],\"settings.about.cli.unavailable_info\":[\"This build does not include the command line helper.\"],\"settings.about.cli.unavailable_subtitle\":[\"Not available in this build\"],\"settings.about.command_line\":[\"Command line\"],\"settings.about.command_line.info_aria\":[\"More information about command line tools\"],\"settings.about.command_line.open_wiki_aria\":[\"Open the command line documentation\"],\"settings.about.faq_help\":[\"FAQ & Help\"],\"settings.about.faq_help_description\":[\"common questions\"],\"settings.about.install_cli\":[\"Install CLI\"],\"settings.about.mode.cloud\":[\"Cloud\"],\"settings.about.mode.local\":[\"Local\"],\"settings.about.restart_onboarding\":[\"Restart Onboarding\"],\"settings.about.restart_onboarding_description\":[\"re-run setup wizard\"],\"settings.about.restart_onboarding_hold_aria\":[\"Restart Onboarding. Hold to confirm.\"],\"settings.about.setup\":[\"Setup\"],\"settings.about.storage\":[\"Storage\"],\"settings.about.storage.database\":[\"Database\"],\"settings.about.storage.library\":[\"Library\"],\"settings.about.storage.models\":[\"Models\"],\"settings.about.storage.recordings\":[\"Recordings\"],\"settings.about.storage.total\":[\"Total\"],\"settings.about.uninstall_cli\":[\"Uninstall\"],\"settings.about.updates\":[\"Updates\"],\"settings.about.version\":[\"Version\"],\"settings.about.version_label\":[\"Glimpse\"],\"settings.account.action.deactivate\":[\"Deactivate this device\"],\"settings.account.action.deactivate_cancel\":[\"Cancel\"],\"settings.account.action.deactivate_confirm_short\":[\"Deactivate\"],\"settings.account.action.refresh\":[\"Refresh\"],\"settings.account.activate.body\":[\"Paste your activation code from Polar.\"],\"settings.account.activate.input_aria\":[\"Activation code\"],\"settings.account.activate.placeholder\":[\"GLIMPSE_…\"],\"settings.account.activate.submit\":[\"Activate\"],\"settings.account.section.activate\":[\"Already bought?\"],\"settings.account.trial.text\":[\"Trial · \",[\"remaining\"],\" of \",[\"total\"],\" days left\"],\"settings.account.trial.text_ended\":[\"Your trial has ended\"],\"settings.account.trial.text_ended_on\":[\"Your trial ended on\"],\"settings.account.trial.text_one\":[\"Trial · 1 day left\"],\"settings.app.accessibility\":[\"Accessibility\"],\"settings.app.accessibility.description\":[\"required for auto-paste\"],\"settings.app.analytics\":[\"Usage Analytics\"],\"settings.app.analytics.body\":[\"anonymous, no transcripts or audio shared.\"],\"settings.app.analytics.more_info\":[\"More info\"],\"settings.app.analytics.toggle_aria\":[\"Toggle usage analytics\"],\"settings.app.appearance\":[\"Appearance\"],\"settings.app.auto_delete\":[\"Auto-delete\"],\"settings.app.auto_delete.after\":[\"after\"],\"settings.app.auto_delete.body\":[\"Deleting transcripts also removes their saved audio.\"],\"settings.app.auto_delete.confirm.apply\":[\"Apply anyway\"],\"settings.app.auto_delete.confirm.audio_only\":[\"This only removes saved audio files, not your transcripts.\"],\"settings.app.auto_delete.confirm.audio_too\":[\"Deleting transcripts also removes the audio they reference.\"],\"settings.app.auto_delete.confirm.immediately.known_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" will immediately delete \",[\"candidateCount\"],\" existing \",[\"noun\"],\".\"],\"settings.app.auto_delete.confirm.immediately.unknown_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" may immediately delete your existing \",[\"noun\"],\".\"],\"settings.app.auto_delete.confirm.threshold.known_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" will immediately delete \",[\"candidateCount\"],\" \",[\"noun\"],\" already older than \",[\"threshold\"],\".\"],\"settings.app.auto_delete.confirm.threshold.unknown_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" may immediately delete \",[\"noun\"],\" already older than \",[\"threshold\"],\".\"],\"settings.app.auto_delete.confirm.title\":[\"Delete older items now?\"],\"settings.app.auto_delete.confirm.unknown_count\":[\"We couldn't count them right now, but auto-delete will still run as soon as you save this change.\"],\"settings.app.auto_delete_recordings\":[\"Auto-delete Recordings\"],\"settings.app.auto_delete_recordings.after_suffix\":[\" after\"],\"settings.app.auto_delete_recordings.body\":[\"Auto deletes local audio files.\"],\"settings.app.auto_delete_recordings.confirm.apply\":[\"Apply anyway\"],\"settings.app.auto_delete_recordings.confirm.audio_only\":[\"This only removes saved local audio files, not your transcript history.\"],\"settings.app.auto_delete_recordings.confirm.immediately.known_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" will immediately delete \",[\"candidateCount\",\"plural\",{\"one\":[\"#\",\" existing local recording\"],\"other\":[\"#\",\" existing local recordings\"]}],\".\"],\"settings.app.auto_delete_recordings.confirm.immediately.unknown_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" may immediately delete your existing local recordings.\"],\"settings.app.auto_delete_recordings.confirm.threshold.known_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" will immediately delete \",[\"candidateCount\",\"plural\",{\"one\":[\"#\",\" local recording that is already older than \",[\"threshold\"]],\"other\":[\"#\",\" local recordings that are already older than \",[\"threshold\"]]}],\".\"],\"settings.app.auto_delete_recordings.confirm.threshold.unknown_count\":[\"Changing auto-delete to \",[\"policyLabel\"],\" may immediately delete local recordings that are already older than \",[\"threshold\"],\".\"],\"settings.app.auto_delete_recordings.confirm.title\":[\"Delete older recordings now?\"],\"settings.app.auto_delete_recordings.confirm.unknown_count\":[\"We couldn't count them right now, but the cleanup will still run as soon as you save this change.\"],\"settings.app.auto_delete_recordings.preview.loading\":[\"Checking how many recordings would be deleted\"],\"settings.app.auto_launch\":[\"Launch at Login\"],\"settings.app.auto_launch.body\":[\"starts Glimpse automatically when you sign in.\"],\"settings.app.auto_launch.toggle_aria\":[\"Toggle launch at login\"],\"settings.app.auto_pause_media\":[\"Auto-pause Media\"],\"settings.app.auto_pause_media.body\":[\"pauses music while recording, resumes when done.\"],\"settings.app.auto_pause_media.body_duck\":[\"Lowers system volume while recording.\"],\"settings.app.auto_pause_media.body_off\":[\"System audio plays while recording.\"],\"settings.app.auto_pause_media.body_pause\":[\"Pauses system audio while recording.\"],\"settings.app.auto_pause_media.lower\":[\"Previous media action\"],\"settings.app.auto_pause_media.off\":[\"Off\"],\"settings.app.auto_pause_media.pause\":[\"Pause\"],\"settings.app.auto_pause_media.raise\":[\"Next media action\"],\"settings.app.auto_pause_media.toggle_aria\":[\"Toggle auto-pause media while recording\"],\"settings.app.auto_update\":[\"Auto-update\"],\"settings.app.auto_update.body\":[\"downloads and installs updates in the background.\"],\"settings.app.auto_update.toggle_aria\":[\"Toggle auto-update\"],\"settings.app.automation\":[\"Automation\"],\"settings.app.cancel\":[\"Cancel\"],\"settings.app.confirm\":[\"Confirm\"],\"settings.app.confirm.checking\":[\"Checking...\"],\"settings.app.input_monitoring\":[\"Input Monitoring\"],\"settings.app.input_monitoring.description\":[\"required for global shortcuts\"],\"settings.app.language.label\":[\"Language\"],\"settings.app.language.search\":[\"Search language...\"],\"settings.app.language.system\":[\"System\"],\"settings.app.microphone\":[\"Microphone\"],\"settings.app.microphone.description\":[\"required for transcription\"],\"settings.app.open_settings\":[\"Open Settings\"],\"settings.app.permission.checking\":[\"Checking permission\"],\"settings.app.permission.enabled\":[\"Enabled\"],\"settings.app.permission.off\":[\"off\"],\"settings.app.permissions_restart_notice\":[\"Permission changes may require a restart.\"],\"settings.app.privacy_permissions\":[\"Privacy & Permissions\"],\"settings.app.prune.day\":[\"A Day\"],\"settings.app.prune.instantly\":[\"Instantly\"],\"settings.app.prune.month\":[\"A Month\"],\"settings.app.prune.never\":[\"Never\"],\"settings.app.prune.noun.audio.one\":[\"audio file\"],\"settings.app.prune.noun.audio.other\":[\"audio files\"],\"settings.app.prune.noun.transcripts.one\":[\"transcript\"],\"settings.app.prune.noun.transcripts.other\":[\"transcripts\"],\"settings.app.prune.threshold.day\":[\"a day\"],\"settings.app.prune.threshold.immediately\":[\"right now\"],\"settings.app.prune.threshold.month\":[\"a month\"],\"settings.app.prune.threshold.week\":[\"a week\"],\"settings.app.prune.threshold.year\":[\"a year\"],\"settings.app.prune.week\":[\"A Week\"],\"settings.app.prune.year\":[\"A Year\"],\"settings.app.prune_target.audio\":[\"Audio\"],\"settings.app.prune_target.transcripts\":[\"Transcripts\"],\"settings.app.start_in_background\":[\"Start in background\"],\"settings.app.start_in_background.toggle_aria\":[\"Toggle start in background\"],\"settings.app.text_size.default\":[\"Default\"],\"settings.app.text_size.label\":[\"Text Size\"],\"settings.app.text_size.large\":[\"Large\"],\"settings.app.text_size.small\":[\"Small\"],\"settings.app.theme.dark\":[\"Dark\"],\"settings.app.theme.label\":[\"Theme\"],\"settings.app.theme.light\":[\"Light\"],\"settings.app.theme.system\":[\"System\"],\"settings.general.account_tab\":[\"Account\"],\"settings.general.auto_dictionary\":[\"Auto Dictionary\"],\"settings.general.auto_dictionary.body\":[\"suggests names and terms after you correct dictated text\"],\"settings.general.auto_dictionary.toggle_aria\":[\"Toggle Auto Dictionary\"],\"settings.general.auto_dictionary.unsupported_body\":[\"requires a model with dictionary support\"],\"settings.general.cloud.aria\":[\"Cloud processing (Coming soon)\"],\"settings.general.cloud.badge\":[\"coming soon\"],\"settings.general.cloud.description\":[\"In development\"],\"settings.general.cloud.label\":[\"Cloud\"],\"settings.general.download_one\":[\"Download one\"],\"settings.general.edit_mode\":[\"Edit Mode\"],\"settings.general.edit_mode.body\":[\"transform selected text with voice\"],\"settings.general.edit_mode.configure_prefix\":[\"Set up AI writing in\"],\"settings.general.edit_mode.help\":[\"Select text in any app, and speak a command like \\\"make this formal\\\" or \\\"fix my grammar\\\".\"],\"settings.general.edit_mode.help_license_requirement\":[\"Requires Glimpse Personal.\"],\"settings.general.edit_mode.help_requirement\":[\"Requires an enabled and configured writing provider.\"],\"settings.general.edit_mode.info_aria\":[\"More information about Edit Mode\"],\"settings.general.edit_mode.license_prefix\":[\"Activate Glimpse Personal in\"],\"settings.general.edit_mode.models_suffix\":[\"to use Edit Mode.\"],\"settings.general.edit_mode.toggle_aria\":[\"Toggle Edit Mode\"],\"settings.general.features\":[\"Features\"],\"settings.general.language_info.active_model\":[\"Language list is filtered to the active transcription model.\"],\"settings.general.language_info_aria\":[\"More information about transcription language support\"],\"settings.general.local.badge\":[\"private\"],\"settings.general.local.description\":[\"Runs entirely on your device\"],\"settings.general.local.label\":[\"Local\"],\"settings.general.microphone\":[\"Microphone\"],\"settings.general.microphone_test.busy_error\":[\"That microphone is already in use.\"],\"settings.general.microphone_test.done\":[\"Done\"],\"settings.general.microphone_test.generic_error\":[\"Couldn't start microphone test.\"],\"settings.general.microphone_test.not_found_error\":[\"No microphone was found.\"],\"settings.general.microphone_test.permission_error\":[\"Microphone access was denied.\"],\"settings.general.microphone_test.start_error\":[\"Couldn't start microphone test.\"],\"settings.general.microphone_test.stop\":[\"Stop\"],\"settings.general.microphone_test.test\":[\"Test\"],\"settings.general.microphone_test.unsupported\":[\"Microphone testing isn't available in this window.\"],\"settings.general.no_model\":[\"No model installed.\"],\"settings.general.processing\":[\"Processing\"],\"settings.general.processing_mode\":[\"Processing Mode\"],\"settings.general.providers_tab\":[\"Providers\"],\"settings.general.search_language\":[\"Search language...\"],\"settings.general.select_microphone\":[\"Select microphone...\"],\"settings.general.shortcut.toggle_aria\":[\"Toggle \",[\"label\"],\" shortcut\"],\"settings.general.shortcuts\":[\"Shortcuts\"],\"settings.general.shortcuts.add_shortcut\":[\"+ Add shortcut\"],\"settings.general.shortcuts.cleanup\":[\"Cleanup\"],\"settings.general.shortcuts.help_cleanup\":[\"Runs Cleanup for that shortcut only.\"],\"settings.general.shortcuts.help_temporary\":[\"Makes a shortcut temporary. It will not save audio, transcript, or history.\"],\"settings.general.shortcuts.hide_shortcuts\":[\"Hide shortcuts\"],\"settings.general.shortcuts.hold\":[\"Hold\"],\"settings.general.shortcuts.hold_description\":[\"hold to talk, release to stop\"],\"settings.general.shortcuts.info_aria\":[\"More information about shortcut options\"],\"settings.general.shortcuts.remove_shortcut\":[\"Remove shortcut\"],\"settings.general.shortcuts.show_shortcuts\":[\"Show shortcuts\"],\"settings.general.shortcuts.smart\":[\"Smart\"],\"settings.general.shortcuts.smart_description\":[\"tap to toggle, hold to talk\"],\"settings.general.shortcuts.temporary\":[\"Temporary\"],\"settings.general.shortcuts.toggle\":[\"Toggle\"],\"settings.general.shortcuts.toggle_description\":[\"tap to start, tap to stop\"],\"settings.general.system_default\":[\"System Default\"],\"settings.general.to_use_local\":[\"to use local.\"],\"settings.general.transcription_language\":[\"Transcription Language\"],\"settings.language_model.api_key\":[\"API Key\"],\"settings.language_model.api_key.aria\":[\"LLM API Key\"],\"settings.language_model.api_key.optional\":[\"Optional\"],\"settings.language_model.api_key.optional_hint\":[\"(if required)\"],\"settings.language_model.api_key.required\":[\"Required\"],\"settings.language_model.description\":[\"Used by Cleanup, Edit Mode, and Personalization.\"],\"settings.language_model.endpoint\":[\"Endpoint\"],\"settings.language_model.endpoint.aria\":[\"LLM Endpoint URL\"],\"settings.language_model.endpoint.autofilled\":[\"(auto-filled)\"],\"settings.language_model.endpoint.placeholder\":[\"https://your-llm-endpoint.com\"],\"settings.language_model.model\":[\"Model\"],\"settings.language_model.model.placeholder\":[\"Model (default: \",[\"0\"],\")\"],\"settings.language_model.model.search\":[\"Search available models...\"],\"settings.language_model.provider\":[\"Provider\"],\"settings.language_model.provider.cloud\":[\"Cloud (API Key)\"],\"settings.language_model.provider.custom\":[\"Custom\"],\"settings.language_model.provider.custom.description\":[\"Enter your own endpoint URL\"],\"settings.language_model.provider.local\":[\"Local\"],\"settings.language_model.provider.search\":[\"Search providers...\"],\"settings.language_model.provider.select\":[\"Select provider...\"],\"settings.language_model.title\":[\"Writing Model Provider\"],\"settings.language_model.toggle\":[\"Use this provider for AI writing features\"],\"settings.local_api.allow_browser_requests\":[\"Allow browser requests\"],\"settings.local_api.allow_browser_requests_aria\":[\"Allow cross-origin browser requests\"],\"settings.local_api.allow_browser_requests_help\":[\"send CORS headers so web apps can call the API\"],\"settings.local_api.api_key\":[\"API key\"],\"settings.local_api.api_key.lan_required\":[\"Required when listening on LAN\"],\"settings.local_api.api_key.placeholder_optional\":[\"Optional - blank to disable\"],\"settings.local_api.api_key.placeholder_required\":[\"Required for LAN access\"],\"settings.local_api.behavior\":[\"Behavior\"],\"settings.local_api.configuration\":[\"Configuration\"],\"settings.local_api.copy.copied_exclaim\":[\"Copied!\"],\"settings.local_api.listen_on_lan\":[\"Listen on LAN\"],\"settings.local_api.listen_on_lan_aria\":[\"Listen on LAN\"],\"settings.local_api.listen_on_lan_help\":[\"expose to other devices on your network\"],\"settings.local_api.logs\":[\"Logs\"],\"settings.local_api.logs.clear\":[\"Clear\"],\"settings.local_api.logs.copied\":[\"Copied\"],\"settings.local_api.logs.copy\":[\"Copy\"],\"settings.local_api.logs.copy_aria\":[\"Copy logs\"],\"settings.local_api.logs.empty\":[\"No logs yet.\"],\"settings.local_api.model.currently_loaded\":[\"currently loaded: \",[\"loadedModelLabel\"]],\"settings.local_api.model.kept_warm\":[\"kept warm in memory for fast responses\"],\"settings.local_api.model.none\":[\"None\"],\"settings.local_api.model.none_held\":[\"no model held in memory between requests\"],\"settings.local_api.port\":[\"Port\"],\"settings.local_api.preloaded_model\":[\"Preloaded model\"],\"settings.local_api.restart_required\":[\"Restart to apply changes\"],\"settings.local_api.start\":[\"Start API\"],\"settings.local_api.start_on_launch\":[\"Start on launch\"],\"settings.local_api.start_on_launch_aria\":[\"Start local API when Glimpse opens\"],\"settings.local_api.start_on_launch_help\":[\"automatically start when Glimpse opens\"],\"settings.local_api.status.running\":[\"Running\"],\"settings.local_api.status.stopped\":[\"Stopped\"],\"settings.local_api.stop\":[\"Stop API\"],\"settings.local_api.title\":[\"API Server\"],\"settings.modal.close_button\":[\"Close settings\"],\"settings.modal.dialog_label\":[\"Settings\"],\"settings.modal.section.core\":[\"Core\"],\"settings.modal.section.developer\":[\"Developer\"],\"settings.modal.section.local\":[\"Local\"],\"settings.modal.tab.about\":[\"About\"],\"settings.modal.tab.account\":[\"Account\"],\"settings.modal.tab.api_server\":[\"API Server\"],\"settings.modal.tab.app\":[\"App\"],\"settings.modal.tab.general\":[\"General\"],\"settings.modal.tab.models\":[\"Models\"],\"settings.modal.tab.providers\":[\"Providers\"],\"settings.modal.title\":[\"Settings\"],\"settings.models.active\":[\"Active\"],\"settings.models.back\":[\"Back\"],\"settings.models.back_overview\":[\"Overview\"],\"settings.models.browse_all\":[\"Browse all models\"],\"settings.models.cancel\":[\"Cancel\"],\"settings.models.cancel_download\":[\"Cancel download\"],\"settings.models.cancelled\":[\"Cancelled\"],\"settings.models.cloud_active\":[\"Glimpse is using \",[\"provider\"],\" to transcribe. Your active local model will be used as a fallback.\"],\"settings.models.cloud_active.provider_fallback\":[\"your speech provider\"],\"settings.models.delete\":[\"Delete\"],\"settings.models.delete_model\":[\"Delete model\"],\"settings.models.description\":[\"Manage local transcription engines and downloaded speech models.\"],\"settings.models.download\":[\"Download\"],\"settings.models.download_model\":[\"Download model\"],\"settings.models.engine.generic.description\":[[\"engineLabel\"],\" transcription engine.\"],\"settings.models.engine.nvidia.description\":[\"NVIDIA local speech models, including Parakeet for transcription and Nemotron for live streaming.\"],\"settings.models.engine.whisper.description\":[\"OpenAI's speech recognition with custom vocabulary support.\"],\"settings.models.installed\":[\"Installed\"],\"settings.models.installed.active\":[\"Active\"],\"settings.models.installed.ane\":[\"ANE\"],\"settings.models.installed.delete\":[\"Delete\"],\"settings.models.installed.delete_model\":[\"Delete model\"],\"settings.models.installed.english\":[\"English\"],\"settings.models.installed.multilingual\":[\"Multilingual\"],\"settings.models.installed.use\":[\"Use\"],\"settings.models.installed_count\":[[\"installedCount\"],\" installed\"],\"settings.models.recommended\":[\"Recommended\"],\"settings.models.remote_speech_active\":[\"Currently using your Speech Provider\",[\"0\"],\". Local models remain available as fallback.\"],\"settings.models.title\":[\"Models\"],\"settings.models.transcription_engines\":[\"Transcription Engines\"],\"settings.models.use\":[\"Use\"],\"settings.providers.description\":[\"Connect local-mode AI providers for Cleanup, Edit Mode, and Personalization.\"],\"settings.providers.language_label\":[\"Language\"],\"settings.providers.speech_label\":[\"Speech\"],\"settings.providers.title\":[\"Providers\"],\"settings.providers.writing_section\":[\"Writing\"],\"settings.speech_model.api_key\":[\"API Key\"],\"settings.speech_model.api_key.aria\":[\"Remote speech API key\"],\"settings.speech_model.api_key.optional\":[\"Optional\"],\"settings.speech_model.api_key.optional_hint\":[\"(if required)\"],\"settings.speech_model.api_key.required\":[\"Required\"],\"settings.speech_model.description\":[\"Transcribe recordings through OpenAI-compatible cloud or self-hosted APIs.\"],\"settings.speech_model.endpoint.aria\":[\"Remote speech endpoint URL\"],\"settings.speech_model.endpoint.placeholder\":[\"https://your-speech-endpoint.com\"],\"settings.speech_model.model\":[\"Model\"],\"settings.speech_model.model.automatic\":[\"Automatic (\",[\"0\"],\")\"],\"settings.speech_model.model.placeholder\":[\"Model (default: \",[\"0\"],\")\"],\"settings.speech_model.model.search\":[\"Search available speech models...\"],\"settings.speech_model.provider\":[\"Provider\"],\"settings.speech_model.provider.cloud\":[\"Cloud (API Key)\"],\"settings.speech_model.provider.custom\":[\"Custom\"],\"settings.speech_model.provider.custom.description\":[\"Enter your own endpoint URL\"],\"settings.speech_model.provider.local\":[\"Local\"],\"settings.speech_model.provider.search\":[\"Search speech providers...\"],\"settings.speech_model.provider.select\":[\"Select provider...\"],\"settings.speech_model.title\":[\"Remote Speech Provider\"],\"settings.speech_model.toggle\":[\"Use this provider for speech-to-text\"],\"toast.close\":[\"Close notification\"],\"toast.copied\":[\"Copied\"],\"toast.copy_message\":[\"Copy message\"],\"toast.retry_failed\":[\"Retry failed. Please try again.\"],\"toast.retry_transcription\":[\"Retry transcription\"],\"toast.retrying\":[\"Retrying...\"],\"transcription.language.auto\":[\"Auto\"],\"transcription.language.unsupported\":[\"Unsupported\"],\"transcription.language.unsupported_description\":[\"Choose a compatible model to enable these.\"],\"transcriptions.filter.sort\":[\"Sort\"],\"transcriptions.filter.when\":[\"When\"],\"transcriptions.group.today\":[\"Today\"],\"transcriptions.group.yesterday\":[\"Yesterday\"],\"transcriptions.item.cleaning\":[\"Cleaning...\"],\"transcriptions.item.copied\":[\"Copied\"],\"transcriptions.item.copy_selection\":[\"Copy selection\"],\"transcriptions.item.copy_transcription\":[\"Copy transcription\"],\"transcriptions.item.delete\":[\"Delete\"],\"transcriptions.item.error.default\":[\"Transcription failed\"],\"transcriptions.item.failed\":[\"Failed\"],\"transcriptions.item.more_options\":[\"More options\"],\"transcriptions.item.restore_original\":[\"Restore original transcript\"],\"transcriptions.item.retry\":[\"Retry\"],\"transcriptions.item.retry_cleanup\":[\"Retry cleanup\"],\"transcriptions.item.retrying\":[\"Retrying...\"],\"transcriptions.item.reverting\":[\"Reverting...\"],\"transcriptions.item.run_cleanup\":[\"Run cleanup\"],\"transcriptions.item.show_less\":[\"Show less\"],\"transcriptions.item.show_more\":[\"Show more\"],\"transcriptions.item.stop_retry\":[\"Stop transcription\"],\"transcriptions.list.empty\":[\"Your recent transcriptions will appear here\"],\"transcriptions.list.filter.aria\":[\"Sort and filter transcriptions\"],\"transcriptions.list.no_results\":[\"No results for \\\"\",[\"resultSearchText\"],\"\\\"\"],\"transcriptions.list.no_results_filters\":[\"No results for selected filters\"],\"transcriptions.list.search.aria\":[\"Search transcriptions\"],\"transcriptions.list.search.clear\":[\"Clear search\"],\"transcriptions.list.search.close\":[\"Close search\"],\"transcriptions.list.search.open\":[\"Search transcriptions\"],\"transcriptions.list.search.placeholder_short\":[\"Search\"],\"transcriptions.list.sort.aria\":[\"Sort transcriptions\"],\"transcriptions.sort.longest\":[\"Longest\"],\"transcriptions.sort.oldest\":[\"Oldest first\"],\"transcriptions.sort.recent\":[\"Newest first\"],\"transcriptions.sort.shortest\":[\"Shortest\"],\"transcriptions.time.7d\":[\"Past 7 days\"],\"transcriptions.time.any\":[\"Any time\"],\"transcriptions.time.today\":[\"Today\"],\"u-NVCP\":[[\"avg\",\"plural\",{\"one\":[\"About \",\"#\",\" word per dictation today\"],\"other\":[\"About \",\"#\",\" words per dictation today\"]}]],\"updates.available_version\":[\"v\",[\"version\"],\" available\"],\"updates.check_aria\":[\"Check for updates\"],\"updates.check_failed\":[\"Update check failed\"],\"updates.check_title\":[\"Check for updates\"],\"updates.checking\":[\"Checking for updates...\"],\"updates.installed\":[\"Update installed\"],\"updates.ready_to_install\":[\"Ready to install\"],\"updates.restart\":[\"Restart\"],\"updates.restart_to_apply\":[\"Restart to apply\"],\"updates.retry\":[\"Retry\"],\"updates.up_to_date\":[\"You're up to date\"],\"updates.update\":[\"Update\"],\"updates.whats_new\":[\"What's new?\"],\"updates.whats_new.couldnt_load\":[\"Couldn't load releases\"],\"updates.whats_new.github_unavailable\":[\"GitHub may be temporarily unavailable. Check your connection and try again.\"],\"updates.whats_new.load_failed\":[\"Failed to load changelog\"],\"updates.whats_new.major_release\":[\"Major Release\"],\"updates.whats_new.no_changelog\":[\"No changelog available.\"],\"updates.whats_new.retry\":[\"Retry\"],\"updates.whats_new.title\":[\"What's New\"],\"updates.whats_new.view_all\":[\"View all releases on GitHub\"]}")};
